### 2023-02-03 #5125 @danielskatz
@nwlandry - thanks for your submission. I see that your paper is ~2500 words, which far beyond the JOSS suggestion of ~1000. Please take a look at the JOSS example paper for what a JOSS paper should contain, and remove some of your current content (eg the API description). Perhaps some parts can be put in the repo and pointed at, or in the documentation?

Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to generate a new PDF and the command `@editorialbot check repository` to do a number of checks, including the length of the paper in words. editorialbot commands need to be the first entry in a new comment.

Once this is done, please let me know, and I'll then work on finding an editor for this submission.
