+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'XGI: A Python package for higher-order interaction networks'
+title: 'XGI: A Python package for research on higher-order interactions'
 tags:
   - python
   - higher-order
@@ -43,132 +43,53 @@ affiliations:
  - name: Max Planck Institute for Mathematics in the Sciences, Germany
    index: 6
 
-date: 3 February 2023
+date: 03/12/2023
 bibliography: references.bib
 
 ---
 
 # Summary
-The Comple**X** **G**roup **I**nteractions (XGI) library provides data structures and algorithms for modeling, analyzing, and visualizing complex systems with group (i.e., higher-order) interactions. At its core are hypergraphs and simplicial complexes. XGI provides multiple ways of building these objects—by hand, from data, or from generative models—and a range of standard and state-of-the-art algorithms and measures to analyze their structure. The library also offers several flexible visualization functions with suitable layouts. Additionally, XGI provides basic functions to simulate dynamical processes on hypergraphs and simplicial complexes. XGI is implemented in pure Python, and is designed to seamlessly interoperate with the rest of the Python scientific stack (Numpy, Scipy, Pandas, Matplotlib, NetworkX). XGI is designed and developed by network scientists with the needs of network scientists in mind.
+The Comple**X** **G**roup **I**nteractions (XGI) library provides data structures and algorithms for modeling, analyzing, and visualizing complex networks with group (i.e., higher-order) interactions. It includes both hypergraphs and simplicial complexes. XGI provides multiple methods to build them, standard and advanced algorithms to analyze their structure, flexible visualization functions. and simulation capabilities for dynamical processes. The library is accompanied by a collection of datasets including group interactions. XGI is implemented in pure Python and interoperates with the rest of the Python scientific stack (Numpy, Scipy, Pandas, Matplotlib, NetworkX). XGI is designed and developed by network scientists with the needs of network scientists in mind.
 
 # Statement of need
-The field of network science bridges across many different disciplines, bringing together theorists, computational scientists, social scientists, and many others, each with vastly different expertise. To facilitate collaboration and cross-disciplinary work, it is important to develop a tool kit that decreases the overhead cost of these collaborations. For pairwise networks, i.e. networks composed by units that interact in pairs, packages such as NetworkX [@hagberg_exploring_2008], graph-tool [@peixoto_graph-tool_2014], and igraph [@igraph] have provided a common framework for cross-disciplinary researchers to collaborate and contribute to the field of network science.
+The field of network science bridges across many different disciplines, bringing together theorists, computational scientists, social scientists, and many others. To facilitate cross-disciplinary collaboration, a common tool kit is crucial. Existing packages like NetworkX [@hagberg_exploring_2008], graph-tool [@peixoto_graph-tool_2014], and igraph [@igraph] have been successful in facilitating collaboration for traditional networks, restricted to pairwise interactions. However, the rapidly growing subfield of higher-order network science, that goes beyond pairwise to model group interactions of any number of units, requires a different approach. 
+Higher-order interaction networks promotes rich dynamical behavior [@iacopini_simplicial_2019;@skardal_abrupt_2019;@neuhauser_multibody_2020;@hickok_bounded-confidence_2022], and can model some empirical interaction patterns more accurately than pairwise networks [@chodrow_configuration_2020]. We anticipate that this field will have lasting impacts on various research areas such as infectious diseases, dynamical systems, and behavioral science. To support the higher-order network science community, we have developed the Comple**X** **G**roup **I**nteractions (XGI)---an open-source solution in Python.
 
[…]
