+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'XGI: A Python package for research on higher-order interactions'
+title: 'XGI: A Python package for higher-order interaction networks'
 tags:
   - python
   - higher-order
@@ -49,10 +49,10 @@ bibliography: references.bib
 ---
 
 # Summary
-The Comple**X** **G**roup **I**nteractions (XGI) library provides data structures and algorithms for modeling, analyzing, and visualizing complex networks with group (i.e., higher-order) interactions. It includes both hypergraphs and simplicial complexes. XGI provides multiple methods to build them, standard and advanced algorithms to analyze their structure, flexible visualization functions. and simulation capabilities for dynamical processes. The library is accompanied by a collection of datasets including group interactions. XGI is implemented in pure Python and interoperates with the rest of the Python scientific stack (Numpy, Scipy, Pandas, Matplotlib, NetworkX). XGI is designed and developed by network scientists with the needs of network scientists in mind.
+Comple**X** **G**roup **I**nteractions (XGI) is a library for higher-order networks. The library provides multiple methods to build hypergraphs and simplicial complexes. It comes with algorithms to analyze their structure, visualize them, and simulate dynamical processes on them. In addition, it is accompanied by a collection of higher-order datasets. XGI is implemented in pure Python and integrates with the rest of the Python scientific stack. XGI is designed and developed by network scientists with the needs of network scientists in mind.
 
 # Statement of need
-The field of network science bridges across many different disciplines, bringing together theorists, computational scientists, social scientists, and many others. To facilitate cross-disciplinary collaboration, a common tool kit is crucial. Existing packages like NetworkX [@hagberg_exploring_2008], graph-tool [@peixoto_graph-tool_2014], and igraph [@igraph] have been successful in facilitating collaboration for traditional networks, restricted to pairwise interactions. However, the rapidly growing subfield of higher-order network science, that goes beyond pairwise to model group interactions of any number of units, requires a different approach. 
+The field of network science bridges across many different disciplines, bringing together theorists, computational scientists, social scientists, and many others. To facilitate cross-disciplinary collaboration, a common tool kit is crucial. Existing packages like NetworkX [@hagberg_exploring_2008], graph-tool [@peixoto_graph-tool_2014], and igraph [@igraph] have been successful in facilitating collaboration for traditional networks, restricted to pairwise interactions. However, the rapidly growing subfield of higher-order network science, which models interactions between any number of entities, requires a different approach. 
 Higher-order interaction networks promotes rich dynamical behavior [@iacopini_simplicial_2019;@skardal_abrupt_2019;@neuhauser_multibody_2020;@hickok_bounded-confidence_2022], and can model some empirical interaction patterns more accurately than pairwise networks [@chodrow_configuration_2020]. We anticipate that this field will have lasting impacts on various research areas such as infectious diseases, dynamical systems, and behavioral science. To support the higher-order network science community, we have developed the Comple**X** **G**roup **I**nteractions (XGI)---an open-source solution in Python.
 
 
@@ -61,16 +61,14 @@ There are several existing packages to represent and analyze higher-order networ
 
 # Overview of the package
 
-The two core classes of the library are those representing hypergraphs and simplicial complexes. The data structure (seen in \autoref{fig:diagram}) employed by XGI for those two is a bipartite graph with entities represented by one node type and relationships among entities (i.e., hyperedges or simplices) represented by a second node type.
[…]
