+++ paper/references.bib
@@ -12,18 +12,6 @@
   archiveprefix = {arXiv}
 }
 
-@article{aksoy_measuring_2017,
-  title={Measuring and modeling bipartite graphs with community structure},
-  author={Aksoy, Sinan G and Kolda, Tamara G and Pinar, Ali},
-  journal={Journal of Complex Networks},
-  volume={5},
-  number={4},
-  pages={581--603},
-  year={2017},
-  publisher={Oxford University Press},
-  doi={10.1093/comnet/cnx001}
-}
-
 @article{arnaudon_connecting_2022,
   title = {Connecting {{Hodge}} and {{Sakaguchi-Kuramoto}} through a Mathematical Framework for Coupled Oscillators on Simplicial Complexes},
   author = {Arnaudon, Alexis and Peach, Robert L. and Petri, Giovanni and Expert, Paul},
@@ -49,21 +37,6 @@
   pages = {101301},
   issn = {2352-7110},
   doi = {10.1016/j.softx.2022.101301}
-}
-
-@article{bagrow_network_2022,
-  title = {Network Cards: Concise, Readable Summaries of Network Data},
-  shorttitle = {Network Cards},
-  author = {Bagrow, James and Ahn, Yong-Yeol},
-  year = {2022},
-  month = dec,
-  journal = {Applied Network Science},
-  volume = {7},
-  number = {1},
-  pages = {1--17},
-  publisher = {{SpringerOpen}},
-  issn = {2364-8228},
-  doi = {10.1007/s41109-022-00514-7}
 }
 
 @article{battiston_networks_2020,
@@ -118,16 +91,6 @@
   year = {2016}
 }
 
-@inproceedings{dewar_subgraphs_2016,
-  title={Subgraphs in non-uniform random hypergraphs},
-  author={Dewar, Megan and Healy, John and P{\'e}rez-Gim{\'e}nez, Xavier and Pra{\l}at, Pawe{\l} and Proos, John and Reiniger, Benjamin and Ternovsky, Kirill},
-  booktitle={Algorithms and Models for the Web Graph: 13th International Workshop, WAW 2016, Montreal, QC, Canada, December 14--15, 2016, Proceedings 13},
-  pages={140--151},
-  year={2016},
-  organization={Springer},
-  doi={10.1007/978-3-319-49787-7_12}
-}
-
 @article{diaz_hypergraphsjl_2022,
   title = {{{HyperGraphs}}.Jl: Representing Higher-Order Relationships in {{Julia}}},
   shorttitle = {{{HyperGraphs}}.Jl},
@@ -145,19 +108,6 @@
 @article{doecode_22160,
   title = {{{PNNL}}/{{HyperNetX}}},
   author = {Praggastis, Brenda and Arendt, Dustin and Joslyn, Cliff and Purvine, Emilie and Aksoy, Sinan and Monson, Kyle}
-}
-
-@article{gebru_datasheets_2021,
-  title = {Datasheets for Datasets},
-  author = {Gebru, Timnit and Morgenstern, Jamie and Vecchione, Briana and Vaughan, Jennifer Wortman and Wallach, Hanna and III, Hal Daum{\'e} and Crawford, Kate},
-  year = {2021},
-  month = nov,
-  journal = {Communications of the ACM},
-  volume = {64},
-  number = {12},
-  pages = {86--92},
-  issn = {0001-0782},
-  doi = {10.1145/3458723}
 }
 
 @inproceedings{hagberg_exploring_2008,
@@ -205,30 +155,6 @@
   url = {https://igraph.org},
 }
 
-@article{kahle_topology_2009,
-  title={Topology of random clique complexes},
-  author={Kahle, Matthew},
-  journal={Discrete mathematics},
[…]
