### 2023-02-07 #5125 @nwlandry
@danielskatz: We have trimmed down the PDF to 1011 words, replaced the thorough description of the API with a brief overview of the package, and verified that the sections that JOSS requires are in the manuscript. Let me know if you need anything else prior to review. Thanks so much!
