### 2023-02-07 #5125 @danielskatz
Can you add something in the intro that defines/explains higher-order networks?  Otherwise, this looks good, and I can start processing the submission while you do this.
