### 2023-02-08 c11a522
added brief description of higher-order networks
