### 2023-02-08 #5125 @nwlandry
> Can you add something in the intro that defines/explains higher-order networks? Otherwise, this looks good, and I can start processing the submission while you do this.

I added a brief description of higher-order networks in the introduction - happy to add a more detailed explanation if that is better.
