### 2023-02-09 #5125 @vissarion
> > Can you add something in the intro that defines/explains higher-order networks? Otherwise, this looks good, and I can start processing the submission while you do this.
>
> I added a brief description of higher-order networks in the introduction - happy to add a more detailed explanation if that is better.

It looks good to me, thanks!

Since I am not an expert in high-order networks, I found the following two introductory references very useful:
- https://doi.org/10.48550/arXiv.2104.11329
- https://doi.org/10.1017/9781108770996
