### 2023-02-09 #5125 @maximelucas
Great!

These are good references indeed. Another one, open access and by some authors of this manuscript:
- https://doi.org/10.1016/j.physrep.2020.05.004
