### 2023-02-13 #5125 @sudarshansudarshan
Not sure I understand what is this all about :)

Regards,
S.R.S. Iyengar

On Wed, 8 Feb 2023 at 14:52, Vissarion Fisikopoulos <
***@***.***> wrote:

> 👋 @sudarshansudarshan <https://github.com/sudarshansudarshan>,
> @saurabh-khanna <https://github.com/saurabh-khanna>, @manmeetkaurbaxi
> <https://github.com/manmeetkaurbaxi>, would any of you be willing to
> review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
