### 2023-02-15 a559929
docs: added existing functions
