+++ docs/source/api/generators/xgi.generators.classic.rst
@@ -8,6 +8,9 @@
    .. rubric:: Functions
    
    .. autofunction:: empty_hypergraph
+   .. autofunction:: empty_simplicial_complex
    .. autofunction:: flag_complex
+   .. autofunction:: flag_complex_d2
    .. autofunction:: star_clique
    .. autofunction:: sunflower
+   .. autofunction:: ring_lattice
