### 2023-02-17 #5162 @editorialbot
[REVIEW]: XGI: A Python package for higher-order interaction networks
**Submitting author:** @nwlandry (<a href="http://orcid.org/0000-0003-1270-4980">Nicholas Landry</a>)
**Repository:** https://github.com/ComplexGroupInteractions/xgi
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.6
