+++ xgi/generators/uniform.py
@@ -63,7 +63,8 @@ def uniform_hypergraph_configuration_model(k, m, seed=None):
     remainder = sum(k.values()) % m
     if remainder != 0:
         warnings.warn(
-            "This degree sequence is not realizable. Increasing the degree of random nodes so that it is."
+            """This degree sequence is not realizable. Increasing the degree of random nodes
+            so that it is."""
         )
         random_ids = random.sample(list(k.keys()), int(round(m - remainder)))
         for id in random_ids:
