+++ xgi/classes/simplicialcomplex.py
@@ -6,14 +6,18 @@ and can associate key/value attribute pairs with each undirected simplex and nod
 Multi-simplices are not allowed.
 """
 
-from collections.abc import Hashable, Iterable
-from itertools import combinations, count
+from collections.abc import Hashable
+from collections.abc import Iterable
+from itertools import combinations
+from itertools import count
 from warnings import warn
 
 from ..exception import XGIError
-from ..utils.utilities import powerset, update_uid_counter
+from ..utils.utilities import powerset
+from ..utils.utilities import update_uid_counter
 from .hypergraph import Hypergraph
-from .reportviews import EdgeView, NodeView
+from .reportviews import EdgeView
+from .reportviews import NodeView
 
 __all__ = ["SimplicialComplex"]
 
@@ -105,9 +109,11 @@ class SimplicialComplex(Hypergraph):
 
         """
         try:
-            return f"{type(self).__name__} named '{self['name']}' with {self.num_nodes} nodes and {self.num_edges} simplices"
+            return f"""{type(self).__name__} named '{self['name']}' with {self.num_nodes}
+        nodes and {self.num_edges} simplices"""
         except XGIError:
-            return f"Unnamed {type(self).__name__} with {self.num_nodes} nodes and {self.num_edges} simplices"
+            return f"""Unnamed {type(self).__name__} with {self.num_nodes} nodes and
+        {self.num_edges} simplices"""
 
     def add_edge(self, edge, id=None, **attr):
         """add_edge is deprecated in SimplicialComplex. Use add_simplex instead"""
@@ -116,17 +122,17 @@ class SimplicialComplex(Hypergraph):
 
     def add_edges_from(self, ebunch_to_add, max_order=None, **attr):
         """add_edges_from is deprecated in SimplicialComplex. Use add_simplices_from instead"""
-        warn(
-            "add_edges_from is deprecated in SimplicialComplex. Use add_simplices_from instead"
-        )
+        warn("add_edges_from is deprecated in SimplicialComplex. Use add_simplices_from instead")
         return self.add_simplices_from(ebunch_to_add, max_order=None, **attr)
 
-    def add_weighted_edges_from(
-        self, ebunch_to_add, max_order=None, weight="weight", **attr
-    ):
-        """add_weighted_edges_from is deprecated in SimplicialComplex. Use add_weighted_simplices_from instead"""
+    def add_weighted_edges_from(self, ebunch_to_add, max_order=None, weight="weight", **attr):
+        """add_weighted_edges_from is deprecated in SimplicialComplex.
+
+        Use add_weighted_simplices_from instead
+        """
         warn(
-            "add_weighted_edges_from is deprecated in SimplicialComplex. Use add_weighted_simplices_from instead"
+            """add_weighted_edges_from is deprecated in SimplicialComplex.
+            Use add_weighted_simplices_from instead"""
         )
         return self.add_weighted_simplices_from(
             ebunch_to_add, max_order=None, weight="weight", **attr
@@ -134,15 +140,17 @@ class SimplicialComplex(Hypergraph):
 
     def remove_edge(self, id):
         """remove_edge is deprecated in SimplicialComplex. Use remove_simplex_id instead"""
-        warn(
-            "remove_edge is deprecated in SimplicialComplex. Use remove_simplex_id instead"
-        )
-        return self.remove_simplex_id(id, **attr)
+        warn("remove_edge is deprecated in SimplicialComplex. Use remove_simplex_id instead")
[…]
