+++ tox.ini
@@ -1,0 +1,86 @@
+[base]
+name = xgi 
+files = {[base]name}
+
+[tox]
+envlist =
+    lint
+    py{38,39,310,311}
+    coverage
+    docs
+
+[testenv]
+setenv =
+    COVERAGE_FILE = {env:COVERAGE_FILE:.coverage-{envname}}
+extras = all
+commands = pytest \
+    --basetemp={envtmpdir} \
+    --cov={[base]name} \
+    --cov-fail-under=84
+
+allowlist_externals = make
+
+[testenv:coverage]
+skip_install = true
+deps = coverage
+allowlist_externals = /bin/bash
+commands =
+    /bin/bash -c 'coverage combine .coverage-py*'
+    coverage xml
+    coverage report
+
+[testenv:lint]
+basepython = python3
+extras = all
+deps =
+    pycodestyle
+    pydocstyle
+    pylint
+    isort
+    black
+
+commands =
+    pycodestyle {[base]files}
+    # pydocstyle {[base]files} --> this one could be added later, it will need a lot of work
+    isort --check-only --sl --diff {[base]files}
+    black --check --line-length 100 {[base]files}
+    pylint {[base]files}
+
+[testenv:format]
+deps = 
+    pybind11
+    black
+    isort
+commands =
+    isort --sl {[base]files}
+    black --line-length 100 {[base]files}
+
+[testenv:docs]
+extras = all
+changedir = docs
+deps = 
+    sphinx
+
+whitelist_externals = make
+commands =
+    make clean
+    make html
+
+[pycodestyle]
+ignore=W503,E731,W605,E203
+max-line-length=100
+
+[pydocstyle]
+add-ignore = D102, D105, D107
+convention = google
+
+[isort]
+line_length=100
+profile=black
+
+[gh-actions]
+python =
+  3.8: py38, lint
+  3.9: py39
+  3.10: py310
+  3.11: py311
