### 2023-02-23 93ef5c3
Drawlims (#279)
* fix: axlim in draw

* docs: added existing functions

* format with black

* fix: used to fill all triangles with p2=0

---------

Co-authored-by: Nicholas Landry <nicholas.landry.91@gmail.com>
