### 2023-03-03 d032a3b
added examples of sorting matrices (#282)
* added examples of sorting matrices
