+++ xgi/linalg/matrix.py
@@ -99,7 +99,7 @@ def incidence_matrix(
         return I
 
 
-def adjacency_matrix(H, order=None, s=1, weighted=False, index=False):
+def adjacency_matrix(H, order=None, sparse=True, s=1, weighted=False, index=False):
     """
     A function to generate an adjacency matrix (N,N) from a Hypergraph object.
 
@@ -110,6 +110,8 @@ def adjacency_matrix(H, order=None, s=1, weighted=False, index=False):
     order: int, optional
         Order of interactions to use. If None (default), all orders are used. If int,
         must be >= 1.
+    sparse: bool, default: True
+        Specifies whether the output matrix is a scipy sparse matrix or a numpy matrix
     s: int, default: 1
         Specifies the number of overlapping edges to be considered connected.
     index: bool, default: False
@@ -123,7 +125,7 @@ def adjacency_matrix(H, order=None, s=1, weighted=False, index=False):
         return A
 
     """
-    I, rowdict, coldict = incidence_matrix(H, index=True, order=order)
+    I, rowdict, coldict = incidence_matrix(H, order=order, sparse=sparse, index=True)
 
     if I.shape == (0,):
         if not rowdict:
@@ -146,7 +148,7 @@ def adjacency_matrix(H, order=None, s=1, weighted=False, index=False):
         return A
 
 
-def intersection_profile(H, order=None, index=False):
+def intersection_profile(H, order=None, sparse=True, index=False):
     """
     A function to generate an intersection profile from a Hypergraph object.
 
@@ -157,6 +159,8 @@ def intersection_profile(H, order=None, index=False):
     order: int, optional
         Order of interactions to use. If None (default), all orders are used. If int,
         must be >= 1.
+    sparse: bool, default: True
+        Specifies whether the output matrix is a scipy sparse matrix or a numpy matrix
     index: bool, default: False
         Specifies whether to output dictionaries mapping the edge IDs to indices
 
@@ -169,11 +173,8 @@ def intersection_profile(H, order=None, index=False):
 
     """
 
-    if index:
-        I, _, coldict = incidence_matrix(H, order=order, index=True)
-    else:
-        I = incidence_matrix(H, order=order, index=False)
-
+    I, _, coldict = incidence_matrix(H, order=order, sparse=sparse, index=True)
+    
     P = I.T.dot(I)
 
     if index:
@@ -182,7 +183,7 @@ def intersection_profile(H, order=None, index=False):
         return P
 
 
-def degree_matrix(H, order=None, index=False):
+def degree_matrix(H, order=None, sparse=True, index=False):
     """Returns the degree of each node as an array
 
     Parameters
@@ -192,6 +193,8 @@ def degree_matrix(H, order=None, index=False):
     order: int, optional
         Order of interactions to use. If None (default), all orders are used. If int,
         must be >= 1.
+    sparse: bool, default: True
+        Specifies whether the output matrix is a scipy sparse matrix or a numpy matrix
     index: bool, default: False
         Specifies whether to output disctionaries mapping the node and edge IDs to indices
 
@@ -203,7 +206,7 @@ def degree_matrix(H, order=None, index=False):
         return K
 
     """
-    I, rowdict, _ = incidence_matrix(H, order=order, index=True)
+    I, rowdict, _ = incidence_matrix(H, order=order, sparse=sparse, index=True)
 
     if I.shape == (0,):
         K = np.zeros(H.num_nodes)
@@ -213,7 +216,7 @@ def degree_matrix(H, order=None, index=False):
     return (K, rowdict) if index else K
 
 
-def laplacian(H, order=1, rescale_per_node=False, index=False):
[…]
