+++ .github/workflows/coverage.yml
@@ -1,0 +1,31 @@
+name: coverage
+
+on:
+  push:
+    branches: [main]
+  pull_request:
+    branches: [main]
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.head_ref || github.run_id }}
+  cancel-in-progress: true
+
+jobs:
+  report:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.10"
+      - name: Install packages
+        run: |
+          python -m pip install --upgrade pip wheel setuptools
+          pip install -r requirements/default.txt -r requirements/test.txt
+          pip install .
+          python -m pip list
+      - name: Run tests and collect coverage
+      run: pytest --cov=xgi --cov-report=xml
+      - name: Upload coverage reports to Codecov with GitHub Action
+        uses: codecov/codecov-action@v3
