+++ .github/workflows/coverage.yml
@@ -7,14 +7,14 @@ on:
     branches: [main]
 
 jobs:
-  report:
+  build:
     runs-on: ubuntu-latest
+    name: Test coverage
     steps:
       - uses: actions/checkout@v3
-      - name: Set up Python
         uses: actions/setup-python@v4
         with:
-          python-version: "3.10"
+          python-version: '3.11'
       - name: Install packages
         run: |
           python -m pip install --upgrade pip wheel setuptools
@@ -22,6 +22,6 @@ jobs:
           pip install .
           python -m pip list
       - name: Run tests and collect coverage
-      run: pytest --cov=xgi --cov-report=xml
+        run: pytest --cov=xgi --cov-report=xml
       - name: Upload coverage reports to Codecov with GitHub Action
         uses: codecov/codecov-action@v3
