+++ .github/workflows/coverage.yml
@@ -12,7 +12,7 @@ jobs:
     name: Test coverage
     steps:
       - uses: actions/checkout@v3
-        uses: actions/setup-python@v4
+      - uses: actions/setup-python@v4
         with:
           python-version: '3.11'
       - name: Install packages
