+++ .github/workflows/coverage.yml
@@ -22,6 +22,8 @@ jobs:
           pip install .
           python -m pip list
       - name: Run tests and collect coverage
-        run: pytest --cov=xgi --cov-report=xml
+        run: |
+        cd tests/
+        pytest --cov=xgi --cov-report=xml
       - name: Upload coverage reports to Codecov with GitHub Action
         uses: codecov/codecov-action@v3
