+++ README.md
@@ -2,6 +2,12 @@
 <img src='logo/logo.svg' width='150px' align="right" style="float:right;margin-left:10pt"></img>
 Comple**X** **G**roup **I**nteractions (**XGI**) is a Python package for the representation, manipulation, and study of the structure, dynamics, and functions of complex systems with group (higher-order) interactions.
 
+[![Test status](https://github.com/ComplexGroupInteractions/xgi/workflows/test/badge.svg?branch=main)](https://github.com/ComplexGroupInteractions/xgi/actions?query=workflow%3A%22test%22)
+
+[![codecov](https://codecov.io/gh/ComplexGroupInteractions/xgi/branch/main/graph/badge.svg?token=BI6TX2WDSG)](https://codecov.io/gh/ComplexGroupInteractions/xgi)
+
+[![Good First Issue](https://img.shields.io/github/labels/networkx/networkx/Good%20First%20Issue?color=green&label=Contribute%20&style=flat-square)](https://github.com/ComplexGroupInteractions/xgi/issues?q=is%3Aopen+is%3Aissue+label%3A%22Good+First+Issue%22)
+
 * [**Source**](../../)
 * [**Bug reports**](../../issues)
 * [**GitHub Discussions**](../../discussions)
