+++ xgi/classes/dihypergraph.py
@@ -71,65 +71,21 @@ class DiHypergraph:
     _hyperedge_attr_dict_factory = IDDict
     _hypergraph_attr_dict_factory = dict
 
-    def __getstate__(self):
-        """Function that allows pickling.
-
-        Returns
-        -------
-        dict
-            The keys label the hyeprgraph dict and the values
-            are dictionarys from the Hypergraph class.
-
-        Notes
-        -----
-        This allows the python multiprocessing module to be used.
-
-        """
-        return {
-            "_edge_uid": self._edge_uid,
-            "_hypergraph": self._hypergraph,
-            "_node": self._node,
-            "_node_attr": self._node_attr,
-            "_edge": self._edge,
-            "_edge_attr": self._edge_attr,
-        }
-
-    def __setstate__(self, state):
-        """Function that allows unpickling of a hypergraph.
-
-        Parameters
-        ----------
-        state
-            The keys access the dictionary names the values are the
-            dictionarys themselves from the Hypergraph class.
-
-        Notes
-        -----
-        This allows the python multiprocessing module to be used.
-        """
-        self._edge_uid = state["_edge_uid"]
-        self._hypergraph = state["_hypergraph"]
-        self._node = state["_node"]
-        self._node_attr = state["_node_attr"]
-        self._edge = state["_edge"]
-        self._edge_attr = state["_edge_attr"]
-        self._nodeview = NodeView(self)
-        self._edgeview = EdgeView(self)
-
     def __init__(self, incoming_data=None, **attr):
         self._edge_uid = count()
         self._hypergraph = self._hypergraph_attr_dict_factory()
-        self._node = self._node_dict_factory()
+        self._node_in = self._node_dict_factory()
+        self._node_out = self._node_dict_factory()
         self._node_attr = self._node_attr_dict_factory()
         self._edge_in = self._hyperedge_dict_factory()
         self._edge_out = self._hyperedge_dict_factory()
         self._edge_attr = self._hyperedge_attr_dict_factory()
 
-        self._nodeview = NodeView(self)
-        """A :class:`~xgi.classes.reportviews.NodeView` of the hypergraph."""
+        # self._nodeview = NodeView(self)
+        # """A :class:`~xgi.classes.reportviews.NodeView` of the hypergraph."""
 
-        self._edgeview = EdgeView(self)
-        """An :class:`~xgi.classes.reportviews.EdgeView` of the hypergraph."""
+        # self._edgeview = EdgeView(self)
+        # """An :class:`~xgi.classes.reportviews.EdgeView` of the hypergraph."""
 
         if incoming_data is not None:
             # This import needs to happen when this function is called, not when it is
@@ -138,155 +94,6 @@ class DiHypergraph:
 
             convert_to_hypergraph(incoming_data, create_using=self)
         self._hypergraph.update(attr)  # must be after convert
-
-    def __str__(self):
-        """Returns a short summary of the hypergraph.
-
-        Returns
-        -------
-        string
-            Hypergraph information
-
-        """
-        try:
-            return f"{type(self).__name__} named {self['name']} with {self.num_nodes} nodes and {self.num_edges} hyperedges"
-        except XGIError:
-            return f"Unnamed {type(self).__name__} with {self.num_nodes} nodes and {self.num_edges} hyperedges"
-
-    def __iter__(self):
-        """Iterate over the nodes.
-
-        Returns
-        -------
-        iterator
-            An iterator over all nodes in the hypergraph.
-        """
-        return iter(self._node)
-
-    def __contains__(self, n):
-        """Check for if a node is in this hypergraph.
-
-        Parameters
-        ----------
-        n : hashable
-            node ID
-
-        Returns
[…]
