+++ xgi/classes/hypergraph.py
@@ -377,6 +377,8 @@ class Hypergraph:
         See Also
         --------
         add_nodes_from
+        set_node_attributes
+        get_node_attributes
 
         Notes
         -----
@@ -406,7 +408,8 @@ class Hypergraph:
         See Also
         --------
         add_node
-
+        set_node_attributes
+        get_node_attributes
         """
         for n in nodes_for_adding:
             try:
@@ -502,6 +505,8 @@ class Hypergraph:
         See Also
         --------
         add_edges_from : Add a collection of edges.
+        set_edge_attributes
+        get_edge_attributes
 
         Examples
         --------
@@ -583,6 +588,8 @@ class Hypergraph:
         --------
         add_edge : Add a single edge.
         add_weighted_edges_from : Convenient way to add weighted edges.
+        set_edge_attributes
+        get_edge_attributes
 
         Notes
         -----
@@ -756,6 +763,8 @@ class Hypergraph:
         --------
         add_edge : Add a single edge.
         add_edges_from : Add multiple edges.
+        set_edge_attributes
+        get_edge_attributes
 
         Notes
         -----
