+++ docs/source/api/generators/xgi.generators.uniform.rst
@@ -8,3 +8,6 @@ xgi.generators.uniform
    .. rubric:: Functions
    
    .. autofunction:: uniform_hypergraph_configuration_model
+   .. autofunction:: uniform_SBM_hypergraph
+   .. autofunction:: uniform_hypergraph_planted_partition_model
+   .. autofunction:: uniform_erdos_renyi_hypergraph
