+++ tox.ini
@@ -61,7 +61,6 @@ changedir = docs
 deps = 
     sphinx
 
-whitelist_externals = make
 commands =
     make clean
     make html
@@ -80,7 +79,7 @@ profile=black
 
 [gh-actions]
 python =
-  3.8: py38, lint
+  3.8: py38, lint, docs
   3.9: py39
   3.10: py310
   3.11: py311
