+++ xgi/linalg/__init__.py
@@ -1,2 +1,3 @@
+"""Linalg module."""
 from . import matrix
 from .matrix import *
