+++ xgi/generators/__init__.py
@@ -1,3 +1,4 @@
+"""Generators module."""
 from . import classic, nonuniform, uniform
 from .classic import *
 from .nonuniform import *
