### 2023-03-11 ff53e29
Line graph and vector centrality functions added
