### 2023-03-11 f96edd5
Update xgi/algorithms/centrality.py
Co-authored-by: Leo Torres <dleonardotn@gmail.com>
