### 2023-03-11 37ff20c
Update xgi.algorithms.centrality.rst
