### 2023-03-13 76e8ff9
Fix `add_edge()` bug (#289)
* Fix `add_edge()` bug

* update test
