### 2023-03-13 4729d7e
Add codecov (#285)
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
