+++ .github/workflows/test.yml
@@ -9,7 +9,6 @@ on:
       - main
 
 jobs:
-
   tutorials:
     runs-on: ubuntu-latest
     strategy:
@@ -19,7 +18,7 @@ jobs:
     - uses: actions/checkout@v2
 
     - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
 
