+++ .github/workflows/coverage.yml
@@ -1,0 +1,29 @@
+name: coverage
+
+on:
+  push:
+    branches: [main]
+  pull_request:
+    branches: [main]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    name: Test coverage
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
+        with:
+          python-version: '3.11'
+      - name: Install packages
+        run: |
+          python -m pip install --upgrade pip wheel setuptools
+          pip install -r requirements/default.txt -r requirements/test.txt
+          pip install .
+          python -m pip list
+      - name: Run tests and collect coverage
+        run: |
+          cd tests/
+          pytest --cov=xgi --cov-report=xml
+      - name: Upload coverage reports to Codecov with GitHub Action
+        uses: codecov/codecov-action@v3
