+++ requirements/default.txt
@@ -1,6 +1,6 @@
-numpy>=1.19
-scipy>=1.5
-pandas>=1.0
+numpy>=1.20
+scipy>=1.8
+pandas>=1.3
 networkx>=2.7
 requests>=2.0
-matplotlib>=3.3
+matplotlib>=3.4
