### 2023-03-13 c70a698
Update xgi/generators/uniform.py
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
