+++ xgi/generators/uniform.py
@@ -257,7 +257,9 @@ def uniform_hypergraph_planted_partition_model(n, m, rho, k, epsilon):
 
 
 def uniform_erdos_renyi_hypergraph(n, m, k):
-    """Generate an m-uniform Erdős–Rényi hypergraph
+    """Generate an m-uniform Erdős–Rényi hypergraph.
+    
+    This creates an hypergraph with `n` nodes where hyperedges of size `m` created at random to obtain a mean degree of `k`. 
 
     Parameters
     ----------
