### 2023-03-13 4b507fb
Update xgi/generators/uniform.py
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
