+++ tests/generators/test_uniform.py
@@ -1,7 +1,8 @@
 import pytest
+import numpy as np
 
 import xgi
-
+from xgi.exception import XGIError
 
 def test_uniform_configuration_model_hypergraph():
     m = 3
@@ -16,3 +17,146 @@ def test_uniform_configuration_model_hypergraph():
         k = {1: 1, 2: 6}
         H = xgi.uniform_hypergraph_configuration_model(k, m)
     assert H.nodes.degree.asnumpy().sum() % m == 0
+
+def test_uniform_HSBM():
+    
+    # sum of sizes != n
+    with pytest.raises(XGIError):
+        m = 2
+        n = 10
+        sizes = [4, 5]
+        p = np.array([[0.5, 0.1], [0.1, 0.5]])
+        xgi.uniform_HSBM(n, m, p, sizes)
+
+    # non-square p
+    with pytest.raises(XGIError):
+        m = 2
+        n = 10
+        sizes = [4, 6]
+        p = np.array([[0.5, 0.1], [0.1, 0.5], [0, 1]])
+        xgi.uniform_HSBM(n, m, p, sizes)
+
+    # length of sizes and length of p don't match
+    with pytest.raises(XGIError):
+        m = 2
+        n = 10
+        sizes = [4, 5, 1]
+        p = np.array([[0.5, 0.1], [0.1, 0.5]])
+        xgi.uniform_HSBM(n, m, p, sizes)
+
+    # dim of p is not m
+    # non-square p
+    with pytest.raises(XGIError):
+        m = 3
+        n = 10
+        sizes = [4, 6]
+        p = np.array([[0.5, 0.1], [0.1, 0.5]])
+        xgi.uniform_HSBM(n, m, p, sizes)
+
+    # test p < 0
+    with pytest.raises(XGIError):
+        m = 2
+        n = 10
+        sizes = [4, 6]
+        p = np.array([[0.5, -0.1], [0.1, 0.5]])
+        xgi.uniform_HSBM(n, m, p, sizes)
+    
+    # test p > 1
+    with pytest.raises(XGIError):
+        m = 2
+        n = 10
+        sizes = [4, 6]
+        p = np.array([[0.5, 1.1], [0.1, 0.5]])
+        xgi.uniform_HSBM(n, m, p, sizes)
+    
+    m = 2
+    n = 10
+    sizes = [4, 6]
+    p = np.array([[0.5, 0.1], [0.1, 0.5]])
+    H1 = xgi.uniform_HSBM(n, m, p, sizes, seed=0)
+
+    assert H1.num_nodes == 10
+    assert xgi.unique_edge_sizes(H1) == [2]
+
+    # test that the seed works
+    H2 = xgi.uniform_HSBM(n, m, p, sizes, seed=0)
+
+    assert H1.edges.members(dtype=dict) == H2.edges.members(dtype=dict)
+
+
+def test_uniform_HPPM():
+    # rho < 0
+    with pytest.raises(XGIError):
+        m = 2
+        n = 10
+        rho = -0.1
+        k = 2
+        epsilon = 0.5
+        xgi.uniform_HPPM(n, m, rho, k, epsilon)
+    
+    # rho > 1
+    with pytest.raises(XGIError):
+        m = 2
+        n = 10
+        rho = 1.1
+        k = 2
+        epsilon = 0.5
+        xgi.uniform_HPPM(n, m, rho, k, epsilon)
+
+    # k < 0
+    with pytest.raises(XGIError):
+        m = 2
+        n = 10
+        rho = 0.5
+        k = -2
+        epsilon = 0.5
+        xgi.uniform_HPPM(n, m, rho, k, epsilon)
+
[…]
