+++ tests/linalg/test_matrix.py
@@ -114,35 +114,35 @@ def test_degree_matrix(edgelist1):
 
     K, node_dict = xgi.degree_matrix(H1, order=None, index=True)
     node_dict = {k: v for v, k in node_dict.items()}
-    assert K.shape == (8,)
+    assert K.shape == (8, 8)
 
-    assert K[node_dict[1]] == 1
-    assert K[node_dict[2]] == 1
-    assert K[node_dict[3]] == 1
-    assert K[node_dict[4]] == 1
-    assert K[node_dict[5]] == 1
-    assert K[node_dict[6]] == 2
-    assert K[node_dict[7]] == 1
-    assert K[node_dict[8]] == 1
+    assert K[node_dict[1], node_dict[1]] == 1
+    assert K[node_dict[2], node_dict[2]] == 1
+    assert K[node_dict[3], node_dict[3]] == 1
+    assert K[node_dict[4], node_dict[4]] == 1
+    assert K[node_dict[5], node_dict[5]] == 1
+    assert K[node_dict[6], node_dict[6]] == 2
+    assert K[node_dict[7], node_dict[7]] == 1
+    assert K[node_dict[8], node_dict[8]] == 1
 
     K1, node_dict1 = xgi.degree_matrix(H1, order=1, index=True)
     node_dict1 = {k: v for v, k in node_dict1.items()}
-    assert K1.shape == (8,)
+    assert K1.shape == (8, 8)
 
-    assert K1[node_dict[1]] == 0
-    assert K1[node_dict[2]] == 0
-    assert K1[node_dict[3]] == 0
-    assert K1[node_dict[4]] == 0
-    assert K1[node_dict[5]] == 1
-    assert K1[node_dict[6]] == 1
-    assert K1[node_dict[7]] == 0
-    assert K1[node_dict[8]] == 0
+    assert K1[node_dict[1], node_dict[1]] == 0
+    assert K1[node_dict[2], node_dict[2]] == 0
+    assert K1[node_dict[3], node_dict[3]] == 0
+    assert K1[node_dict[4], node_dict[4]] == 0
+    assert K1[node_dict[5], node_dict[5]] == 1
+    assert K1[node_dict[6], node_dict[6]] == 1
+    assert K1[node_dict[7], node_dict[7]] == 0
+    assert K1[node_dict[8], node_dict[8]] == 0
 
     K3, node_dict3 = xgi.degree_matrix(H1, order=3, index=True)
     node_dict3 = {k: v for v, k in node_dict3.items()}
-    assert K3.shape == (8,)
+    assert K3.shape == (8, 8)
     for i in range(8):
-        assert K3[i] == 0
+        assert K3[i, i] == 0
 
 
 def test_adjacency_matrix(edgelist1, edgelist4):
@@ -241,6 +241,13 @@ def test_adjacency_matrix(edgelist1, edgelist4):
     assert A4[node_dict4[4], node_dict4[5]] == 0
     assert A4[node_dict4[4], node_dict4[6]] == 0
 
+    A5 = xgi.adjacency_matrix(H1, sparse=False)
+    assert isinstance(A5, np.ndarray)
+    assert np.all(A5 == A1.todense())
+
+    A6 = xgi.adjacency_matrix(H1, order=1, sparse=False)
+    assert np.all(A6 == A3.todense())
+
 
 def test_laplacian(edgelist2, edgelist6):
     el1 = edgelist6
@@ -250,6 +257,8 @@ def test_laplacian(edgelist2, edgelist6):
 
     L1, node_dict1 = xgi.laplacian(H1, order=2, index=True)
     node_dict1 = {k: v for v, k in node_dict1.items()}
+
[…]
