+++ xgi/linalg/matrix.py
@@ -93,10 +93,8 @@ def incidence_matrix(
             members = H.edges.members(edge)
             for node in members:
                 I[node_dict[node], edge_dict[edge]] = weight(node, edge, H)
-    if index:
-        return I, rowdict, coldict
-    else:
-        return I
+
+    return (I, rowdict, coldict) if index else I
 
 
 def adjacency_matrix(H, order=None, sparse=True, s=1, weighted=False, index=False):
@@ -135,17 +133,21 @@ def adjacency_matrix(H, order=None, sparse=True, s=1, weighted=False, index=Fals
         return (A, {}) if index else A
 
     A = I.dot(I.T)
-    A = A - diags(A.diagonal())
+
+    if sparse:
+        A.setdiag(0)
+    else:
+        np.fill_diagonal(A, 0)
 
     if not weighted:
         A = (A >= s) * 1
     else:
-        A.data[np.where(A.data < s)] = 0
+        A = (A >= s) * A
 
-    if index:
-        return A, rowdict
-    else:
-        return A
+    if sparse:
+        A.eliminate_zeros()
+
+    return (A, rowdict) if index else A
 
 
 def intersection_profile(H, order=None, sparse=True, index=False):
@@ -172,18 +174,12 @@ def intersection_profile(H, order=None, sparse=True, index=False):
         return P
 
     """
-
     I, _, coldict = incidence_matrix(H, order=order, sparse=sparse, index=True)
-    
     P = I.T.dot(I)
-
-    if index:
-        return P, coldict
-    else:
-        return P
+    return (P, coldict) if index else P
 
 
-def degree_matrix(H, order=None, sparse=True, index=False):
+def degree_matrix(H, order=None, sparse=False, index=False):
     """Returns the degree of each node as an array
 
     Parameters
@@ -193,7 +189,7 @@ def degree_matrix(H, order=None, sparse=True, index=False):
     order: int, optional
         Order of interactions to use. If None (default), all orders are used. If int,
         must be >= 1.
-    sparse: bool, default: True
+    sparse: bool, default: False
         Specifies whether the output matrix is a scipy sparse matrix or a numpy matrix
     index: bool, default: False
         Specifies whether to output disctionaries mapping the node and edge IDs to indices
@@ -213,10 +209,15 @@ def degree_matrix(H, order=None, sparse=True, index=False):
     else:
         K = np.ravel(np.sum(I, axis=1))  # flatten
 
-    return (K, rowdict) if index else K
+    if sparse:
+        D = csr_array(diags(K))
+    else:
+        D = np.diag(K)
+
+    return (D, rowdict) if index else D
 
 
-def laplacian(H, order=1, sparse=True, rescale_per_node=False, index=False):
+def laplacian(H, order=1, sparse=False, rescale_per_node=False, index=False):
     """Laplacian matrix of order d, see [1].
 
     Parameters
@@ -226,7 +227,7 @@ def laplacian(H, order=1, sparse=True, rescale_per_node=False, index=False):
     order : int
         Order of interactions to consider. If order=1 (default),
         returns the usual graph Laplacian
-    sparse: bool, default: True
+    sparse: bool, default: False
         Specifies whether the output matrix is a scipy sparse matrix or a numpy matrix
     index: bool, default: False
         Specifies whether to output disctionaries mapping the node and edge IDs to indices
@@ -249,25 +250,25 @@ def laplacian(H, order=1, sparse=True, rescale_per_node=False, index=False):
         Physical Review Research, 2(3), 033410.
 
     """
[…]
