### 2023-03-14 940a3ad
Update xgi.generators.uniform.rst
