+++ tests/generators/test_nonuniform.py
@@ -1,4 +1,7 @@
 import pytest
+import random
+
+import numpy as np
 
 import xgi
 
@@ -14,8 +17,23 @@ def test_chung_lu_hypergraph():
         k2 = {1: 2, 1: 2}
         H = xgi.chung_lu_hypergraph(k1, k2)
 
+def test_dcsbm_hypergraph():
 
-def test_random_seed():
+    n = 50
+    k1 = {i : random.randint(1, n) for i in range(n)}
+    k2 = {i : sorted(k1.values())[i] for i in range(n)}
+    g1 = {i : random.choice([0, 1]) for i in range(n)}
+    g2 = {i : random.choice([0, 1]) for i in range(n)}
+    omega = np.array([[n//2, 10], [10, n//2]])
+
+    H = xgi.dcsbm_hypergraph(k1, k2, g1, g2, omega)
+
+    assert H.num_nodes == 50
+
+
+def test_random_hypergraph():
+
+    # seed
     H1 = xgi.random_hypergraph(10, [0.1, 0.001], seed=1)
     H2 = xgi.random_hypergraph(10, [0.1, 0.001], seed=2)
     H3 = xgi.random_hypergraph(10, [0.1, 0.001], seed=2)
@@ -23,9 +41,64 @@ def test_random_seed():
     assert H1._edge != H2._edge
     assert H2._edge == H3._edge
 
+    # wrong input
+    with pytest.raises(ValueError):
+        H1 = xgi.random_hypergraph(10, [1, 1.1])
+    with pytest.raises(ValueError):
+        H1 = xgi.random_hypergraph(10, [1, -2])
+
+def test_random_simplicial_complex():
+
+    # seed
     S1 = xgi.random_simplicial_complex(10, [0.1, 0.001], seed=1)
     S2 = xgi.random_simplicial_complex(10, [0.1, 0.001], seed=2)
     S3 = xgi.random_simplicial_complex(10, [0.1, 0.001], seed=2)
 
     assert S1._edge != S2._edge
     assert S2._edge == S3._edge
+
+    # wrong input 
+    with pytest.raises(ValueError):
+        S1 = xgi.random_simplicial_complex(10, [1, 1.1])
+    with pytest.raises(ValueError):
+        S1 = xgi.random_simplicial_complex(10, [1, -2])
+
+
+def test_random_flag_complex():
+
+    # seed
+    S1 = xgi.random_flag_complex(10, 0.1, seed=1)
+    S2 = xgi.random_flag_complex(10, 0.1, seed=2)
+    S3 = xgi.random_flag_complex(10, 0.1, seed=2)
+
+    assert S1._edge != S2._edge
+    assert S2._edge == S3._edge
+
+    # wrong input
+    with pytest.raises(ValueError):
+        S1 = xgi.random_flag_complex(10, 1.1)
+    with pytest.raises(ValueError):
+        S1 = xgi.random_flag_complex(10, -2)
+
+
+def test_random_flag_complex_d2():
+
+    # seed
+    S1 = xgi.random_flag_complex_d2(10, 0.1, seed=1)
+    S2 = xgi.random_flag_complex_d2(10, 0.1, seed=2)
+    S3 = xgi.random_flag_complex_d2(10, 0.1, seed=2)
+
+    assert S1._edge != S2._edge
+    assert S2._edge == S3._edge
+
+    # wrong input
+    with pytest.raises(ValueError):
+        S1 = xgi.random_flag_complex_d2(10, 1.1)
+    with pytest.raises(ValueError):
+        S1 = xgi.random_flag_complex_d2(10, -2)
+
+
+
+
+
+
