### 2023-03-14 d6824c8
added uniform generators (#286)
* added uniform generators

* fixed grammar

* update docs

* Update xgi/generators/uniform.py

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>

* Update xgi/generators/uniform.py

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>

* response to PR review

* added unit tests

* added two args to uniform_erdos_renyi_hypergraph

* Update xgi.generators.uniform.rst

---------

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
