### 2023-03-14 81805ef
changed `degree_matrix()` back to a vector
