+++ tests/generators/test_nonuniform.py
@@ -7,13 +7,13 @@ import xgi
 
 
 def test_chung_lu_hypergraph():
-    
+
     k1 = {1: 1, 2: 2, 3: 3, 4: 4}
     k2 = {1: 2, 2: 2, 3: 3, 4: 3}
     H = xgi.chung_lu_hypergraph(k1, k2)
     assert H.num_nodes == 4
 
-    # seed 
+    # seed
     H1 = xgi.chung_lu_hypergraph(k1, k2, seed=1)
     H2 = xgi.chung_lu_hypergraph(k1, k2, seed=2)
     H3 = xgi.chung_lu_hypergraph(k1, k2, seed=2)
@@ -23,7 +23,8 @@ def test_chung_lu_hypergraph():
     with pytest.warns(Warning):
         k1 = {1: 1, 2: 2}
         k2 = {1: 2, 1: 2}
-        H = xgi.chung_lu_hypergraph(k1, k2)    
+        H = xgi.chung_lu_hypergraph(k1, k2)
+
 
 def test_dcsbm_hypergraph():
 
@@ -38,7 +39,7 @@ def test_dcsbm_hypergraph():
 
     assert H.num_nodes == 50
 
-    # seed 
+    # seed
     H1 = xgi.dcsbm_hypergraph(k1, k2, g1, g2, omega, seed=1)
     H2 = xgi.dcsbm_hypergraph(k1, k2, g1, g2, omega, seed=2)
     H3 = xgi.dcsbm_hypergraph(k1, k2, g1, g2, omega, seed=2)
