### 2023-03-14 01d6729
Add sparse argument to matrix functions (#284)
* initial commit

* response to PR review

* changed `degree_matrix()` back to a vector
