+++ xgi/linalg/matrix.py
@@ -137,13 +137,11 @@ def incidence_matrix(
             members = H.edges.members(edge)
             for node in members:
                 I[node_dict[node], edge_dict[edge]] = weight(node, edge, H)
-    if index:
-        return I, rowdict, coldict
-    else:
-        return I
+
+    return (I, rowdict, coldict) if index else I
 
 
-def adjacency_matrix(H, order=None, s=1, weighted=False, index=False):
+def adjacency_matrix(H, order=None, sparse=True, s=1, weighted=False, index=False):
     """
     A function to generate an adjacency matrix (N,N) from a Hypergraph object.
 
@@ -154,6 +152,8 @@ def adjacency_matrix(H, order=None, s=1, weighted=False, index=False):
     order: int, optional
         Order of interactions to use. If None (default), all orders are used. If int,
         must be >= 1.
+    sparse: bool, default: True
+        Specifies whether the output matrix is a scipy sparse matrix or a numpy matrix
     s: int, default: 1
         Specifies the number of overlapping edges to be considered connected.
     index: bool, default: False
@@ -167,7 +167,7 @@ def adjacency_matrix(H, order=None, s=1, weighted=False, index=False):
         return A
 
     """
-    I, rowdict, coldict = incidence_matrix(H, index=True, order=order)
+    I, rowdict, coldict = incidence_matrix(H, order=order, sparse=sparse, index=True)
 
     if I.shape == (0,):
         if not rowdict:
@@ -177,20 +177,24 @@ def adjacency_matrix(H, order=None, s=1, weighted=False, index=False):
         return (A, {}) if index else A
 
     A = I.dot(I.T)
-    A = A - diags(A.diagonal())
+
+    if sparse:
+        A.setdiag(0)
+    else:
+        np.fill_diagonal(A, 0)
 
     if not weighted:
         A = (A >= s) * 1
     else:
-        A.data[np.where(A.data < s)] = 0
+        A = (A >= s) * A
 
-    if index:
-        return A, rowdict
-    else:
-        return A
+    if sparse:
+        A.eliminate_zeros()
+
+    return (A, rowdict) if index else A
 
 
-def intersection_profile(H, order=None, index=False):
+def intersection_profile(H, order=None, sparse=True, index=False):
     """
     A function to generate an intersection profile from a Hypergraph object.
 
@@ -201,6 +205,8 @@ def intersection_profile(H, order=None, index=False):
     order: int, optional
         Order of interactions to use. If None (default), all orders are used. If int,
         must be >= 1.
+    sparse: bool, default: True
+        Specifies whether the output matrix is a scipy sparse matrix or a numpy matrix
     index: bool, default: False
         Specifies whether to output dictionaries mapping the edge IDs to indices
 
@@ -212,18 +218,9 @@ def intersection_profile(H, order=None, index=False):
         return P
 
     """
-
-    if index:
-        I, _, coldict = incidence_matrix(H, order=order, index=True)
-    else:
-        I = incidence_matrix(H, order=order, index=False)
-
+    I, _, coldict = incidence_matrix(H, order=order, sparse=sparse, index=True)
     P = I.T.dot(I)
-
-    if index:
-        return P, coldict
-    else:
-        return P
+    return (P, coldict) if index else P
 
 
 def degree_matrix(H, order=None, index=False):
@@ -257,7 +254,7 @@ def degree_matrix(H, order=None, index=False):
     return (K, rowdict) if index else K
 
 
-def laplacian(H, order=1, rescale_per_node=False, index=False):
+def laplacian(H, order=1, sparse=False, rescale_per_node=False, index=False):
     """Laplacian matrix of order d, see [1].
[…]
