### 2023-03-14 e37c299
tests: added basic for layouts (#293)
* tests: added basic for layouts
* added seed argument to all layout functions
