### 2023-03-14 e8b7c84
tests: more for nonuniform generators. fix: keep isolated nodes when converting maximal SC to HG
