### 2023-03-14 eff5f70
feat: added conversion to max SC in random layout
