### 2023-03-15 b378e37
add tutorial reqs to test
