+++ .github/workflows/test.yml
@@ -68,7 +68,7 @@ jobs:
     - name: Install packages
       run: |
         python -m pip install --upgrade pip wheel setuptools
-        python -m pip install -r requirements/default.txt -r requirements/test.txt -r requirements/documentation.txt -r requirements/tutorials.txt
+        python -m pip install -r requirements/default.txt -r requirements/test.txt -r requirements/documentation.txt -r requirements/tutorial.txt
         python -m pip install .
         python -m pip list
 
