+++ tox.ini
@@ -82,4 +82,4 @@ python =
   3.8: py38, lint, docs
   3.9: py39
   3.10: py310
-  3.11: py311
+  3.11: py311, coverage
