+++ .github/workflows/run-tox.yml
@@ -30,8 +30,8 @@ jobs:
       run: |
         tox
     - name: Upload to codecov
-      if: ${{matrix.python-version == '3.10'}}
-      uses: codecov/codecov-action@v1
+      if: ${{matrix.python-version == '3.11'}}
+      uses: codecov/codecov-action@v3
       with:
         fail_ci_if_error: false
         files: ./coverage.xml
