+++ .github/workflows/run-tox.yml
@@ -9,9 +9,10 @@ on:
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}-latest
     strategy:
       matrix:
+        os: [ubuntu, macos, windows]
         python-version: ["3.8", "3.9", "3.10", "3.11"]
 
     steps:
