### 2023-03-15 bb619d2
remove coverage from tox (not windows compat)
