### 2023-03-15 e36b9d8
More tests for layouts (#296)
* tests: more for layouts. fix: keep isolated nodes when converting maximal SC to HG

* feat: added conversion to max SC in random layout
