### 2023-03-17 b167a46
Apply suggestions from code review
Two suggestions added, the rest have been replied to.

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
