### 2023-03-17 4b2e328
Added fixed suggestion from code review
