+++ requirements/developer.txt
@@ -1,4 +1,3 @@
 black[jupyter]==22.3.0
-pre-commit>=2.12
 isort==5.10.1
 pylint>=2.10
