### 2023-03-17 f548300
fixed add_nodes_from and add_edge
