+++ xgi/classes/dihypergraph.py
@@ -190,11 +190,11 @@ class DiHypergraph:
         """
         for n in nodes_for_adding:
             try:
-                newnode = n not in self._node
+                newnode = n not in self._node_in
                 newdict = attr
             except TypeError:
                 n, ndict = n
-                newnode = n not in self._node
+                newnode = n not in self._node_in
                 newdict = attr.copy()
                 newdict.update(ndict)
             if newnode:
@@ -274,66 +274,39 @@ class DiHypergraph:
             self.remove_node(n)
 
     def add_edge(self, members, id=None, **attr):
-        """Add one edge with optional attributes.
+        try:
+            head = set(members["head"])
+            tail = set(members["tail"])
+        except TypeError:
+            raise XGIError("Edge must be a dictionary with head and tail keys!")
 
-        Parameters
-        ----------
-        members : Iterable
-            An iterable of the ids of the nodes contained in the new edge.
-        id : hashable, default None
-            Id of the new edge. If None, a unique numeric ID will be created.
-        **attr : dict, optional
-            Attributes of the new edge.
-
-        Raises
-        -----
-        XGIError
-            If `members` is empty.
-
-        See Also
-        --------
-        add_edges_from : Add a collection of edges.
-
-        Examples
-        --------
-
-        Add edges with or without specifying an edge id.
-
-        >>> import xgi
-        >>> H = xgi.Hypergraph()
-        >>> H.add_edge([1, 2, 3])
-        >>> H.add_edge([3, 4], id='myedge')
-        >>> H.edges
-        EdgeView((0, 'myedge'))
-
-        Access attributes using square brackets.  By default no attributes are created.
-
-        >>> H.edges[0]
-        {}
-        >>> H.add_edge([1, 4], color='red', place='peru')
-        >>> H.edges
-        EdgeView((0, 'myedge', 1))
-        >>> H.edges[1]
-        {'color': 'red', 'place': 'peru'}
-
-        """
-        members = set(members)
         if not members:
             raise XGIError("Cannot add an empty edge")
 
-        if id in self._edge.keys():  # check that uid is not present yet
+        if id in self._edge_in.keys():  # check that uid is not present yet
             warn(f"uid {id} already exists, cannot add edge {members}")
             return
 
-        uid = next(self._edge_uid) if not id else id
+        uid = next(self._edge_uid) if id is None else id
 
-        self._edge[uid] = set()
-        for node in members:
-            if node not in self._node:
-                self._node[node] = set()
+        self._edge_in[uid] = set()
+        self._edge_out[uid] = set()
+
+        for node in head:
+            if node not in self._node_out:
+                self._node_in[node] = set()
+                self._node_out[node] = set()
                 self._node_attr[node] = self._node_attr_dict_factory()
-            self._node[node].add(uid)
-            self._edge[uid].add(node)
+            self._node_out[node].add(uid)
+            self._edge_in[uid].add(node)
+        
+        for node in tail:
+            if node not in self._node_in:
+                self._node_in[node] = set()
+                self._node_out[node] = set()
+                self._node_attr[node] = self._node_attr_dict_factory()
+            self._node_in[node].add(uid)
+            self._edge_out[uid].add(node)
 
         self._edge_attr[uid] = self._hyperedge_attr_dict_factory()
         self._edge_attr[uid].update(attr)
@@ -342,103 +315,6 @@ class DiHypergraph:
[…]
