+++ xgi/classes/dihypergraph.py
@@ -274,6 +274,36 @@ class DiHypergraph:
             self.remove_node(n)
 
     def add_edge(self, members, id=None, **attr):
+        """Add one edge with optional attributes.
+
+        Parameters
+        ----------
+        members : Iterable
+            An iterable of the ids of the nodes contained in the new edge.
+        id : hashable, default None
+            Id of the new edge. If None, a unique numeric ID will be created.
+        **attr : dict, optional
+            Attributes of the new edge.
+
+        Raises
+        -----
+        XGIError
+            If `members` is empty.
+
+        See Also
+        --------
+        add_edges_from : Add a collection of edges.
+
+        Examples
+        --------
+
+        Add edges with or without specifying an edge id.
+
+        >>> import xgi
+        >>> H = xgi.DiHypergraph()
+        >>> H.add_edge({"head": [1, 2, 3], "tail": [2, 3, 4]})
+        >>> H.add_edge({"head": [3, 4], "tail":{}}, id='myedge')
+        """
         try:
             head = set(members["head"])
             tail = set(members["tail"])
@@ -315,21 +345,134 @@ class DiHypergraph:
             update_uid_counter(self, id)
 
     def add_edges_from(self, ebunch_to_add, **attr):
+        """Add multiple edges with optional attributes.
+
+        Parameters
+        ----------
+        ebunch_to_add : Iterable
+
+            An iterable of edges.  This may be an iterable of iterables (Format 1),
+            where each element contains the members of the edge specified as valid node IDs.
+            Alternatively, each element could also be a tuple in any of the following
+            formats:
+
+            * Format 2: 2-tuple (members, edge_id), or
+            * Format 3: 2-tuple (members, attr), or
+            * Format 4: 3-tuple (members, edge_id, attr),
+
+            where `members` is an iterable of node IDs, `edge_id` is a hashable to use
+            as edge ID, and `attr` is a dict of attributes. Finally, `ebunch_to_add`
+            may be a dict of the form `{edge_id: edge_members}` (Format 5).
+
+            Formats 2 and 3 are unambiguous because `attr` dicts are not hashable, while `id`s must be.
+            In Formats 2-4, each element of `ebunch_to_add` must have the same length,
+            i.e. you cannot mix different formats.  The iterables containing edge
+            members cannot be strings.
+
+        attr : \*\*kwargs, optional
+            Additional attributes to be assigned to all edges. Attribues specified via
+            `ebunch_to_add` take precedence over `attr`.
+
+        See Also
+        --------
+        add_edge : Add a single edge.
+        add_weighted_edges_from : Convenient way to add weighted edges.
+
+        Notes
+        -----
+        Adding the same edge twice will create a multi-edge. Currently
+        cannot add empty edges; the method skips over them.
+
+        Examples
+        --------
+        >>> import xgi
+        >>> H = xgi.Hypergraph()
+
+        When specifying edges by their members only, numeric edge IDs will be assigned
+        automatically.
+
+        >>> H.add_edges_from([[0, 1], [1, 2], [2, 3, 4]])
+        >>> H.edges.members(dtype=dict)
+        {0: {0, 1}, 1: {1, 2}, 2: {2, 3, 4}}
+
+        Custom edge ids can be specified using a dict.
+
+        >>> H = xgi.Hypergraph()
+        >>> H.add_edges_from({'one': [0, 1], 'two': [1, 2], 'three': [2, 3, 4]})
+        >>> H.edges.members(dtype=dict)
+        {'one': {0, 1}, 'two': {1, 2}, 'three': {2, 3, 4}}
+
+        You can use the dict format to easily add edges from another hypergraph.
+
[…]
