### 2023-03-17 cdff9b7
removed call to nodeview
