+++ xgi/classes/hypergraph.py
@@ -898,7 +898,7 @@ class Hypergraph:
         remove_edges_from : Remove multiple edges.
 
         """
-        for node in self.edges.members(id):
+        for node in self._edge[id].copy():
             self._node[node].remove(id)
         del self._edge[id]
         del self._edge_attr[id]
@@ -922,7 +922,7 @@ class Hypergraph:
 
         """
         for id in ebunch:
-            for node in self.edges.members(id):
+            for node in self._edge[id].copy():
                 self._node[node].remove(id)
             del self._edge[id]
             del self._edge_attr[id]
