### 2023-03-17 cedb69a
added ability to remove edges
