### 2023-03-17 a9b2daa
Update dihypergraph.py
