### 2023-03-17 badb28f
Update dihypergraph.py
