### 2023-03-21 6a5fb60
Line graph and vector centrality functions added (#290)
* Line graph and vector centrality functions added

* Update xgi/algorithms/centrality.py

Co-authored-by: Leo Torres <dleonardotn@gmail.com>

* Update xgi.algorithms.centrality.rst

* Update centrality.py

Changed vector_centrality to line_vector_centrality

* Apply suggestions from code review

Co-authored-by: Leo Torres <dleonardotn@gmail.com>

* Apply suggestions from code review

Two suggestions added, the rest have been replied to.

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>

* Added fixed suggestion from co
[…]
