+++ .github/workflows/test.yml
@@ -35,7 +35,7 @@ jobs:
         # go into tutorial location
         cd tutorials/
         # convert each notebook into a python file
-        jupyter nbconvert --to script Tutorial*ipynb
+        jupyter nbconvert --to script *.ipynb
         # rename the files so they are found by pytest: add 'test' at the beginning
         rename 's/(.*)\.py/test $1.py/' *
         # rename the files so they are found by pytest: change spaces to underscores
