+++ tests/drawing/test_xgi_pylab.py
@@ -1,0 +1,26 @@
+def test_draw():
+    assert 0 == 0
+
+def test_draw_nodes():
+    assert 0 == 0
+
+def test_draw_hyperedges():
+    assert 0 == 0
+
+def test_draw_simplices():
+    assert 0 == 0
+
+def test_CCW_sort():
+    assert 0 == 0
+
+def test_scalar_arg_to_dict():
+    assert 0 == 0
+
+def test_color_arg_to_dict():
+    assert 0 == 0
+
+def test_draw_node_labels():
+    assert 0 == 0
+
+def test_draw_hyperedge_labels():
+    assert 0 == 0
