+++ tests/generators/test_nonuniform.py
@@ -1,7 +1,7 @@
-import pytest
 import random
 
 import numpy as np
+import pytest
 
 import xgi
 
