### 2023-03-21 a451564
rename from `xgi_pylab` to `draw`
