+++ docs/source/api/drawing/xgi.drawing.draw.rst
@@ -8,8 +8,8 @@ xgi.drawing.xgi_pylab
    .. rubric:: Functions
    
    .. autofunction:: draw
-   .. autofunction:: draw_xgi_nodes
-   .. autofunction:: draw_xgi_hyperedges
-   .. autofunction:: draw_xgi_simplices
+   .. autofunction:: draw_nodes
+   .. autofunction:: draw_hyperedges
+   .. autofunction:: draw_simplices
    .. autofunction:: draw_node_labels
    .. autofunction:: draw_hyperedge_labels
