### 2023-03-22 6f4c35f
perf: cleaner and faster incidence_matrix
