### 2023-03-22 893299e
fix: return sparse (if asked) for empty H (sparse and non-sparse now have dim (0,0)). fixed corresponding tests.
