+++ codecov.yml
@@ -1,0 +1,10 @@
+coverage:
+  precision: 2
+  round: down
+  range: "60..85"
+  status:
+    project:
+      default:
+        # basic
+        target: 0
+        threshold: 1
