### 2023-03-22 3ac3a94
tests: more to check sparse equals non sparse
