### 2023-03-22 7a0500f
Return axes (#300)
* return axes and rename

* style: format with isort and black

* rename from `xgi_pylab` to `draw`

* update docs

* added pca transform function

* added tests

* style: format with isort and black

* update docs

* response to review

* response to review

* Create codecov.yml
