+++ xgi/convert.py
@@ -136,7 +136,7 @@ def convert_to_graph(H):
 
 def convert_to_line_graph(H):
     """Line graph of the hypergraph.
-    
+
     The line graph of the hypergraph `H` is the graph whose
     nodes correspond to each hyperedge in `H`, linked together
     if they share at least one vertex.
@@ -155,7 +155,7 @@ def convert_to_line_graph(H):
     LG = nx.Graph()
 
     edge_label_dict = {tuple(edge): index for index, edge in H._edge.items()}
-    
+
     nodes = sorted(set(edge_label_dict.values()))
     LG.add_nodes_from(nodes)
 
@@ -492,7 +492,7 @@ def from_simplicial_complex_to_hypergraph(SC):
 
     max_simplices = maximal_simplices(SC)
     H = Hypergraph()
-    H.add_nodes_from(SC.nodes) # to keep node order and isolated nodes
+    H.add_nodes_from(SC.nodes)  # to keep node order and isolated nodes
     H.add_edges_from([list(SC.edges.members(e)) for e in max_simplices])
     return H
 
