+++ docs/source/api/drawing/xgi.drawing.xgi_pylab.rst
@@ -1,15 +1,0 @@
-xgi.drawing.xgi_pylab
-=====================
-
-.. currentmodule:: xgi.drawing.xgi_pylab
-
-.. automodule:: xgi.drawing.xgi_pylab
-   
-   .. rubric:: Functions
-   
-   .. autofunction:: draw
-   .. autofunction:: draw_xgi_nodes
-   .. autofunction:: draw_xgi_hyperedges
-   .. autofunction:: draw_xgi_simplices
-   .. autofunction:: draw_node_labels
-   .. autofunction:: draw_hyperedge_labels
