+++ docs/source/api/drawing/xgi.drawing.draw.rst
@@ -1,0 +1,15 @@
+xgi.drawing.draw
+================
+
+.. currentmodule:: xgi.drawing.draw
+
+.. automodule:: xgi.drawing.draw
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: draw
+   .. autofunction:: draw_nodes
+   .. autofunction:: draw_hyperedges
+   .. autofunction:: draw_simplices
+   .. autofunction:: draw_node_labels
+   .. autofunction:: draw_hyperedge_labels
