### 2023-03-22 5efe433
Speed and memory improvements in linalg (#303)
* perf: cleaner and faster incidence_matrix

* fix: return sparse (if asked) for empty H (sparse and non-sparse now have dim (0,0)). fixed corresponding tests.

* tests: more to check sparse equals non sparse

* tests: more for sparse

* perf: made multiorder_laplacian truly sparse and faster #301

* style: ran isort and black
