+++ CITATION.cff
@@ -26,4 +26,4 @@ cff-version: "1.1.0"
 license: "BSD-3"
 message: "If you use this software, please cite it using these metadata."
 title: XGI
-version: "0.5.4"
+version: "0.5.5"
