### 2023-03-27 ff5c8e6
minor: renamed to from_max_simplices
