### 2023-03-27 4a998c8
fix: raise error for assortativity of empty hypergraph
