### 2023-03-27 530c5ae
minor: renamed to from_max_simplices (#306)
* minor: renamed to from_max_simplices

* Update codecov.yml

* Update codecov.yml

---------

Co-authored-by: Nicholas Landry <nicholas.landry.91@gmail.com>
