### 2023-03-27 bbc9877
feat: keep_isolates argument in subhypergraph
