### 2023-03-27 6cb328c
fix: raise error for assortativity of empty hypergraph (#307)
* fix: raise error for assortativity of empty hypergraph

* fix: review comment
