+++ xgi/algorithms/assortativity.py
@@ -37,8 +37,10 @@ def dynamical_assortativity(H):
     DOI: 10.1063/5.0086905
 
     """
-    if H.num_nodes == 0 or H.num_edges == 0:
-        raise XGIError("Hypergraph must contain nodes and edges!")
+    if H.num_nodes == 0:
+        raise XGIError("Hypergraph must contain nodes")
+    elif H.num_edges == 0:
+        raise XGIError("Hypergraph must contain edges!")
 
     if not is_uniform(H):
         raise XGIError("Hypergraph must be uniform!")
@@ -82,6 +84,11 @@ def degree_assortativity(H, kind="uniform", exact=False, num_samples=1000):
     float
         the degree assortativity
 
+    Raises
+    ------
+    XGIError
+        If there are no nodes or no edges
+
     References
     ----------
     Phil Chodrow,
@@ -89,6 +96,12 @@ def degree_assortativity(H, kind="uniform", exact=False, num_samples=1000):
     Journal of Complex Networks 2020.
     DOI: 10.1093/comnet/cnaa018
     """
+
+    if H.num_nodes == 0:
+        raise XGIError("Hypergraph must contain nodes")
+    elif H.num_edges == 0:
+        raise XGIError("Hypergraph must contain edges!")
+
     degs = H.degree()
     if exact:
         k1k2 = [
