### 2023-03-27 bcb84b7
keep_isolates argument in subhypergraph (#308)
* feat: keep_isolates argument in subhypergraph

* style: black and isort

* format with isort and updated config

* made coverage more strict

---------

Co-authored-by: Nicholas Landry <nicholas.landry.91@gmail.com>
