+++ .isort.cfg
@@ -1,2 +1,3 @@
 [settings]
 skip=__init__.py, benchmarks/env
+profile=black
