### 2023-03-28 25c71f8
tests: clustering. drawing: giving default params values to all functions
