+++ xgi/drawing/draw.py
@@ -81,7 +81,7 @@ def draw(
         If a dict, must contain (node_id: width) pairs.  If iterable, assume the widths are
         specified in the same order as the nodes are found in H.nodes. If NodeStat, use a monotonic
         linear interpolation defined between min_node_lw and max_node_lw.
-    node_size : int, float, dict, iterable, or NodeStat (default=10)
+    node_size : int, float, dict, iterable, or NodeStat (default=15)
         Radius of the nodes in pixels.  If int or float, use the same radius for all nodes.
         If a dict, must contain (node_id: radius) pairs.  If iterable, assume the radiuses are
         specified in the same order as the nodes are found in H.nodes. If NodeStat, use a monotonic
@@ -213,15 +213,15 @@ def draw(
 
 def draw_nodes(
     H,
-    pos,
-    ax,
-    node_fc,
-    node_ec,
-    node_lw,
-    node_size,
-    zorder,
-    settings,
-    node_labels,
+    pos=None,
+    ax=None,
+    node_fc="white",
+    node_ec="black",
+    node_lw=1,
+    node_size=15,
+    zorder=None,
+    settings=None,
+    node_labels=False,
     **kwargs,
 ):
     """Draw the nodes of a hypergraph
@@ -231,9 +231,10 @@ def draw_nodes(
     H : Hypergraph or SimplicialComplex
         Higher-order network to plot.
     ax : matplotlib.pyplot.axes
-    pos : dict
-        Dictionary of positions node_id:(x,y) of every node.
-    node_fc : str, dict, iterable, or NodeStat
+    pos : dict (default=None)
+        If passed, this dictionary of positions node_id:(x,y) is used for placing the 0-simplices.
+        If None (default), use the `barycenter_spring_layout` to compute the positions.
+    node_fc : str, dict, iterable, or NodeStat (default='white')
         Color of the nodes.  If str, use the same color for all nodes.  If a dict, must
         contain (node_id: color_str) pairs.  If other iterable, assume the colors are
         specified in the same order as the nodes are found in H.nodes. If NodeStat,
@@ -248,7 +249,7 @@ def draw_nodes(
         If a dict, must contain (node_id: width) pairs.  If iterable, assume the widths are
         specified in the same order as the nodes are found in H.nodes. If NodeStat, use a monotonic
         linear interpolation defined between min_node_lw and max_node_lw.
-    node_size : int, float, dict, iterable, or NodeStat (default=10)
+    node_size : int, float, dict, iterable, or NodeStat (default=15)
         Radius of the nodes in pixels.  If int or float, use the same radius for all nodes.
         If a dict, must contain (node_id: radius) pairs.  If iterable, assume the radiuses are
         specified in the same order as the nodes are found in H.nodes. If NodeStat, use a monotonic
@@ -276,6 +277,25 @@ def draw_nodes(
     draw_node_labels
     draw_hyperedge_labels
     """
+
+    if settings is None:
+        settings = {
+            "min_node_size": 10.0,
+            "max_node_size": 30.0,
+            "min_node_lw": 1.0,
+            "max_node_lw": 5.0,
+            "node_fc_cmap": cm.Reds,
+            "node_ec_cmap": cm.Greys,
+        }
+
+    settings.update(kwargs)
+
+    if pos is None:
+        pos = barycenter_spring_layout(H)
+
+    if ax is None:
+        ax = plt.gca()
+
     # Note Iterable covers lists, tuples, ranges, generators, np.ndarrays, etc
[…]
