### 2023-03-28 d22413e
added handling for empty and disconnected hypergraphs
