+++ tests/algorithms/test_centrality.py
@@ -7,6 +7,16 @@ from xgi.exception import XGIError
 
 
 def test_cec_centrality():
+    # test empty hypergraph
+    H = xgi.Hypergraph()
+    assert xgi.CEC_centrality(H) == dict()
+    # Test no edges
+    H.add_nodes_from([0, 1, 2])
+    assert xgi.CEC_centrality(H) == {0: np.NaN, 1: np.NaN, 2: np.NaN}
+    # test disconnected
+    H.add_edge([0, 1])
+    assert xgi.CEC_centrality(H) == {0: np.NaN, 1: np.NaN, 2: np.NaN}
+
     H = xgi.sunflower(3, 1, 3)
     c = H.nodes.cec_centrality.asnumpy()
     assert norm(c[1:] - c[1]) < 1e-4
@@ -18,8 +28,21 @@ def test_cec_centrality():
     assert abs(c[0] / c[1] - ratio(5, 7, kind="CEC")) < 1e-4
 
 
-@pytest.mark.slow
+# @pytest.mark.slow
 def test_hec_centrality():
+    # test empty hypergraph
+    H = xgi.Hypergraph()
+    c = xgi.HEC_centrality(H)
+    assert c == dict()
+    # Test no edges
+    H.add_nodes_from([0, 1, 2])
+    c = xgi.HEC_centrality(H)
+    assert c == {0: np.NaN, 1: np.NaN, 2: np.NaN}
+    # test disconnected
+    H.add_edge([0, 1])
+    c = xgi.HEC_centrality(H)
+    assert c == {0: np.NaN, 1: np.NaN, 2: np.NaN}
+
     H = xgi.sunflower(3, 1, 5)
     c = H.nodes.hec_centrality(max_iter=1000).asnumpy()
     assert norm(c[1:] - c[1]) < 1e-4
@@ -36,6 +59,16 @@ def test_hec_centrality():
 
 
 def test_node_edge_centrality():
+    # test empty hypergraph
+    H = xgi.Hypergraph()
+    assert xgi.node_edge_centrality(H) == (dict(), dict())
+    # Test no edges
+    H.add_nodes_from([0, 1, 2])
+    assert xgi.node_edge_centrality(H) == ({0: np.NaN, 1: np.NaN, 2: np.NaN}, dict())
+    # test disconnected
+    H.add_edge([0, 1])
+    assert xgi.node_edge_centrality(H) == ({0: np.NaN, 1: np.NaN, 2: np.NaN}, {0: np.NaN})
+
     H = xgi.Hypergraph([[0, 1, 2, 3, 4]])
     c = H.nodes.node_edge_centrality.asnumpy()
     assert norm(c - c[0]) < 1e-6
