### 2023-03-28 8808126
added tests for line_vector_centrality()
