+++ tests/algorithms/test_centrality.py
@@ -28,7 +28,7 @@ def test_cec_centrality():
     assert abs(c[0] / c[1] - ratio(5, 7, kind="CEC")) < 1e-4
 
 
-# @pytest.mark.slow
+@pytest.mark.slow
 def test_hec_centrality():
     # test empty hypergraph
     H = xgi.Hypergraph()
@@ -67,7 +67,10 @@ def test_node_edge_centrality():
     assert xgi.node_edge_centrality(H) == ({0: np.NaN, 1: np.NaN, 2: np.NaN}, dict())
     # test disconnected
     H.add_edge([0, 1])
-    assert xgi.node_edge_centrality(H) == ({0: np.NaN, 1: np.NaN, 2: np.NaN}, {0: np.NaN})
+    assert xgi.node_edge_centrality(H) == (
+        {0: np.NaN, 1: np.NaN, 2: np.NaN},
+        {0: np.NaN},
+    )
 
     H = xgi.Hypergraph([[0, 1, 2, 3, 4]])
     c = H.nodes.node_edge_centrality.asnumpy()
@@ -79,6 +82,28 @@ def test_node_edge_centrality():
     H = xgi.Hypergraph([[0, 1], [1, 2]])
     c = H.edges.node_edge_centrality.asnumpy()
     assert abs(c[0] - c[1]) < 1e-6
+
+
+def test_line_vector_centrality():
+    H = xgi.Hypergraph()
+    c = xgi.line_vector_centrality(H)
+    assert c == dict()
+
+    with pytest.raises(XGIError):
+        H = xgi.Hypergraph()
+        H.add_nodes_from([0, 1, 2])
+        H.add_edge([0, 1])
+        xgi.line_vector_centrality(H)
+
+    H = xgi.sunflower(3, 1, 3) << xgi.sunflower(3, 1, 5)
+    c = xgi.line_vector_centrality(H)
+    assert len(c[0]) == 4  # sizes 2 through 5
+    assert np.allclose(c[0], [0, 0.40824829, 0, 0.24494897])
+    assert set(c.keys()) == set(H.nodes)
+
+    with pytest.raises(Exception):
+        H = xgi.Hypergraph([["a", "b"], ["b", "c"], ["a", "c"]])
+        xgi.line_vector_centrality(H)
 
 
 def ratio(r, m, kind="CEC"):
@@ -112,9 +137,5 @@ def ratio(r, m, kind="CEC"):
     """
     if kind == "CEC":
         return 2 * r * (m - 1) / (np.sqrt(m**2 + 4 * (m - 1) * (r - 1)) + m - 2)
-    elif kind == "ZEC":
-        if m > 4:
-            raise XGIError("Choose a larger m value.")
-        return r**0.5
     elif kind == "HEC":
         return r ** (1.0 / m)
