### 2023-03-28 ec356c6
tests: initial for draw functions
