+++ tests/drawing/test_draw.py
@@ -1,9 +1,11 @@
+import matplotlib.pyplot as plt
 import numpy as np
 import pytest
 from matplotlib import cm
 
 import xgi
 from xgi.drawing.draw import _CCW_sort, _color_arg_to_dict, _scalar_arg_to_dict
+from xgi.exception import XGIError
 
 
 def test_CCW_sort():
@@ -85,3 +87,55 @@ def test_color_arg_to_dict(edgelist4):
     assert np.allclose(d[1], np.array([[0.99692426, 0.89619377, 0.84890427, 1.0]]))
     assert np.allclose(d[2], np.array([[0.98357555, 0.41279508, 0.28835063, 1.0]]))
     assert np.allclose(d[3], np.array([[0.59461745, 0.0461361, 0.07558631, 1.0]]))
+
+
+def test_draw(edgelist8):
+
+    H = xgi.Hypergraph(edgelist8)
+
+    ax = xgi.draw(H)
+
+    assert len(ax.lines) == len(H.edges.filterby("size", 2))  # dyads
+    assert len(ax.patches) == len(H.edges.filterby("size", 2, mode="gt"))  # hyperedges
+    assert len(ax.collections[0].get_sizes()) == H.num_nodes  # nodes
+    plt.close()
+
+
+def test_draw_nodes(edgelist8):
+
+    H = xgi.Hypergraph(edgelist8)
+
+    ax = xgi.draw_nodes(H)
+
+    assert len(ax.lines) == 0  # dyads
+    assert len(ax.patches) == 0  # hyperedges
+    assert len(ax.collections[0].get_sizes()) == H.num_nodes  # nodes
+    plt.close()
+
+
+def test_draw_hyperedges(edgelist8):
+
+    H = xgi.Hypergraph(edgelist8)
+
+    ax = xgi.draw_hyperedges(H)
+
+    assert len(ax.lines) == len(H.edges.filterby("size", 2))  # dyads
+    assert len(ax.patches) == len(H.edges.filterby("size", 2, mode="gt"))  # hyperedges
+    assert len(ax.collections) == 0  # nodes
+    plt.close()
+
+
+def test_draw_simplices(edgelist8):
+
+    with pytest.raises(XGIError):
+        H = xgi.Hypergraph(edgelist8)
+        ax = xgi.draw_simplices(H)
+        plt.close()
+
+    S = xgi.SimplicialComplex(edgelist8)
+    ax = xgi.draw_simplices(S)
+
+    assert len(ax.lines) == 18  # dyads
+    assert len(ax.patches) == 3  # hyperedges
+    assert len(ax.collections) == 0  # nodes
+    plt.close()
