### 2023-03-28 a9813e1
Update test_xgi_data.py
