+++ docs/source/api/stats/xgi.stats.nodestats.rst
@@ -14,4 +14,3 @@
    .. autofunction:: degree
    .. autofunction:: hec_centrality
    .. autofunction:: node_edge_centrality
-   .. autofunction:: zec_centrality
