### 2023-03-28 7d5accb
Centrality update (#313)
* format with black

* added handling for empty and disconnected hypergraphs

* removed ZEC centrality

* added unit tests for centrality

* added tests for line_vector_centrality()

* Update test_xgi_data.py

* removed more instances of ZEC centrality
