+++ tests/algorithms/test_centrality.py
@@ -7,6 +7,16 @@ from xgi.exception import XGIError
 
 
 def test_cec_centrality():
+    # test empty hypergraph
+    H = xgi.Hypergraph()
+    assert xgi.CEC_centrality(H) == dict()
+    # Test no edges
+    H.add_nodes_from([0, 1, 2])
+    assert xgi.CEC_centrality(H) == {0: np.NaN, 1: np.NaN, 2: np.NaN}
+    # test disconnected
+    H.add_edge([0, 1])
+    assert xgi.CEC_centrality(H) == {0: np.NaN, 1: np.NaN, 2: np.NaN}
+
     H = xgi.sunflower(3, 1, 3)
     c = H.nodes.cec_centrality.asnumpy()
     assert norm(c[1:] - c[1]) < 1e-4
@@ -20,6 +30,19 @@ def test_cec_centrality():
 
 @pytest.mark.slow
 def test_hec_centrality():
+    # test empty hypergraph
+    H = xgi.Hypergraph()
+    c = xgi.HEC_centrality(H)
+    assert c == dict()
+    # Test no edges
+    H.add_nodes_from([0, 1, 2])
+    c = xgi.HEC_centrality(H)
+    assert c == {0: np.NaN, 1: np.NaN, 2: np.NaN}
+    # test disconnected
+    H.add_edge([0, 1])
+    c = xgi.HEC_centrality(H)
+    assert c == {0: np.NaN, 1: np.NaN, 2: np.NaN}
+
     H = xgi.sunflower(3, 1, 5)
     c = H.nodes.hec_centrality(max_iter=1000).asnumpy()
     assert norm(c[1:] - c[1]) < 1e-4
@@ -36,6 +59,19 @@ def test_hec_centrality():
 
 
 def test_node_edge_centrality():
+    # test empty hypergraph
+    H = xgi.Hypergraph()
+    assert xgi.node_edge_centrality(H) == (dict(), dict())
+    # Test no edges
+    H.add_nodes_from([0, 1, 2])
+    assert xgi.node_edge_centrality(H) == ({0: np.NaN, 1: np.NaN, 2: np.NaN}, dict())
+    # test disconnected
+    H.add_edge([0, 1])
+    assert xgi.node_edge_centrality(H) == (
+        {0: np.NaN, 1: np.NaN, 2: np.NaN},
+        {0: np.NaN},
+    )
+
     H = xgi.Hypergraph([[0, 1, 2, 3, 4]])
     c = H.nodes.node_edge_centrality.asnumpy()
     assert norm(c - c[0]) < 1e-6
@@ -46,6 +82,28 @@ def test_node_edge_centrality():
     H = xgi.Hypergraph([[0, 1], [1, 2]])
     c = H.edges.node_edge_centrality.asnumpy()
     assert abs(c[0] - c[1]) < 1e-6
+
+
+def test_line_vector_centrality():
+    H = xgi.Hypergraph()
+    c = xgi.line_vector_centrality(H)
+    assert c == dict()
+
+    with pytest.raises(XGIError):
+        H = xgi.Hypergraph()
+        H.add_nodes_from([0, 1, 2])
+        H.add_edge([0, 1])
+        xgi.line_vector_centrality(H)
+
+    H = xgi.sunflower(3, 1, 3) << xgi.sunflower(3, 1, 5)
+    c = xgi.line_vector_centrality(H)
+    assert len(c[0]) == 4  # sizes 2 through 5
+    assert np.allclose(c[0], [0, 0.40824829, 0, 0.24494897])
+    assert set(c.keys()) == set(H.nodes)
+
+    with pytest.raises(Exception):
+        H = xgi.Hypergraph([["a", "b"], ["b", "c"], ["a", "c"]])
[…]
