### 2023-03-29 472c92b
Tests for draw functions (#312)
* tests: clustering. drawing: giving default params values to all functions

* tests: initial for draw functions

* review comments
