+++ xgi/drawing/draw.py
@@ -81,7 +81,7 @@ def draw(
         If a dict, must contain (node_id: width) pairs.  If iterable, assume the widths are
         specified in the same order as the nodes are found in H.nodes. If NodeStat, use a monotonic
         linear interpolation defined between min_node_lw and max_node_lw.
-    node_size : int, float, dict, iterable, or NodeStat (default=10)
+    node_size : int, float, dict, iterable, or NodeStat (default=15)
         Radius of the nodes in pixels.  If int or float, use the same radius for all nodes.
         If a dict, must contain (node_id: radius) pairs.  If iterable, assume the radiuses are
         specified in the same order as the nodes are found in H.nodes. If NodeStat, use a monotonic
@@ -195,33 +195,22 @@ def draw(
     )
 
     # compute axis limits
-    pos_arr = np.asarray([[x, y] for n, (x, y) in pos.items()])
-
-    maxx, maxy = np.max(pos_arr, axis=0)
-    minx, miny = np.min(pos_arr, axis=0)
-    w = maxx - minx
-    h = maxy - miny
-
-    # update view after drawing
-    padx, pady = 0.05 * w, 0.05 * h
-    corners = (minx - padx, miny - pady), (maxx + padx, maxy + pady)
-    ax.update_datalim(corners)
-    ax.autoscale_view()
+    _update_lims(pos, ax)
 
     return ax
 
 
 def draw_nodes(
     H,
-    pos,
-    ax,
-    node_fc,
-    node_ec,
-    node_lw,
-    node_size,
-    zorder,
-    settings,
-    node_labels,
+    pos=None,
+    ax=None,
+    node_fc="white",
+    node_ec="black",
+    node_lw=1,
+    node_size=15,
+    zorder=None,
+    settings=None,
+    node_labels=False,
     **kwargs,
 ):
     """Draw the nodes of a hypergraph
@@ -231,9 +220,10 @@ def draw_nodes(
     H : Hypergraph or SimplicialComplex
         Higher-order network to plot.
     ax : matplotlib.pyplot.axes
-    pos : dict
-        Dictionary of positions node_id:(x,y) of every node.
-    node_fc : str, dict, iterable, or NodeStat
+    pos : dict (default=None)
+        If passed, this dictionary of positions node_id:(x,y) is used for placing the 0-simplices.
+        If None (default), use the `barycenter_spring_layout` to compute the positions.
+    node_fc : str, dict, iterable, or NodeStat (default='white')
         Color of the nodes.  If str, use the same color for all nodes.  If a dict, must
         contain (node_id: color_str) pairs.  If other iterable, assume the colors are
         specified in the same order as the nodes are found in H.nodes. If NodeStat,
@@ -248,7 +238,7 @@ def draw_nodes(
         If a dict, must contain (node_id: width) pairs.  If iterable, assume the widths are
         specified in the same order as the nodes are found in H.nodes. If NodeStat, use a monotonic
         linear interpolation defined between min_node_lw and max_node_lw.
-    node_size : int, float, dict, iterable, or NodeStat (default=10)
+    node_size : int, float, dict, iterable, or NodeStat (default=15)
         Radius of the nodes in pixels.  If int or float, use the same radius for all nodes.
         If a dict, must contain (node_id: radius) pairs.  If iterable, assume the radiuses are
         specified in the same order as the nodes are found in H.nodes. If NodeStat, use a monotonic
@@ -276,6 +266,29 @@ def draw_nodes(
     draw_node_labels
     draw_hyperedge_labels
     """
+
[…]
