### 2023-03-29 90e1a2a
Update draw.py
Convex hull drawing for hypergraphs.
