+++ xgi/drawing/draw.py
@@ -15,6 +15,12 @@ from ..exception import XGIError
 from ..stats import EdgeStat, NodeStat
 from .layout import barycenter_spring_layout
 
+#added for convex hull drawing
+import networkx as nx
+import seaborn as sns
+from scipy.spatial import ConvexHull
+from ..convert import convert_to_graph
+
 __all__ = [
     "draw",
     "draw_nodes",
@@ -985,3 +991,155 @@ def _update_lims(pos, ax):
     corners = (minx - padx, miny - pady), (maxx + padx, maxy + pady)
     ax.update_datalim(corners)
     ax.autoscale_view()
+    
+def draw_hull(node_positions, radius, ax, contour_color, filling_color, alpha_filling):
+    """ Draws an hyperedge as a convex hull encompassing the nodes participating in it
+    
+    
+    Parameters
+    ----------
+    node_positions : np.array
+        array containing the positions (in 2d) of the nodes participating in the hyperedge
+    
+    radius : float
+        radius of the circle created around each node to avoid the covex hull 'touching' the nodes
+    
+    ax : matplotlib.pyplot.axes
+    
+    contour_color : string
+    
+    filling_color : string
+    
+    alpha_filling : float
+
+    Returns
+    -------
+    ax : matplotlib.pyplot.axes
+    """
+    points = node_positions.copy()
+    for i in node_positions:
+        center = i
+        circle = [center + radius * np.array([np.cos(theta), np.sin(theta)]) for theta in np.linspace(0, 2 * np.pi, num=100, endpoint=False)]
+        points = np.vstack([circle, points])
+    
+    hull = ConvexHull(points)
+
+    for simplex in hull.simplices:
+        plt.plot(points[simplex,0], points[simplex,1], color=contour_color)
+    plt.fill(points[hull.vertices,0], points[hull.vertices,1], color=filling_color, alpha=alpha_filling)
+
+    return ax
+
+def draw_hypergraph_hull(
+        H, 
+        pos=None, 
+        ax=None, 
+        max_order=None, 
+        radius=0.05, 
+        alpha_filling=0.4, 
+        palette=None, 
+        node_size=60, 
+        node_color='tab:blue', 
+        node_shape='o', 
+        node_linewidths=1, 
+        node_edgecolors='black',
+        dyad_color='black'
+
+):
+    """Draw hypergraph with hyperedges of >3 elements displayed as convex hull
+
+    Parameters
+    ----------
+    H : Hypergraph
+    
+    pos : dict (default=None)
+        If passed, this dictionary of positions node_id:(x,y) is used for placing the nodes.
+        If None (default), use the `barycenter_spring_layout` to compute the positions.
+    
+    ax : matplotlib.pyplot.axes (default=None)
+    
+    max_order : int (default=None)
+        Maximum of hyperedges to plot. If None (default), plots all orders.
+    
+    radius : flot (default = 0.05)
+        radius of the convex hull in the vicinity of the nodes.
+    
+    alpha_filling : float (default=0.4)
+        The hyperedges transparency.
+    palette : None or array of colors (default=None)
+        Array of colors given by the hyperedges of different orders
+    
+    node_size : scalar or array (default=300)
+        Size of nodes.  If an array it must be the same length as nodelist.
+    
+    node_color : color or array of colors (default='tab:blue')
+        Node color. Can be a single color or a sequence of colors with the same
+        length as nodelist. Color can be string or rgb (or rgba) tuple of
+        floats from 0-1. If numeric values are specified they will be
+        mapped to colors using the cmap and vmin,vmax parameters. See
+        matplotlib.scatter for more details.
+    
[…]
