### 2023-03-29 9a2e3be
Hypergraph hull drawing
created .ipynb tutorial and minor update to draw.py
