+++ xgi/drawing/draw.py
@@ -994,114 +994,54 @@ def _update_lims(pos, ax):
     ax.update_datalim(corners)
     ax.autoscale_view()
     
-def draw_hull(node_positions, radius, ax, contour_color, filling_color, alpha_filling):
-    """ Draws an hyperedge as a convex hull encompassing the nodes participating in it
-    
-    
-    Parameters
-    ----------
-    node_positions : np.array
-        array containing the positions (in 2d) of the nodes participating in the hyperedge
-    
-    radius : float
-        radius of the circle created around each node to avoid the covex hull 'touching' the nodes
-    
-    ax : matplotlib.pyplot.axes
-    
-    contour_color : string
-    
-    filling_color : string
-    
-    alpha_filling : float
-
-    Returns
-    -------
-    ax : matplotlib.pyplot.axes
-    """
-    points = node_positions.copy()
-    for i in node_positions:
+def _draw_hull(node_pos, radius, ax, edges_ec, facecolor, alpha):
+    #add_description
+    points = node_pos.copy()
+    for i in node_pos:
         center = i
-        circle = [center + radius * np.array([np.cos(theta), np.sin(theta)]) for theta in np.linspace(0, 2 * np.pi, num=100, endpoint=False)]
+        thetas = np.linspace(0, 2 * np.pi, num=100, endpoint=False)
+        circle = [center + radius * np.array([np.cos(theta), np.sin(theta)]) for theta in thetas]
         points = np.vstack([circle, points])
     
     hull = ConvexHull(points)
 
     for simplex in hull.simplices:
-        plt.plot(points[simplex,0], points[simplex,1], color=contour_color)
-    plt.fill(points[hull.vertices,0], points[hull.vertices,1], color=filling_color, alpha=alpha_filling)
+        ax.plot(points[simplex,0], points[simplex,1], color=edges_ec)
+    ax.fill(points[hull.vertices,0], points[hull.vertices,1], color=facecolor, alpha=alpha)
 
     return ax
 
 def draw_hypergraph_hull(
-        H, 
-        pos=None, 
-        ax=None, 
-        max_order=None, 
-        radius=0.05, 
-        alpha_filling=0.4, 
-        palette=None, 
-        node_size=60, 
-        node_color='tab:blue', 
-        node_shape='o', 
-        node_linewidths=1, 
-        node_edgecolors='black',
-        dyad_color='black'
-
+        H,
+        pos=None,
+        ax=None,
+        dyad_color = "black",
+        node_fc="white",
+        node_ec="black",
+        node_lw=1,
+        node_size=15,
+        edge_fc=None,
+        edges_ec="tab:gray",
+        max_order=None,
+        node_labels=False,
+        radius=0.05,
+        **kwargs
 ):
-    """Draw hypergraph with hyperedges of >3 elements displayed as convex hull
+    #add description
+    settings = {
+        "edge_fc_cmap": cm.Blues,
+        "alpha": 0.4
+    }
+    
 
-    Parameters
-    ----------
-    H : Hypergraph
+    alpha = settings["alpha"]
     
-    pos : dict (default=None)
-        If passed, this dictionary of positions node_id:(x,y) is used for placing the nodes.
-        If None (default), use the `barycenter_spring_layout` to compute the positions.
+    if edge_fc is None:
+        edge_fc = H.edges.size
+        
+    edge_fc = _color_arg_to_dict(edge_fc, H.edges, settings["edge_fc_cmap"])
     
-    ax : matplotlib.pyplot.axes (default=None)
-    
-    max_order : int (default=None)
-        Maximum of hyperedges to plot. If None (default), plots all orders.
-    
-    radius : flot (default = 0.05)
-        radius of the convex hull in the vicinity of the nodes.
-    
-    alpha_filling : float (default=0.4)
[…]
