+++ xgi/drawing/draw.py
@@ -16,10 +16,7 @@ from ..stats import EdgeStat, NodeStat
 from .layout import barycenter_spring_layout
 
 #added for convex hull drawing
-import networkx as nx
-import seaborn as sns
 from scipy.spatial import ConvexHull
-from ..convert import convert_to_graph
 
 __all__ = [
     "draw",
@@ -993,7 +990,28 @@ def _update_lims(pos, ax):
     ax.update_datalim(corners)
     ax.autoscale_view()
     
-def _draw_hull(node_pos, radius, ax, edges_ec, facecolor, alpha):
+def _draw_hull(node_pos, ax, edges_ec, facecolor, alpha, radius):
+    """Draw a convex hull encompassing the nodes in node_pos
+
+    Parameters
+    ----------
+    node_pos : np.array
+        nx2 dimensional array containing positions of the nodes
+    ax : matplotlib.pyplot.axes
+    edges_ec : str
+        Color of the border of the convex hull
+    facecolor : str
+        Filling color of the convex hull
+    alpha : float
+        Transparency of the convex hull
+    radius : float
+        Radius of the convex hull in the vicinity of the nodes.
+
+    Returns
+    -------
+    ax : matplotlib.pyplot.axes
+
+    """
     #add_description
     points = node_pos.copy()
     for i in node_pos:
@@ -1015,18 +1033,88 @@ def draw_hypergraph_hull(
         pos=None,
         ax=None,
         dyad_color = "black",
+        edge_fc=None,
+        edges_ec="tab:gray",
         node_fc="tab:blue",
         node_ec="black",
         node_lw=1,
         node_size=7,
-        edge_fc=None,
-        edges_ec="tab:gray",
         max_order=None,
         node_labels=False,
         radius=0.05,
         **kwargs
 ):
-    #add description
+    """Draw hypergraphs displaying the hyperedges of order k>1 as convex hulls
+    
+
+    Parameters
+    ----------
+    H : Hypergraph
+    pos : dict (default=None)
+        If passed, this dictionary of positions node_id:(x,y) is used for placing the nodes.
+        If None (default), use the `barycenter_spring_layout` to compute the positions.
+    ax : matplotlib.pyplot.axes (default=None)
+    dyad_color : str, dict, iterable, or EdgeStat (default='black')
+        Color of the dyadic links.  If str, use the same color for all edges. If a dict, must
+        contain (edge_id: color_str) pairs.  If iterable, assume the colors are
+        specified in the same order as the edges are found in H.edges. If EdgeStat, use a colormap
+        (specified with dyad_color_cmap) associated to it.
+    edge_fc : str, dict, iterable, or EdgeStat (default=None)
+        str, dict, iterable, or EdgeStat (default=None)
+        Color of the hyperedges.  If str, use the same color for all nodes.  If a dict, must
+        contain (edge_id: color_str) pairs.  If other iterable, assume the colors are
+        specified in the same order as the hyperedges are found in H.edges. If EdgeStat,
+        use the colormap specified with edge_fc_cmap. If None (default),
+        use the H.edges.size.
+    edges_ec : str, dict, iterable, or EdgeStat (default='black')
+        Color of the borders of the hyperdges of order k>1.  If str, use the same color for all edges. If a dict, must
+        contain (edge_id: color_str) pairs.  If iterable, assume the colors are
+        specified in the same order as the edges are found in H.edges. If EdgeStat, use a colormap
+        (specified with dyad_color_cmap) associated to it.
+    node_fc : node_fc : str, dict, iterable, or NodeStat (default='tab:blue')
+        Color of the nodes.  If str, use the same color for all nodes.  If a dict, must
[…]
