### 2023-03-30 a3338fd
Update draw.py
added edge_ec
