+++ xgi/drawing/draw.py
@@ -1061,16 +1061,17 @@ def draw_hypergraph_hull(
         (specified with dyad_color_cmap) associated to it.
     edge_fc : str, dict, iterable, or EdgeStat (default=None)
         str, dict, iterable, or EdgeStat (default=None)
-        Color of the hyperedges.  If str, use the same color for all nodes.  If a dict, must
+        Color of the hyperedges of order k>1.  If str, use the same color for all hyperedges of order k>1.  If a dict, must
         contain (edge_id: color_str) pairs.  If other iterable, assume the colors are
         specified in the same order as the hyperedges are found in H.edges. If EdgeStat,
         use the colormap specified with edge_fc_cmap. If None (default),
         use the H.edges.size.
-    edge_ec : str, dict, iterable, or EdgeStat (default='black')
+    edge_ec : str, dict, iterable, or EdgeStat (default='tab:gray')
         Color of the borders of the hyperdges of order k>1.  If str, use the same color for all edges. If a dict, must
         contain (edge_id: color_str) pairs.  If iterable, assume the colors are
         specified in the same order as the edges are found in H.edges. If EdgeStat, use a colormap
-        (specified with dyad_color_cmap) associated to it.
+        (specified with edge_ec_cmap) associated to it. If None,
+        use the H.edges.size.
     node_fc : node_fc : str, dict, iterable, or NodeStat (default='tab:blue')
         Color of the nodes.  If str, use the same color for all nodes.  If a dict, must
         contain (node_id: color_str) pairs.  If other iterable, assume the colors are
