### 2023-03-30 480d9bb
Added normalized laplacian
