### 2023-03-30 c79775c
Added normalized laplacian (#314)
* Added normalized laplacian
