+++ docs/source/api/algorithms/xgi.algorithms.clustering.rst
@@ -1,0 +1,12 @@
+xgi.algorithms.clustering
+=========================
+
+.. currentmodule:: xgi.algorithms.clustering
+
+.. automodule:: xgi.algorithms.clustering
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: local_clustering_coefficient
+   .. autofunction:: clustering_coefficient
+   .. autofunction:: two_node_clustering_coefficient
