+++ tests/algorithms/test_clustering.py
@@ -1,0 +1,124 @@
+import numpy as np
+import pytest
+
+import xgi
+from xgi.exception import XGIError
+
+
+def test_local_clustering_coefficient(edgelist8):
+    H = xgi.random_hypergraph(3, [1])
+
+    cc = xgi.local_clustering_coefficient(H)
+    true_cc = {0: 1.0, 1: 1.0, 2: 1.0}
+    assert cc == true_cc
+
+    H = xgi.random_hypergraph(3, [1, 1])
+    cc = xgi.local_clustering_coefficient(H)
+    true_cc = {0: 1.0, 1: 1.0, 2: 1.0}
+    assert cc == true_cc
+
+    H = xgi.random_hypergraph(3, [0, 1])
+    cc = xgi.local_clustering_coefficient(H)
+    true_cc = {0: 0.0, 1: 0.0, 2: 0.0}
+    assert cc == true_cc
+
+    H = xgi.Hypergraph()
+    cc = xgi.local_clustering_coefficient(H)
+    assert cc == {}
+
+    H = xgi.Hypergraph()
+    H.add_nodes_from(range(3))
+    cc = xgi.local_clustering_coefficient(H)
+    assert cc == {0: np.NaN, 1: np.NaN, 2: np.NaN}
+
+    H = xgi.Hypergraph(edgelist8)
+    cc = xgi.local_clustering_coefficient(H)
+    true_cc = {
+        0: 0.6777777777777778,
+        1: 0.575,
+        2: 0.3333333333333333,
+        3: 0.3333333333333333,
+        4: 0.6666666666666666,
+        5: 0.0,
+        6: 0.0,
+    }
+    assert cc == true_cc
+
+
+def test_clustering_coefficient(edgelist1):
+    H = xgi.random_hypergraph(3, [1])
+
+    cc = xgi.clustering_coefficient(H)
+    true_cc = {0: 1.0, 1: 1.0, 2: 1.0}
+    assert cc == true_cc
+
+    H = xgi.random_hypergraph(3, [1, 1])
+    cc = xgi.clustering_coefficient(H)
+    true_cc = {0: 1.0, 1: 1.0, 2: 1.0}
+    assert cc == true_cc
+
+    H = xgi.random_hypergraph(3, [0, 1])
+    cc = xgi.clustering_coefficient(H)
+    true_cc = {0: 1.0, 1: 1.0, 2: 1.0}
+    assert cc == true_cc
+
+    H = xgi.Hypergraph()
+    cc = xgi.clustering_coefficient(H)
+    assert cc == {}
+
+    H = xgi.Hypergraph()
+    H.add_nodes_from(range(3))
+    cc = xgi.clustering_coefficient(H)
+    assert cc == {0: np.NaN, 1: np.NaN, 2: np.NaN}
+
+    H = xgi.Hypergraph(edgelist1)
+    cc = xgi.clustering_coefficient(H)
+    true_cc = {1: 1.0, 2: 1.0, 3: 1.0, 4: np.NaN, 5: np.NaN, 6: 1 / 3, 8: 1.0, 7: 1.0}
+    assert cc == true_cc
+
+
+def test_two_node_clustering_coefficient(edgelist1, edgelist8):
+    H = xgi.random_hypergraph(3, [1])
+
+    cc = xgi.two_node_clustering_coefficient(H)
+    true_cc = {0: 1 / 3, 1: 1 / 3, 2: 1 / 3}
+    assert cc == true_cc
+
+    # check default keyword
+    cc1 = xgi.two_node_clustering_coefficient(H, kind="union")
+    assert cc == cc1
+
[…]
