### 2023-04-02 68e53dc
Added line graph tests. Fixes #298.
