### 2023-04-03 4148d07
Update xgi/algorithms/clustering.py
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
