### 2023-04-03 6a2a280
Test line graph (#318)
* Added line graph tests. Fixes #298.

* added s-parameter
