+++ tests/test_convert.py
@@ -1,5 +1,6 @@
 import pandas as pd
 import pytest
+from networkx import Graph
 
 import xgi
 from xgi.exception import XGIError
@@ -179,3 +180,36 @@ def test_from_bipartite_graph(
     # not bipartite
     with pytest.raises(XGIError):
         H = xgi.from_bipartite_graph(bipartite_graph4, dual=True)
+
+
+def test_to_line_graph(edgelist1, hypergraph1):
+    H = xgi.Hypergraph(edgelist1)
+    L = xgi.to_line_graph(H)
+
+    assert isinstance(L, Graph)
+    assert set(L.nodes) == {0, 1, 2, 3}
+    assert [set(e) for e in L.edges] == [{2, 3}]
+
+    L = xgi.to_line_graph(hypergraph1)
+    assert set(L.nodes) == {"e1", "e2", "e3"}
+    assert [set(e) for e in L.edges] == [{"e1", "e2"}, {"e2", "e3"}]
+
+    L = xgi.to_line_graph(hypergraph1,s=2)
+    assert set(L.nodes) == {"e1", "e2", "e3"}
+    assert [set(e) for e in L.edges] == [{"e1", "e2"}]
+
+    L = xgi.to_line_graph(hypergraph1,s=3)
+    assert set(L.nodes) == {"e1", "e2", "e3"}
+    assert [set(e) for e in L.edges] == []
+
+    H = xgi.Hypergraph()
+    L = xgi.to_line_graph(H)
+
+    assert L.number_of_nodes() == 0
+    assert L.number_of_edges() == 0
+
+    H.add_nodes_from([0, 1, 2])
+    L = xgi.to_line_graph(H)
+
+    assert L.number_of_nodes() == 0
+    assert L.number_of_edges() == 0
