### 2023-04-04 e757e39
refact: new folder structure for linalg
