+++ xgi/linalg/laplacian_matrix.py
@@ -1,0 +1,234 @@
+"""Matrices associated to hypergraphs.
+
+Note that the order of the rows and columns of the
+matrices in this module correspond to the order in
+which nodes/edges are added to the hypergraph or
+simplicial complex. If the node and edge IDs are
+able to be sorted, the following is an example to sort
+by the node and edge IDs.
+
+>>> import xgi
+>>> import pandas as pd
+>>> H = xgi.Hypergraph([[1, 2, 3, 7], [4], [5, 6, 7]])
+>>> I, nodedict, edgedict = xgi.incidence_matrix(H, sparse=False, index=True)
+>>> # Sorting the resulting numpy array:
+>>> sortedI = I.copy()
+>>> sortedI = sortedI[sorted(nodedict, key=nodedict.get), :]
+>>> sortedI = sortedI[:, sorted(edgedict, key=edgedict.get)]
+>>> sortedI
+array([[1, 0, 0],
+       [1, 0, 0],
+       [1, 0, 0],
+       [0, 1, 0],
+       [0, 0, 1],
+       [0, 0, 1],
+       [1, 0, 1]])
+>>> # Indexing a Pandas dataframe by the node/edge IDs
+>>> df = pd.DataFrame(I, index=nodedict.values(), columns=edgedict.values())
+
+If the nodes are already sorted, this order can be preserved by adding the nodes
+to the hypergraph prior to adding edges. For example,
+
+>>> import xgi
+>>> H = xgi.Hypergraph()
+>>> H.add_nodes_from(range(1, 8))
+>>> H.add_edges_from([[1, 2, 3, 7], [4], [5, 6, 7]])
+>>> xgi.incidence_matrix(H, sparse=False)
+array([[1, 0, 0],
+       [1, 0, 0],
+       [1, 0, 0],
+       [0, 1, 0],
+       [0, 0, 1],
+       [0, 0, 1],
+       [1, 0, 1]])
+
+"""
+from warnings import warn
+
+import numpy as np
+from scipy.sparse import csr_array, diags
+
+from ..exception import XGIError
+from .hypergraph_matrix import adjacency_matrix, clique_motif_matrix, degree_matrix
+
+__all__ = [
+    "laplacian",
+    "multiorder_laplacian",
+    "normalized_hypergraph_laplacian",
+]
+
+
+def laplacian(H, order=1, sparse=False, rescale_per_node=False, index=False):
+    """Laplacian matrix of order d, see [1].
+
+    Parameters
+    ----------
+    HG : Hypergraph
+        Hypergraph
+    order : int
+        Order of interactions to consider. If order=1 (default),
+        returns the usual graph Laplacian
+    sparse: bool, default: False
+        Specifies whether the output matrix is a scipy sparse matrix or a numpy matrix
+    index: bool, default: False
+        Specifies whether to output disctionaries mapping the node and edge IDs to indices
+
+    Returns
+    -------
+    L_d : numpy array
+        Array of dim (N, N)
+    if index is True:
+        return rowdict
+
+    See also
+    --------
+    multiorder_laplacian
+
+    References
+    ----------
+    .. [1] Lucas, M., Cencetti, G., & Battiston, F. (2020).
+        Multiorder Laplacian for synchronization in higher-order networks.
+        Physical Review Research, 2(3), 033410.
+
+    """
+    A, row_dict = adjacency_matrix(
+        H, order=order, sparse=sparse, weighted=True, index=True
+    )
+
+    if A.shape == (0, 0):
+        L = csr_array((0, 0)) if sparse else np.empty((0, 0))
+        return (L, {}) if index else L
+
+    if sparse:
+        K = csr_array(diags(degree_matrix(H, order=order)))
[…]

[…]
