### 2023-04-04 e92c0ca
removed for loop for circles around nodes
