### 2023-04-04 e09f209
refact: new folder structure for generators step 1g
