+++ xgi/generators/classic.py
@@ -5,25 +5,10 @@ hypergraph).
 
 """
 
-import random
-from collections import defaultdict
-from itertools import combinations
-from warnings import warn
-
-import networkx as nx
-
-from ..classes.function import subfaces
-from ..exception import XGIError
-from ..utils.utilities import find_triangles
 
 __all__ = [
     "empty_hypergraph",
     "empty_simplicial_complex",
-    "star_clique",
-    "flag_complex",
-    "flag_complex_d2",
-    "sunflower",
-    "ring_lattice",
 ]
 
 
@@ -114,282 +99,3 @@ def empty_simplicial_complex(create_using=None, default=None):
     if default is None:
         default = xgi.SimplicialComplex
     return _empty_network(create_using, default)
-
-
-def star_clique(n_star, n_clique, d_max):
-    """Generate a star-clique structure
-
-    That is a star network and a clique network,
-    connected by one pairwise edge connecting the centre of the star to the clique.
-    network, the each clique is promoted to a hyperedge
-    up to order d_max.
-
-    Parameters
-    ----------
-    n_star : int
-        Number of legs of the star
-    n_clique : int
-        Number of nodes in the clique
-    d_max : int
-        Maximum order up to which to promote
-        cliques to hyperedges
-
-    Returns
-    -------
-    H : Hypergraph
-
-    Examples
-    --------
-    >>> import xgi
-    >>> H = xgi.star_clique(6, 7, 2)
-
-    Notes
-    -----
-    The total number of nodes is n_star + n_clique.
-
-    """
-
-    if n_star <= 0:
-        raise ValueError("n_star must be an integer > 0.")
-    if n_clique <= 0:
-        raise ValueError("n_clique must be an integer > 0.")
-    if d_max < 0:
-        raise ValueError("d_max must be an integer >= 0.")
-    elif d_max > n_clique - 1:
-        raise ValueError("d_max must be <= n_clique - 1.")
-
-    nodes_star = range(n_star)
-    nodes_clique = range(n_star, n_star + n_clique)
-    nodes = list(nodes_star) + list(nodes_clique)
-
-    H = empty_hypergraph()
-    H.add_nodes_from(nodes)
-
-    # add star edges (center of the star is 0-th node)
-    H.add_edges_from([(nodes_star[0], nodes_star[i]) for i in range(1, n_star)])
-
-    # connect clique and star by adding last star leg
-    H.add_edge((nodes_star[0], nodes_clique[0]))
-
-    # add clique hyperedges up to order d_max
-    H.add_edges_from(
-        [e for d in range(1, d_max + 1) for e in combinations(nodes_clique, d + 1)]
-    )
-
-    return H
-
-
-def flag_complex(G, max_order=2, ps=None, seed=None):
-    """Generate a flag (or clique) complex from a
-    NetworkX graph by filling all cliques up to dimension max_order.
-
-    Parameters
-    ----------
-    G : Networkx Graph
-
-    max_order : int
-        maximal dimension of simplices to add to the output simplicial complex
-    ps: list of float
-        List of probabilities (between 0 and 1) to create a
-        hyperedge from a clique, at each order d. For example,
-        ps[0] is the probability of promoting any 3-node clique (triangle) to
-        a 3-hyperedge.
-    seed: int or None (default)
-        The seed for the random number generator
-
-    Returns
-    -------
-    S : SimplicialComplex
-
-    Notes
-    -----
-    Computing all cliques quickly becomes heavy for large networks. `flag_complex_d2`
-    is faster to compute up to order 2.
-
-    See also
-    --------
-    flag_complex_d2
-
-    """
-    # This import needs to happen when this function is called, not when it is
[…]
