+++ codecov.yml
@@ -1,14 +1,12 @@
 coverage:
   precision: 2
   round: down
-  range: 60..85
+  range: 70..100
   status:
     project:
       default:
         # basic
         target: auto
         threshold: 0.1
-    patch:
-      default:
-        # basic
-        target: 0
+        removed_code_behavior: removals_only # off, removals_only, adjust_base
+    patch: off
