+++ xgi/generators/nonuniform.py
@@ -1,471 +1,0 @@
-"""Generate random (non-uniform) hypergraphs."""
-import math
-import random
-import warnings
-from collections import defaultdict
-from itertools import combinations
-
-import networkx as nx
-import numpy as np
-from scipy.special import comb
-
-from ..classes import SimplicialComplex
-from .classic import empty_hypergraph
-from .lattice import ring_lattice
-from .simplicial_complexes import flag_complex_d2
-
-__all__ = [
-    "chung_lu_hypergraph",
-    "dcsbm_hypergraph",
-    "random_hypergraph",
-    "random_simplicial_complex",
-    "random_flag_complex_d2",
-    "random_flag_complex",
-    "watts_strogatz_hypergraph",
-]
-
-
-def chung_lu_hypergraph(k1, k2, seed=None):
-    """A function to generate a Chung-Lu hypergraph
-
-    Parameters
-    ----------
-    k1 : dictionary
-        Dictionary where the keys are node ids
-        and the values are node degrees.
-    k2 : dictionary
-        Dictionary where the keys are edge ids
-        and the values are edge sizes.
-    seed : integer or None (default)
-            The seed for the random number generator.
-
-    Returns
-    -------
-    Hypergraph object
-        The generated hypergraph
-
-    Warns
-    -----
-    warnings.warn
-        If the sums of the edge sizes and node degrees are not equal, the
-        algorithm still runs, but raises a warning.
-
-    Notes
-    -----
-    The sums of k1 and k2 should be the same. If they are not the same,
-    this function returns a warning but still runs.
-
-    References
-    ----------
-    Implemented by Mirah Shi in HyperNetX and described for
-    bipartite networks by Aksoy et al. in https://doi.org/10.1093/comnet/cnx001
-
-    Example
-    -------
-    >>> import xgi
-    >>> import random
-    >>> n = 100
-    >>> k1 = {i : random.randint(1, 100) for i in range(n)}
-    >>> k2 = {i : sorted(k1.values())[i] for i in range(n)}
-    >>> H = xgi.chung_lu_hypergraph(k1, k2)
-
-    """
-    if seed is not None:
-        random.seed(seed)
-
-    # sort dictionary by degree in decreasing order
-    node_labels = [n for n, _ in sorted(k1.items(), key=lambda d: d[1], reverse=True)]
-    edge_labels = [m for m, _ in sorted(k2.items(), key=lambda d: d[1], reverse=True)]
-
-    m = len(k2)
-
-    if sum(k1.values()) != sum(k2.values()):
-        warnings.warn(
-            "The sum of the degree sequence does not match the sum of the size sequence"
-        )
-
-    S = sum(k1.values())
-
-    H = empty_hypergraph()
-    H.add_nodes_from(node_labels)
-
-    for u in node_labels:
-        j = 0
-        v = edge_labels[j]  # start from beginning every time
-        p = min((k1[u] * k2[v]) / S, 1)
-
-        while j < m:
-            if p != 1:
-                r = random.random()
-                try:
-                    j = j + math.floor(math.log(r) / math.log(1 - p))
-                except ZeroDivisionError:
-                    j = np.inf
-
-            if j < m:
-                v = edge_labels[j]
-                q = min((k1[u] * k2[v]) / S, 1)
-                r = random.random()
-                if r < q / p:
-                    # no duplicates
-                    H.add_node_to_edge(v, u)
-                p = q
-                j = j + 1
-
-    return H
-
-
-def dcsbm_hypergraph(k1, k2, g1, g2, omega, seed=None):
-    """A function to generate a DCSBM hypergraph.
-
-    Parameters
-    ----------
-    k1 : dict
-        This is a dictionary where the keys are node ids
-        and the values are node degrees.
-    k2 : dict
[…]
