+++ tests/generators/test_classic.py
@@ -1,8 +1,4 @@
-import networkx as nx
-import pytest
-
 import xgi
-from xgi.exception import XGIError
 
 
 def test_empty_hypergraph():
@@ -13,110 +9,3 @@ def test_empty_hypergraph():
 def test_empty_hypergraph():
     SC = xgi.empty_simplicial_complex()
     assert (SC.num_nodes, SC.num_edges) == (0, 0)
-
-
-def test_star_clique():
-    with pytest.raises(ValueError):
-        H = xgi.star_clique(-1, 7, 3)
-    with pytest.raises(ValueError):
-        H = xgi.star_clique(6, -1, 3)
-    with pytest.raises(ValueError):
-        H = xgi.star_clique(6, 7, -1)
-    with pytest.raises(ValueError):
-        H = xgi.star_clique(6, 7, 7)
-
-    H = xgi.star_clique(6, 7, 3)
-    assert H.num_nodes == 13
-    assert H.num_edges == 97
-    assert xgi.max_edge_order(H) == 3
-
-
-def test_flag_complex():
-    edges = [[0, 1], [1, 2], [2, 0], [0, 3]]
-    G = nx.Graph(edges)
-
-    S = xgi.flag_complex(G)
-
-    simplices_2 = [
-        frozenset({0, 1}),
-        frozenset({0, 2}),
-        frozenset({0, 3}),
-        frozenset({1, 2}),
-    ]
-
-    simplices_3 = simplices_2 + [frozenset({0, 1, 2})]
-
-    assert S.edges.members() == simplices_3
-
-    S1 = xgi.flag_complex(G, ps=[1], seed=42)
-    S2 = xgi.flag_complex(G, ps=[0.5], seed=42)
-    S3 = xgi.flag_complex(G, ps=[0], seed=42)
-
-    assert S1.edges.members() == simplices_3
-    assert S2.edges.members() == simplices_2
-    assert S3.edges.members() == simplices_2
-
-    G1 = nx.complete_graph(4)
-    S4 = xgi.flag_complex(G1)
-    S5 = xgi.flag_complex(G1, ps=[1])
-    assert S4.num_nodes == S5.num_nodes
-    assert S4.num_edges == S5.num_edges
-    assert set(S4.edges.members()) == set(S5.edges.members())
-
-
-def test_flag_complex_d2():
-    G = nx.erdos_renyi_graph(15, 0.3, seed=3)
-
-    S = xgi.flag_complex(G, max_order=2)
-    S2 = xgi.flag_complex_d2(G)
-
-    assert set(S.edges.members()) == set(S2.edges.members())
-
-
-def test_ring_lattice():
-    H = xgi.ring_lattice(5, 2, 2, 0)
-    assert H.num_nodes == 5
-    assert H.num_edges == 5
-    assert xgi.unique_edge_sizes(H) == [2]
-
-    H = xgi.ring_lattice(5, 3, 4, 1)
-    edges = H.edges.members()
-    for i in range(H.num_edges - 1):
-        assert len(set(edges[i]).intersection(set(edges[i + 1]))) == 2  # d-l
-    assert xgi.unique_edge_sizes(H) == [3]
-
-    # k < 2 test
-    with pytest.warns(Warning):
-        H = xgi.ring_lattice(5, 2, 1, 0)
-    assert H.num_nodes == 5
-    assert H.num_edges == 0
-
-    # k % 2 != 0 test
-    with pytest.warns(Warning):
-        xgi.ring_lattice(5, 2, 3, 0)
-
-    # k < 0 test
-    with pytest.raises(XGIError):
-        xgi.ring_lattice(5, 2, -1, 0)
-
-
-def test_sunflower():
[…]
