### 2023-04-04 06aa7e2
Update xgi/drawing/draw.py
Co-authored-by: Leo Torres <dleonardotn@gmail.com>
