### 2023-04-04 5c012cf
corrected error in faster implementation
