+++ docs/source/api/generators.rst
@@ -8,5 +8,9 @@ generators package
    :toctree: generators
 
    ~xgi.generators.classic
-   ~xgi.generators.nonuniform
+   ~xgi.generators.simple
+   ~xgi.generators.lattice
+   ~xgi.generators.random
    ~xgi.generators.uniform
+   ~xgi.generators.simplicial_complexes
+   
