+++ codecov.yml
@@ -1,7 +1,7 @@
 coverage:
   precision: 2
   round: down
-  range: 70..100
+  range: 60..85
   status:
     project:
       default:
