+++ docs/source/api/stats/xgi.stats.nodestats.rst
@@ -10,7 +10,9 @@
    .. autofunction:: attrs
    .. autofunction:: average_neighbor_degree
    .. autofunction:: clique_eigenvector_centrality
-   .. autofunction:: clustering
+   .. autofunction:: clustering_coefficient
+   .. autofunction:: local_clustering_coefficient
+   .. autofunction:: two_node_clustering_coefficient
    .. autofunction:: degree
    .. autofunction:: h_eigenvector_centrality
    .. autofunction:: node_edge_centrality
