+++ docs/source/api/stats/xgi.stats.nodestats.rst
@@ -11,8 +11,8 @@
    .. autofunction:: average_neighbor_degree
    .. autofunction:: clique_eigenvector_centrality
    .. autofunction:: clustering_coefficient
-   .. autofunction:: local_clustering_coefficient
-   .. autofunction:: two_node_clustering_coefficient
    .. autofunction:: degree
    .. autofunction:: h_eigenvector_centrality
+   .. autofunction:: local_clustering_coefficient
    .. autofunction:: node_edge_centrality
+   .. autofunction:: two_node_clustering_coefficient
