### 2023-04-05 b77c206
Optionally return the index-to-unit mappings from to_bipartite_graph. The default has been set to False to mirror the behavior of linalg.
