### 2023-04-05 ec7903e
changed NaN to nan
https://stackoverflow.com/questions/53436339/difference-between-np-nan-and-np-nan
