### 2023-04-05 759114e
Drawing hypergraphs with hyperdedges as convex hulls  (#320)
* Update draw.py

Convex hull drawing for hypergraphs.

* Hypergraph hull drawing

created .ipynb tutorial and minor update to draw.py

* update draw hull

* Update convex_hull_drawing_example.ipynb

docs

added edge_ec

modify doc

Update tutorial

hyperedges labelss


line 1180: zorder = max_order -1

allow controlling the zorder of the hulls

* updated tutorials

move convex_hull_drawing_example.ipynb to tutorials

* Update convex_hull_drawing_example.ipynb

minor update to tutorial

* Update Tutorial 5 - Plotting.ipynb

minor upd
[…]
