+++ xgi/drawing/draw.py
@@ -8,6 +8,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 from matplotlib import cm
 from matplotlib.colors import LinearSegmentedColormap, ListedColormap
+from scipy.spatial import ConvexHull
 
 from .. import convert
 from ..classes import Hypergraph, SimplicialComplex, max_edge_order
@@ -22,6 +23,7 @@ __all__ = [
     "draw_simplices",
     "draw_node_labels",
     "draw_hyperedge_labels",
+    "draw_hypergraph_hull",
 ]
 
 
@@ -985,3 +987,241 @@ def _update_lims(pos, ax):
     corners = (minx - padx, miny - pady), (maxx + padx, maxy + pady)
     ax.update_datalim(corners)
     ax.autoscale_view()
+
+
+def _draw_hull(node_pos, ax, edges_ec, facecolor, alpha, zorder, radius):
+    """Draw a convex hull encompassing the nodes in node_pos
+
+    Parameters
+    ----------
+    node_pos : np.array
+        nx2 dimensional array containing positions of the nodes
+    ax : matplotlib.pyplot.axes
+    edges_ec : str
+        Color of the border of the convex hull
+    facecolor : str
+        Filling color of the convex hull
+    alpha : float
+        Transparency of the convex hull
+    radius : float
+        Radius of the convex hull in the vicinity of the nodes.
+
+    Returns
+    -------
+    ax : matplotlib.pyplot.axes
+
+    """
+    
+    thetas = np.linspace(0, 2*np.pi, num=100, endpoint=False)
+    offsets = radius * np.array([np.cos(thetas), np.sin(thetas)]).T
+    points = np.vstack([p + offsets for p in node_pos])
+    points = np.vstack([node_pos, points])
+
+    hull = ConvexHull(points)
+
+    for simplex in hull.simplices:
+        ax.plot(points[simplex, 0], points[simplex, 1], color=edges_ec, zorder=zorder)
+    ax.fill(
+        points[hull.vertices, 0],
+        points[hull.vertices, 1],
+        color=facecolor,
+        alpha=alpha,
+        zorder=zorder,
+    )
+
+    return ax
+
+
+def draw_hypergraph_hull(
+    H,
+    pos=None,
+    ax=None,
+    dyad_color="black",
+    edge_fc=None,
+    edge_ec=None,
+    node_fc="tab:blue",
+    node_ec="black",
+    node_lw=1,
+    node_size=7,
+    max_order=None,
+    node_labels=False,
+    hyperedge_labels=False,
+    radius=0.05,
+    **kwargs,
+):
+    """Draw hypergraphs displaying the hyperedges of order k>1 as convex hulls
+
+
+    Parameters
+    ----------
+    H : Hypergraph
+    pos : dict (default=None)
+        If passed, this dictionary of positions node_id:(x,y) is used for placing the nodes.
+        If None (default), use the `barycenter_spring_layout` to compute the positions.
+    ax : matplotlib.pyplot.axes (default=None)
+    dyad_color : str, dict, iterable, or EdgeStat (default='black')
+        Color of the dyadic links.  If str, use the same color for all edges. If a dict, must
+        contain (edge_id: color_str) pairs.  If iterable, assume the colors are
+        specified in the same order as the edges are found in H.edges. If EdgeStat, use a colormap
+        (specified with dyad_color_cmap) associated to it.
+    edge_fc : str, dict, iterable, or EdgeStat (default=None)
+        str, dict, iterable, or EdgeStat (default=None)
+        Color of the hyperedges of order k>1.  If str, use the same color for all hyperedges of order k>1.  If a dict, must
+        contain (edge_id: color_str) pairs.  If other iterable, assume the colors are
+        specified in the same order as the hyperedges are found in H.edges. If EdgeStat,
[…]
