+++ tests/algorithms/test_clustering.py
@@ -74,7 +74,11 @@ def test_clustering_coefficient(edgelist1):
     H = xgi.Hypergraph(edgelist1)
     cc = xgi.clustering_coefficient(H)
     true_cc = {1: 1.0, 2: 1.0, 3: 1.0, 4: np.nan, 5: np.nan, 6: 1 / 3, 8: 1.0, 7: 1.0}
-    assert cc == true_cc
+    assert np.isnan(cc[4])
+    assert np.isnan(cc[5])
+    assert {i: cc[i] for i in cc if i not in [4, 5]} == {
+        i: true_cc[i] for i in true_cc if i not in [4, 5]
+    }
 
 
 def test_two_node_clustering_coefficient(edgelist1, edgelist8):
@@ -101,9 +105,21 @@ def test_two_node_clustering_coefficient(edgelist1, edgelist8):
     true_cc1 = {1: 1.0, 2: 1.0, 3: 1.0, 4: np.nan, 5: 0.5, 6: 0.5, 8: 0.75, 7: 0.75}
     true_cc2 = {1: 1.0, 2: 1.0, 3: 1.0, 4: np.nan, 5: 1.0, 6: 1.0, 8: 1.0, 7: 1.0}
     true_cc3 = {1: 1.0, 2: 1.0, 3: 1.0, 4: np.nan, 5: 0.5, 6: 0.5, 8: 0.75, 7: 0.75}
-    assert cc1 == true_cc1
-    assert cc2 == true_cc2
-    assert cc3 == true_cc3
+
+    assert {i: cc1[i] for i in cc1 if i != 4} == {
+        i: true_cc1[i] for i in true_cc1 if i != 4
+    }
+    assert np.isnan(cc1[4])
+
+    assert {i: cc2[i] for i in cc2 if i != 4} == {
+        i: true_cc2[i] for i in true_cc2 if i != 4
+    }
+    assert np.isnan(cc2[4])
+
+    assert {i: cc3[i] for i in cc3 if i != 4} == {
+        i: true_cc3[i] for i in true_cc3 if i != 4
+    }
+    assert np.isnan(cc3[4])
 
     with pytest.raises(XGIError):
         xgi.two_node_clustering_coefficient(H, kind="test")
@@ -121,4 +137,7 @@ def test_two_node_clustering_coefficient(edgelist1, edgelist8):
         6: 0.5,
         10: np.nan,
     }
-    assert cc == true_cc
+    assert {i: cc[i] for i in cc if i != 10} == {
+        i: true_cc[i] for i in true_cc if i != 10
+    }
+    assert np.isnan(cc[10])
