### 2023-04-05 80a5e4b
change to isnan
