### 2023-04-05 ca4fb51
Index-to-unit mappings from to_bipartite_graph. Fixes #322 (#323)
Optionally return the index-to-unit mappings from to_bipartite_graph. The default has been set to False to mirror the behavior of linalg. Fixes 322.
