+++ tests/algorithms/test_clustering.py
@@ -29,9 +29,7 @@ def test_local_clustering_coefficient(edgelist8):
     H = xgi.Hypergraph()
     H.add_nodes_from(range(3))
     cc = xgi.local_clustering_coefficient(H)
-    assert set(cc) == {0, 1, 2}
-    for i in cc:
-        assert np.isnan(cc[i])
+    assert cc == {0: 0, 1: 0, 2: 0}
 
     H = xgi.Hypergraph(edgelist8)
     cc = xgi.local_clustering_coefficient(H)
@@ -71,18 +69,12 @@ def test_clustering_coefficient(edgelist1):
     H = xgi.Hypergraph()
     H.add_nodes_from(range(3))
     cc = xgi.clustering_coefficient(H)
-    assert set(cc) == {0, 1, 2}
-    for i in cc:
-        assert np.isnan(cc[i])
+    assert {0: 0, 1: 0, 2: 0}
 
     H = xgi.Hypergraph(edgelist1)
     cc = xgi.clustering_coefficient(H)
-    true_cc = {1: 1.0, 2: 1.0, 3: 1.0, 4: np.nan, 5: np.nan, 6: 1 / 3, 8: 1.0, 7: 1.0}
-    assert np.isnan(cc[4])
-    assert np.isnan(cc[5])
-    assert {i: cc[i] for i in cc if i not in [4, 5]} == {
-        i: true_cc[i] for i in true_cc if i not in [4, 5]
-    }
+    true_cc = {1: 1.0, 2: 1.0, 3: 1.0, 4: 0, 5: 0, 6: 1 / 3, 8: 1.0, 7: 1.0}
+    assert cc == true_cc
 
 
 def test_two_node_clustering_coefficient(edgelist1, edgelist8):
@@ -106,24 +98,13 @@ def test_two_node_clustering_coefficient(edgelist1, edgelist8):
     cc2 = xgi.two_node_clustering_coefficient(H, kind="min")
     cc3 = xgi.two_node_clustering_coefficient(H, kind="max")
 
-    true_cc1 = {1: 1.0, 2: 1.0, 3: 1.0, 4: np.nan, 5: 0.5, 6: 0.5, 8: 0.75, 7: 0.75}
-    true_cc2 = {1: 1.0, 2: 1.0, 3: 1.0, 4: np.nan, 5: 1.0, 6: 1.0, 8: 1.0, 7: 1.0}
-    true_cc3 = {1: 1.0, 2: 1.0, 3: 1.0, 4: np.nan, 5: 0.5, 6: 0.5, 8: 0.75, 7: 0.75}
+    true_cc1 = {1: 1.0, 2: 1.0, 3: 1.0, 4: 0, 5: 0.5, 6: 0.5, 8: 0.75, 7: 0.75}
+    true_cc2 = {1: 1.0, 2: 1.0, 3: 1.0, 4: 0, 5: 1.0, 6: 1.0, 8: 1.0, 7: 1.0}
+    true_cc3 = {1: 1.0, 2: 1.0, 3: 1.0, 4: 0, 5: 0.5, 6: 0.5, 8: 0.75, 7: 0.75}
 
-    assert {i: cc1[i] for i in cc1 if i != 4} == {
-        i: true_cc1[i] for i in true_cc1 if i != 4
-    }
-    assert np.isnan(cc1[4])
-
-    assert {i: cc2[i] for i in cc2 if i != 4} == {
-        i: true_cc2[i] for i in true_cc2 if i != 4
-    }
-    assert np.isnan(cc2[4])
-
-    assert {i: cc3[i] for i in cc3 if i != 4} == {
-        i: true_cc3[i] for i in true_cc3 if i != 4
-    }
[…]
