### 2023-04-05 7ee8aa2
Add clustering coefficient definitions (#316)
* moved modules

* add clustering code

* added clustering functions

* add tests

* fixed tests

* Update xgi/algorithms/clustering.py

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>

* updated centrality names

* added more info on clustering

* updated all names

* fixed name

* update docstrings

* added tests

* Update clustering.py

* changed NaN to nan

https://stackoverflow.com/questions/53436339/difference-between-np-nan-and-np-nan

* response to review

* change to isnan

* fixed more tests

---------

Co-authored-by:
[…]
