+++ xgi/algorithms/__init__.py
@@ -1,4 +1,5 @@
-from . import assortativity, centrality, connected
+from . import assortativity, centrality, clustering, connected
 from .assortativity import *
 from .centrality import *
+from .clustering import *
 from .connected import *
