+++ docs/source/api/algorithms.rst
@@ -9,4 +9,5 @@ algorithms package
    
    ~xgi.algorithms.assortativity
    ~xgi.algorithms.centrality
+   ~xgi.algorithms.clustering
    ~xgi.algorithms.connected
