+++ docs/source/api/stats/xgi.stats.nodestats.rst
@@ -9,8 +9,10 @@
 
    .. autofunction:: attrs
    .. autofunction:: average_neighbor_degree
-   .. autofunction:: cec_centrality
-   .. autofunction:: clustering
+   .. autofunction:: clique_eigenvector_centrality
+   .. autofunction:: clustering_coefficient
    .. autofunction:: degree
-   .. autofunction:: hec_centrality
+   .. autofunction:: h_eigenvector_centrality
+   .. autofunction:: local_clustering_coefficient
    .. autofunction:: node_edge_centrality
+   .. autofunction:: two_node_clustering_coefficient
