+++ docs/source/api/linalg/xgi.linalg.hypergraph_matrix.rst
@@ -1,0 +1,14 @@
+﻿xgi.linalg.hypergraph_matrix
+============================
+
+.. currentmodule:: xgi.linalg.hypergraph_matrix
+
+.. automodule:: xgi.linalg.hypergraph_matrix
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: adjacency_matrix
+   .. autofunction:: clique_motif_matrix
+   .. autofunction:: degree_matrix
+   .. autofunction:: incidence_matrix
+   .. autofunction:: intersection_profile
