+++ docs/source/api/generators/xgi.generators.simple.rst
@@ -1,0 +1,11 @@
+xgi.generators.simple
+=====================
+
+.. currentmodule:: xgi.generators.simple
+
+.. automodule:: xgi.generators.simple
+   
+    .. rubric:: Functions
+
+    .. autofunction:: star_clique
+    .. autofunction:: sunflower
