+++ docs/source/api/generators/xgi.generators.nonuniform.rst
@@ -1,15 +1,0 @@
-xgi.generators.nonuniform
-=========================
-
-.. currentmodule:: xgi.generators.nonuniform
-
-.. automodule:: xgi.generators.nonuniform
-   
-   .. rubric:: Functions
-   
-   .. autofunction:: chung_lu_hypergraph
-   .. autofunction:: dcsbm_hypergraph
-   .. autofunction:: random_flag_complex
-   .. autofunction:: random_flag_complex_d2
-   .. autofunction:: random_hypergraph
-   .. autofunction:: random_simplicial_complex
