+++ docs/source/api/linalg/xgi.linalg.hodge_matrix.rst
@@ -1,0 +1,11 @@
+xgi.linalg.hodge_matrix
+=======================
+
+.. currentmodule:: xgi.linalg.hodge_matrix
+
+.. automodule:: xgi.linalg.hodge_matrix
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: boundary_matrix
+   .. autofunction:: hodge_laplacian
