### 2023-04-05 3435409
Restructure folder for generators and linalg (#321)
* refact: new folder structure for linalg

* refact: new folder structure for generators step 1g

* fix: imports

* fix: add missing files

* updated codecov params to avoid annoying warnings

* refact: moved more generators around

* refact: merged (non)uniform random_hypergraph() and remove nonuniform file

* fix: tests

* PR review comments

* minor adjustments

* minor adjustments

* docs: added missing documentation

* added the uniform Erdos-Renyi back in

---------

Co-authored-by: Nicholas Landry <nicholas.landry.91@gmail.com>
