+++ xgi/generators/random.py
@@ -5,22 +5,89 @@ import warnings
 from collections import defaultdict
 from itertools import combinations
 
-import networkx as nx
 import numpy as np
 from scipy.special import comb
 
-from ..classes import SimplicialComplex
-from .classic import empty_hypergraph, flag_complex_d2, ring_lattice
+from .classic import empty_hypergraph
+from .lattice import ring_lattice
 
 __all__ = [
+    "random_hypergraph",
     "chung_lu_hypergraph",
     "dcsbm_hypergraph",
-    "random_hypergraph",
-    "random_simplicial_complex",
-    "random_flag_complex_d2",
-    "random_flag_complex",
     "watts_strogatz_hypergraph",
 ]
+
+
+def random_hypergraph(N, ps, order=None, seed=None):
+    """Generates a random hypergraph
+
+    Generate N nodes, and connect any d+1 nodes
+    by a hyperedge with probability ps[d-1].
+
+    Parameters
+    ----------
+    N : int
+        Number of nodes
+    ps : list of float
+        List of probabilities (between 0 and 1) to create a
+        hyperedge at each order d between any d+1 nodes. For example,
+        ps[0] is the wiring probability of any edge (2 nodes), ps[1]
+        of any triangles (3 nodes).
+    order: int of None (default)
+        If None, ignore. If int, generates a uniform hypergraph with edges
+        of order `order` (ps must have only one element).
+    seed : integer or None (default)
+            Seed for the random number generator.
+
+    Returns
+    -------
+    Hypergraph object
+        The generated hypergraph
+
+    References
+    ----------
+    Described as 'random hypergraph' by M. Dewar et al. in https://arxiv.org/abs/1703.07686
+
+    Example
+    -------
+    >>> import xgi
+    >>> H = xgi.random_hypergraph(50, [0.1, 0.01])
+
+    """
+    if seed is not None:
+        np.random.seed(seed)
+
+    if order is not None:
+        if len(ps) != 1:
+            raise ValueError("ps must contain a single element if order is an int")
+
+    if (np.any(np.array(ps) < 0)) or (np.any(np.array(ps) > 1)):
+        raise ValueError("All elements of ps must be between 0 and 1 included.")
+
+    nodes = range(N)
+    hyperedges = []
+
+    for i, p in enumerate(ps):
+
+        if order is not None:
+            d = order
+        else:
+            d = i + 1  # order, ps[0] is prob of edges (d=1)
+
+        potential_edges = combinations(nodes, d + 1)
+        n_comb = comb(N, d + 1, exact=True)
+        mask = np.random.random(size=n_comb) <= p  # True if edge to keep
+
+        edges_to_add = [e for e, val in zip(potential_edges, mask) if val]
+
+        hyperedges += edges_to_add
+
+    H = empty_hypergraph()
+    H.add_nodes_from(nodes)
+    H.add_edges_from(hyperedges)
+
+    return H
 
 
 def chung_lu_hypergraph(k1, k2, seed=None):
@@ -114,7 +181,8 @@ def chung_lu_hypergraph(k1, k2, seed=None):
 
 
 def dcsbm_hypergraph(k1, k2, g1, g2, omega, seed=None):
-    """A function to generate a DCSBM hypergraph.
+    """A function to generate a Degree-Corrected Stochastic Block Model
+    (DCSBM) hypergraph.
 
     Parameters
     ----------
@@ -249,207 +317,6 @@ def dcsbm_hypergraph(k1, k2, g1, g2, omega, seed=None):
                         p = q
                         j = j + 1
     return H
-
-
-def random_hypergraph(N, ps, seed=None):
-    """Generates a random hypergraph
-
-    Generate N nodes, and connect any d+1 nodes
[…]

[…]
