### 2023-04-08 b0914ab
added the maximal method to the EdgeView
