### 2023-04-11 03240d8
fix: documented max_order in add_simplices_from
