### 2023-04-11 13e283c
fix: improve degree_counts doc
