### 2023-04-11 97240c5
Update xgi/classes/reportviews.py
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
