### 2023-04-11 6f09560
Added `maximal` method to EdgeView (#324)
* added the maximal method to the EdgeView

* style: format with black

* removed references to the old function

* Update xgi.classes.reportviews.EdgeView.rst

* response to review

* added more documentation

* Update xgi/classes/reportviews.py

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>

---------

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
