+++ xgi/drawing/draw.py
@@ -1017,8 +1017,8 @@ def _draw_hull(node_pos, ax, edges_ec, facecolor, alpha, zorder, radius):
     ax : matplotlib.pyplot.axes
 
     """
-    
-    thetas = np.linspace(0, 2*np.pi, num=100, endpoint=False)
+
+    thetas = np.linspace(0, 2 * np.pi, num=100, endpoint=False)
     offsets = radius * np.array([np.cos(thetas), np.sin(thetas)]).T
     points = np.vstack([p + offsets for p in node_pos])
     points = np.vstack([node_pos, points])
