### 2023-04-11 b42a04f
fix: improve degree_counts doc (#329)
