### 2023-04-11 9b55fc1
added multiedges argument and tests
