+++ xgi/stats/__init__.py
@@ -40,7 +40,7 @@ NodeView((1, 4))
 
 Many other features are available, including edge-statistics, and user-defined
 statistics.  For more details, see the `tutorial
-<https://github.com/ComplexGroupInteractions/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+<https://github.com/xgi/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
 
 """
 
@@ -206,7 +206,7 @@ class NodeStat(IDStat):
 
     `NodeStat` objects represent a mapping that assigns a value to each node in a
     network.  For more details, see the `tutorial
-    <https://github.com/ComplexGroupInteractions/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+    <https://github.com/xgi/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
 
     """
 
@@ -216,7 +216,7 @@ class EdgeStat(IDStat):
 
     `EdgeStat` objects represent a mapping that assigns a value to each edge in a
     network.  For more details, see the `tutorial
-    <https://github.com/ComplexGroupInteractions/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+    <https://github.com/xgi/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
 
     """
 
@@ -393,7 +393,7 @@ class MultiNodeStat(MultiIDStat):
     """Multiple node-quantity mappings.
 
     For more details, see the `tutorial
-    <https://github.com/ComplexGroupInteractions/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+    <https://github.com/xgi/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
 
     """
 
@@ -405,7 +405,7 @@ class MultiEdgeStat(MultiIDStat):
     """Multiple edge-quantity mappings.
 
     For more details, see the `tutorial
-    <https://github.com/ComplexGroupInteractions/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+    <https://github.com/xgi/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
 
     """
 
