### 2023-04-13 7aceaa7
change name to xgi-org (#334)
* change name to xgi-org

* fix codecov error
