+++ long_description.rst
@@ -1,7 +1,7 @@
 XGI
 ===
 
-.. image:: https://github.com/ComplexGroupInteractions/xgi/raw/main/logo/logo.svg
+.. image:: https://github.com/xgi-org/xgi/raw/main/logo/logo.svg
   :width: 200
 
 CompleX Group Interactions (XGI) is a Python package for the representation, manipulation,
@@ -26,15 +26,15 @@ To install this package locally:
 
 Getting Started
 ---------------
-To get started, take a look at the `tutorials <https://github.com/ComplexGroupInteractions/xgi/tree/main/tutorials>`_
+To get started, take a look at the `tutorials <https://github.com/xgi-org/xgi/tree/main/tutorials>`_
 illustrating the library's basic functionality.
 
 Contributing
 ------------
-Contributions are always welcome. Please report any bugs that you find `here <https://github.com/ComplexGroupInteractions/xgi/issues>`_.
-Or, even better, fork the repository on `GitHub <https://github.com/ComplexGroupInteractions/xgi>`_ and create a pull request (PR).
+Contributions are always welcome. Please report any bugs that you find `here <https://github.com/xgi-org/xgi/issues>`_.
+Or, even better, fork the repository on `GitHub <https://github.com/xgi-org/xgi>`_ and create a pull request (PR).
 We welcome all changes, big or small, and we will help you make the PR if you are new to `git`
-(just ask on the issue and/or see our `contributing guidelines <https://github.com/ComplexGroupInteractions/xgi/tree/main/CONTRIBUTING.md>`_.
+(just ask on the issue and/or see our `contributing guidelines <https://github.com/xgi-org/xgi/tree/main/CONTRIBUTING.md>`_.
 
 How to Cite
 -----------
@@ -42,17 +42,17 @@ We acknowledge the importance of good software to support research, and we note
 that research becomes more valuable when it is communicated effectively. To
 demonstrate the value of XGI, we ask that you cite XGI in your work.
 Currently, the best way to cite XGI is to go to our
-`repository page <https://github.com/ComplexGroupInteractions/xgi>`_ and
+`repository page <https://github.com/xgi-org/xgi>`_ and
 click the "cite this repository" button on the right sidebar. This will generate
 a citation in your preferred format, and will also integrate well with citation managers.
 
 Code of Conduct
 ---------------
-Our full code of conduct, and how we enforce it, can be read in `our repository <https://github.com/ComplexGroupInteractions/xgi/tree/main/CODE_OF_CONDUCT.md>`_.
+Our full code of conduct, and how we enforce it, can be read in `our repository <https://github.com/xgi-org/xgi/tree/main/CODE_OF_CONDUCT.md>`_.
 
 License
 -------
-Released under the 3-Clause BSD license (see the `license file <https://github.com/ComplexGroupInteractions/xgi/tree/main/license.md>`_)
+Released under the 3-Clause BSD license (see the `license file <https://github.com/xgi-org/xgi/tree/main/license.md>`_)
 
 Copyright (C) 2021-2023 XGI Developers
 
