+++ .github/workflows/coverage.yml
@@ -27,3 +27,5 @@ jobs:
           pytest --cov=xgi --cov-report=xml
       - name: Upload coverage reports to Codecov with GitHub Action
         uses: codecov/codecov-action@v3
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}

[…]
