### 2023-04-13 ebe51d9
feat: trivial hypergraph
