+++ tests/generators/test_classic.py
@@ -9,3 +9,7 @@ def test_empty_hypergraph():
 def test_empty_hypergraph():
     SC = xgi.empty_simplicial_complex()
     assert (SC.num_nodes, SC.num_edges) == (0, 0)
+
+def test_trivial_hypergraph():
+    SC = xgi.trivial_hypergraph()
+    assert (SC.num_nodes, SC.num_edges) == (1, 0)
