+++ tests/generators/test_classic.py
@@ -11,5 +11,11 @@ def test_empty_hypergraph():
     assert (SC.num_nodes, SC.num_edges) == (0, 0)
 
 def test_trivial_hypergraph():
-    SC = xgi.trivial_hypergraph()
-    assert (SC.num_nodes, SC.num_edges) == (1, 0)
+    H = xgi.trivial_hypergraph()
+    assert (H.num_nodes, H.num_edges) == (1, 0)
+
+    H = xgi.trivial_hypergraph(n=1)
+    assert (H.num_nodes, H.num_edges) == (1, 0)
+
+    H = xgi.trivial_hypergraph(n=2)
+    assert (H.num_nodes, H.num_edges) == (2, 0)
