### 2023-04-14 112defd
feat: added trivial hypergraph (#335)
* feat: trivial hypergraph

* fix

* review commets

* review comments 2
