### 2023-04-14 5607375
Added multiedge keyword to `maximal()` method (#332)
* added multiedges argument and tests

* removed bad test

* response to review
