+++ tests/generators/test_classic.py
@@ -10,6 +10,7 @@ def test_empty_hypergraph():
     SC = xgi.empty_simplicial_complex()
     assert (SC.num_nodes, SC.num_edges) == (0, 0)
 
+
 def test_trivial_hypergraph():
     H = xgi.trivial_hypergraph()
     assert (H.num_nodes, H.num_edges) == (1, 0)
