### 2023-04-17 4e8a9b3
feat: added complete hypergraph
