### 2023-04-17 e6d5c1f
docs: added new function. style: black isort
