+++ docs/source/api/generators/xgi.generators.classic.rst
@@ -10,3 +10,4 @@
    .. autofunction:: empty_hypergraph
    .. autofunction:: empty_simplicial_complex
    .. autofunction:: trivial_hypergraph
+   .. autofunction:: complete_hypergraph
