+++ xgi/generators/classic.py
@@ -182,18 +182,28 @@ def complete_hypergraph(N, order=None, max_order=None, include_singletons=False)
     Hypergraph object
         A complete hypergraph with `N` nodes
 
-    Note
+    Notes
     ----
     Only one of `order` and `max_order` can be specified by and int (not None).
     Additionally, at least one of either must be specified.
+
+    The number of possible edges grows exponentially as $\approx 2^N$ for large `N` and 
+    quickly becomes impractically long to compute, especially when using `max_order`. For 
+    example, `N=100` and `max_order=5` already yields $8 10^7$ edges. Increasing `N=1000`
+    makes it $8 10^{12}$. `N=100` and with a larger `max_order=6` yields $8 10^8$ edgfes.
+
     """
     # this import needs to happen when the function runs, not when the module is first
     # imported, to avoid circular imports
     import xgi
 
-    if bool(order) == bool(max_order):
+    if (order is None) and (max_order is None):
         raise ValueError(
-            "One (and one only) among order and max_order must be specified (not None)"
+            "At least one among order and max_order must be specified (not None)"
+        )
+    if (order is not None) and (max_order is not None):
+        raise ValueError(
+            "Both order and max_order cannot be specified (not None) at the same time."
         )
 
     H = xgi.Hypergraph()
