### 2023-04-17 1191b60
feat: added complete_hypergraph (#337)
* feat: added complete hypergraph

* tests: added corresponding

* docs: added new function. style: black isort

* review comments

* review comments

* fix docs maths
