+++ docs/source/conf.py
@@ -163,6 +163,8 @@ html_theme = "sphinx_rtd_theme"
 # html_theme_path = ["_static"]
 # html_static_path = ["_static"]
 
+intersphinx_mapping = {'xgi': ('https://xgi.readthedocs.io', None)}
+
 html_show_sphinx = True
 
 htmlhelp_basename = "XGIDoc"
