+++ xgi/convert.py
@@ -52,7 +52,7 @@ def convert_to_hypergraph(data, create_using=None):
     ----------
     data : object to be converted
         Current known types are:
-         * a Hypergraph object
+         * a SimplicialComplex object
          * list-of-iterables
          * dict-of-iterables
          * Pandas DataFrame (bipartite edgelist)
@@ -71,23 +71,19 @@ def convert_to_hypergraph(data, create_using=None):
     if data is None:
         return empty_hypergraph(create_using)
 
-    elif isinstance(data, Hypergraph):
-        H = empty_hypergraph(create_using)
-        H.add_nodes_from((n, attr) for n, attr in data.nodes.items())
-        ee = data.edges
-        H.add_edges_from((ee.members(e), e, deepcopy(attr)) for e, attr in ee.items())
-        H._hypergraph = deepcopy(data._hypergraph)
+    elif isinstance(data, SimplicialComplex):
+        return from_max_simplices(data)
 
     elif isinstance(data, list):
         # edge list
-        from_hyperedge_list(data, create_using)
+        return from_hyperedge_list(data, create_using)
 
     elif isinstance(data, pd.DataFrame):
-        from_bipartite_pandas_dataframe(data, create_using)
+        return from_bipartite_pandas_dataframe(data, create_using)
 
     elif isinstance(data, dict):
         # edge dict in the form we need
-        from_hyperedge_dict(data, create_using)
+        return from_hyperedge_dict(data, create_using)
 
     elif isinstance(
         data,
@@ -207,13 +203,23 @@ def convert_to_simplicial_complex(data, create_using=None):
             (ee.members(e), e, deepcopy(attr)) for e, attr in ee.items()
         )
         H._hypergraph = deepcopy(data._hypergraph)
+        return H
+    
+    elif isinstance(data, Hypergraph):
+        H = empty_simplicial_complex(create_using)
+        H.add_nodes_from((n, attr) for n, attr in data.nodes.items())
+        ee = data.edges
+        H.add_simplices_from(
+            (ee.members(e), e, deepcopy(attr)) for e, attr in ee.items()
+        )
+        return H
 
     elif isinstance(data, list):
         # edge list
-        from_hyperedge_list(data, create_using)
+        return from_hyperedge_list(data, create_using)
 
     elif isinstance(data, pd.DataFrame):
-        from_bipartite_pandas_dataframe(data, create_using)
+        return from_bipartite_pandas_dataframe(data, create_using)
 
     elif isinstance(data, dict):
         # edge dict in the form we need
