### 2023-04-27 4ea03b6
docs: fix docs for clustering and connected
