### 2023-04-27 #5162 @MridulS
Thanks for the ping @vissarion and sorry for the delay!

I have finished the review and this paper should definitely be included in JOSS! The paper is well written and the package and documentation covers all the necessary details. There could be a some more information about things like "How to contribute", as currently it still leaves a lot to the reader. But I think this is good to go :) We (NetworkX) have also been thinking about how to interact with HyperGraphs in the future so this was a fun read!
