### 2023-04-27 25d4239
docs: fix hypergraphviews and reportviews
