### 2023-04-27 3f2a5e2
docs: fix simplicialcomplex
