+++ xgi/stats/nodestats.py
@@ -287,10 +287,11 @@ def two_node_clustering_coefficient(net, bunch, kind="union"):
         The type of two-node clustering coefficient.
         Types are:
 
-        - "union"
-        - "min"
-        - "max"
-        by default, "union".
+        * "union"
+        * "min"
+        * "max"
+
+    by default, "union".
 
     Returns
     -------
