+++ requirements/tutorial.txt
@@ -1,3 +1,4 @@
 jupyter>=1.0
 matplotlib>=3.3
 seaborn>=0.10
+ipython<=8.13
