### 2023-04-28 5497778
Fix IPython error (#344)
Update tutorial requirements file.
