+++ xgi/convert.py
@@ -73,10 +73,10 @@ def convert_to_hypergraph(data, create_using=None):
         return empty_hypergraph(create_using)
     
     elif isinstance(data, Hypergraph):
-        H = empty_simplicial_complex(create_using)
+        H = empty_hypergraph(create_using)
         H.add_nodes_from((n, attr) for n, attr in data.nodes.items())
         ee = data.edges
-        H.add_simplices_from(
+        H.add_edges_from(
             (ee.members(e), e, deepcopy(attr)) for e, attr in ee.items()
         )
         return H
