### 2023-04-29 4d04161
Add test for SC->H and H->SC
