+++ xgi/convert.py
@@ -71,7 +71,10 @@ def convert_to_hypergraph(data, create_using=None):
     """
     if data is None:
         return empty_hypergraph(create_using)
-    
+
+    elif isinstance(data, SimplicialComplex):
+        return from_max_simplices(data)
+
     elif isinstance(data, Hypergraph):
         H = empty_hypergraph(create_using)
         H.add_nodes_from((n, attr) for n, attr in data.nodes.items())
@@ -80,9 +83,6 @@ def convert_to_hypergraph(data, create_using=None):
             (ee.members(e), e, deepcopy(attr)) for e, attr in ee.items()
         )
         return H
-
-    elif isinstance(data, SimplicialComplex):
-        return from_max_simplices(data)
 
     elif isinstance(data, list):
         # edge list
