+++ xgi/classes/simplicialcomplex.py
@@ -1,9 +1,10 @@
 """Base class for undirected simplicial complexes.
 
-The SimplicialComplex class allows any hashable object as a node
-and can associate key/value attribute pairs with each undirected simplex and node.
+The SimplicialComplex class allows any hashable object as a node and can associate
+key/value attribute pairs with each undirected simplex and node.
 
 Multi-simplices are not allowed.
+
 """
 
 from collections.abc import Hashable, Iterable
@@ -22,24 +23,23 @@ class SimplicialComplex(Hypergraph):
     r"""A class to represent undirected simplicial complexes.
 
     A simplicial complex is a collection of subsets of a set of *nodes* or *vertices*.
-    It is a pair :math:`(V, E)`, where :math:`V` is a set of elements called
-    *nodes* or *vertices*, and :math:`E` is a set whose elements are subsets of
-    :math:`V`, that is, each :math:`e \in E` satisfies :math:`e \subset V`.  The
-    elements of :math:`E` are called *simplices*. Additionally, if a simplex is part of
-    a simplicial complex, all its faces must be too. This makes simplicial complexes
-    a special case of hypergraphs.
+    It is a pair :math:`(V, E)`, where :math:`V` is a set of elements called *nodes* or
+    *vertices*, and :math:`E` is a set whose elements are subsets of :math:`V`, that is,
+    each :math:`e \in E` satisfies :math:`e \subset V`.  The elements of :math:`E` are
+    called *simplices*. Additionally, if a simplex is part of a simplicial complex, all
+    its faces must be too. This makes simplicial complexes a special case of
+    hypergraphs.
 
     The SimplicialComplex class allows any hashable object as a node and can associate
-    attributes to each node, simplex, or the simplicial complex itself, in the form of key/value
-    pairs.
-
+    attributes to each node, simplex, or the simplicial complex itself, in the form of
+    key/value pairs.
 
     Parameters
     ----------
     incoming_data : input simplicial complex data, optional
         Data to initialize the simplicial complex. If None (default), an empty
-        simplicial complex is created, i.e. one with no nodes or simplices.
-        The data can be in the following formats:
+        simplicial complex is created, i.e. one with no nodes or simplices.  The data
+        can be in the following formats:
 
         * simplex list
         * simplex dictionary
@@ -110,44 +110,54 @@ class SimplicialComplex(Hypergraph):
 
         """
         try:
-            return f"{type(self).__name__} named '{self['name']}' with {self.num_nodes} nodes and {self.num_edges} simplices"
+            return (
+                f"{type(self).__name__} named '{self['name']}' "
+                f"with {self.num_nodes} nodes and {self.num_edges} simplices"
+            )
         except XGIError:
-            return f"Unnamed {type(self).__name__} with {self.num_nodes} nodes and {self.num_edges} simplices"
+            return (
+                f"Unnamed {type(self).__name__} with {self.num_nodes}"
+                f" nodes and {self.num_edges} simplices"
+            )
 
     def add_edge(self, edge, id=None, **attr):
-        """add_edge is deprecated in SimplicialComplex. Use add_simplex instead"""
+        """Deprecated in SimplicialComplex. Use add_simplex instead"""
         warn("add_edge is deprecated in SimplicialComplex. Use add_simplex instead")
         return self.add_simplex(edge, id=None, **attr)
 
     def add_edges_from(self, ebunch_to_add, max_order=None, **attr):
-        """add_edges_from is deprecated in SimplicialComplex. Use add_simplices_from instead"""
+        """Deprecated in SimplicialComplex. Use add_simplices_from instead"""
         warn(
-            "add_edges_from is deprecated in SimplicialComplex. Use add_simplices_from instead"
[…]
