+++ xgi/readwrite/json.py
@@ -28,18 +28,18 @@ def write_json(H, path):
 
     # get node data
     try:
-        data["node-data"] = {str(id): H.nodes[id] for id in H.nodes}
-    except:
+        data["node-data"] = {str(idx): H.nodes[idx] for idx in H.nodes}
+    except KeyError:
         raise XGIError("Node attributes not saved!")
 
     try:
-        data["edge-data"] = {str(id): H.edges[id] for id in H.edges}
-    except:
+        data["edge-data"] = {str(idx): H.edges[idx] for idx in H.edges}
+    except KeyError:
         raise XGIError("Edge attributes not saved!")
 
     # hyperedge dict
     data["edge-dict"] = {
-        str(id): [str(n) for n in H.edges.members(id)] for id in H.edges
+        str(idx): [str(n) for n in H.edges.members(idx)] for idx in H.edges
     }
 
     datastring = json.dumps(data, indent=2)
