### 2023-05-02 a0fda35
prefer function definition over assigning to lambdas
