+++ tests/generators/test_classic.py
@@ -10,7 +10,7 @@ def test_empty_hypergraph():
     assert (H.num_nodes, H.num_edges) == (0, 0)
 
 
-def test_empty_hypergraph():
+def test_empty_simplicial_complex():
     SC = xgi.empty_simplicial_complex()
     assert (SC.num_nodes, SC.num_edges) == (0, 0)
 
