### 2023-05-03 16793a1
remove hardcoded alpha
