### 2023-05-03 b0b2d29
Attempt at fixing convert (#345)
Update convert.py and add tests. Fixes #327.
