+++ xgi/convert.py
@@ -53,6 +53,7 @@ def convert_to_hypergraph(data, create_using=None):
     data : object to be converted
         Current known types are:
          * a Hypergraph object
+         * a SimplicialComplex object
          * list-of-iterables
          * dict-of-iterables
          * Pandas DataFrame (bipartite edgelist)
@@ -71,23 +72,36 @@ def convert_to_hypergraph(data, create_using=None):
     if data is None:
         return empty_hypergraph(create_using)
 
-    elif isinstance(data, Hypergraph):
+    elif isinstance(data, Hypergraph) and not isinstance(data, SimplicialComplex):
         H = empty_hypergraph(create_using)
         H.add_nodes_from((n, attr) for n, attr in data.nodes.items())
         ee = data.edges
-        H.add_edges_from((ee.members(e), e, deepcopy(attr)) for e, attr in ee.items())
+        H.add_edges_from(
+            (ee.members(e), e, deepcopy(attr)) for e, attr in ee.items()
+        )
         H._hypergraph = deepcopy(data._hypergraph)
+        return H
+    
+
+    elif isinstance(data, SimplicialComplex):
+        return from_max_simplices(data)
 
     elif isinstance(data, list):
         # edge list
-        from_hyperedge_list(data, create_using)
+        result = from_hyperedge_list(data, create_using)
+        if not isinstance(create_using, Hypergraph):
+            return result
 
     elif isinstance(data, pd.DataFrame):
-        from_bipartite_pandas_dataframe(data, create_using)
+        result = from_bipartite_pandas_dataframe(data, create_using)
+        if not isinstance(create_using, Hypergraph):
+            return result
 
     elif isinstance(data, dict):
         # edge dict in the form we need
-        from_hyperedge_dict(data, create_using)
+        result = from_hyperedge_dict(data, create_using)
+        if not isinstance(create_using, Hypergraph):
+            return result
 
     elif isinstance(
         data,
@@ -180,6 +194,7 @@ def convert_to_simplicial_complex(data, create_using=None):
     ----------
     data : object to be converted
         Current known types are:
+         * a SimplicialComplex object
          * a Hypergraph object
          * list-of-iterables
          * dict-of-iterables
@@ -197,7 +212,7 @@ def convert_to_simplicial_complex(data, create_using=None):
     """
 
     if data is None:
-        return empty_hypergraph(create_using)
+        return empty_simplicial_complex(create_using)
 
     elif isinstance(data, SimplicialComplex):
         H = empty_simplicial_complex(create_using)
@@ -207,14 +222,28 @@ def convert_to_simplicial_complex(data, create_using=None):
             (ee.members(e), e, deepcopy(attr)) for e, attr in ee.items()
         )
         H._hypergraph = deepcopy(data._hypergraph)
+        return H
+    
+    elif isinstance(data, Hypergraph):
+        H = empty_simplicial_complex(create_using)
+        H.add_nodes_from((n, attr) for n, attr in data.nodes.items())
+        ee = data.edges
+        H.add_simplices_from(
+            (ee.members(e), e, deepcopy(attr)) for e, attr in ee.items()
+        )
+        return H
 
     elif isinstance(data, list):
         # edge list
-        from_hyperedge_list(data, create_using)
+        result = from_hyperedge_list(data, create_using)
+        if not isinstance(create_using, SimplicialComplex):
+            return convert_to_simplicial_complex(result)
 
     elif isinstance(data, pd.DataFrame):
-        from_bipartite_pandas_dataframe(data, create_using)
-
+        result = from_bipartite_pandas_dataframe(data, create_using)
+        if not isinstance(create_using, SimplicialComplex):
+            return convert_to_simplicial_complex(result)
+            
[…]
