### 2023-05-03 fd7113b
update docs requirements
