+++ requirements/documentation.txt
@@ -1,6 +1,6 @@
-sphinx~=4.0
+sphinx>=7.0
 sphinx_copybutton
-sphinx-rtd-theme>=0.4.2
+pydata-sphinx-theme
 numpydoc>=1.1
 pillow>=8.2
 matplotlib>=3.3
