+++ docs/source/index.rst
@@ -64,8 +64,8 @@ designed and developed by network scientists with the needs of network scientist
 mind.
 
 - Repository: https://github.com/xgi-org/xgi
-- PyPI: https://pypi.org/project/xgi/
-- Documentation: https://xgi.readthedocs.io/
+- PyPI: `latest release <https://pypi.org/project/xgi/>`_
+- Twitter: `@xginets <https://twitter.com/xginets>`_
 
 Sign up for our `mailing list <http://eepurl.com/igE6ez>`_ and follow XGI on `Twitter <https://twitter.com/xginets>`_ or `Mastodon <https://mathstodon.xyz/@xginets>`_!
 
