### 2023-05-03 ebbe588
updated the landing page for GH, PyPI, and RTD
