+++ docs/source/index.rst
@@ -93,11 +93,43 @@ If that command does not work, you may try the following instead
 
 XGI was developed and tested for Python 3.8-3.11 on Mac OS, Windows, and Ubuntu.
 
+
 Corresponding Data
 ==================
 
 A number of higher-order datasets are available in the `XGI-DATA repository <https://gitlab.com/complexgroupinteractions/xgi-data>`_ and can be easily accessed with the ``load_xgi_data()`` function.
 
+
+Contributing
+============
+
+If you want to contribute to this project, please make sure to read the
+`contributing guidelines <CONTRIBUTING.md>`_.
+We expect respectful and kind interactions by all contributors and users
+as laid out in our `code of conduct <CODE_OF_CONDUCT.md>`_.
+
+The XGI community always welcomes contributions, no matter how small.
+We're happy to help troubleshoot XGI issues you run into,
+assist you if you would like to add functionality or fixes to the codebase,
+or answer any questions you may have.
+
+Some concrete ways that you can get involved:
+
+* **Get XGI updates** by following the XGI `Twitter <https://twitter.com/xginets>`_ account, signing up for our `mailing list <http://eepurl.com/igE6ez>`_, or starring this repository.
+* **Spread the word** when you use XGI by sharing with your colleagues and friends.
+* **Request a new feature or report a bug** by raising a `new issue <https://github.com/xgi-org/xgi/issues/new>`_.
+* **Create a Pull Request (PR)** to address an `open issue <../../issues>`_ or add a feature.
+* **Join our Zulip channel** to be a part of the `daily goings-on of XGI <https://xgi.zulipchat.com/join/7agfwo7dh7jo56ppnk5kc23r/>`_.
+
+How to Cite
+===========
+We acknowledge the importance of good software to support research, and we note
+that research becomes more valuable when it is communicated effectively. To
+demonstrate the value of XGI, we ask that you cite XGI in your work.
+Currently, the best way to cite XGI is to go to our
+`repository page <https://github.com/xgi-org/xgi>`_ and
+click the "cite this repository" button on the right sidebar. This will generate
+a citation in your preferred format, and will also integrate well with citation managers.
 
 Academic References
 ===================
@@ -119,27 +151,6 @@ Academic References
   Rosvall, and Ingo Scholtes.
 
 
-Contributing
-============
-
-If you want to contribute to this project, please make sure to read the
-`code of conduct
-<https://github.com/xgi-org/xgi/blob/main/CODE_OF_CONDUCT.md>`_
-and the `contributing guidelines
-<https://github.com/xgi-org/xgi/blob/main/CONTRIBUTING.md>`_.
-
-The best way to contribute to XGI is by submitting a bug or request a new feature by
-opening a `new issue <https://github.com/xgi-org/xgi/issues/new>`_.
-
-To get more actively involved, you are invited to browse the `issues page
-<https://github.com/xgi-org/xgi/issues>`_ and choose one that you can
-work on.  The core developers will be happy to help you understand the codebase and any
-other doubts you may have while working on your contribution.
-
-If you are interested in the daily goings-on of XGI, you are invited to join our `Zulip
-channel <https://xgi.zulipchat.com/join/7agfwo7dh7jo56ppnk5kc23r/>`_.
-
-
 Contributors
 ============
 
@@ -159,6 +170,9 @@ Other contributors:
 * Tim LaRock
 * Sabina Adhikari
 * Marco Nurisso
+* Alexis Arnaudon
+* Thomas Robiglio
+* Gonzalo Contreras Aso
 
 Funding
 =======
