### 2023-05-03 4e59510
added additional unit tests
