+++ xgi/utils/utilities.py
@@ -13,32 +13,32 @@ __all__ = [
 
 def dual_dict(edge_dict):
     """Given a dictionary with IDs as keys
-    and lists as values, return the dual.
+    and sets as values, return the dual.
 
     Parameters
     ----------
     edge_dict : dict
         A dictionary where the keys are
-        IDs and the values are lists of hashables
+        IDs and the values are sets of hashables
 
     Returns
     -------
     dict
         A dictionary with IDs as keys
-        and lists as values, but the reverse of
+        and sets as values, but the reverse of
         the original dict.
 
     Examples
     --------
     >>> import xgi
     >>> xgi.dual_dict({0 : [1, 2, 3], 1 : [0, 2]})
-    {1: [0], 2: [0, 1], 3: [0], 0: [1]}
+    {1: {0}, 2: {0, 1}, 3: {0}, 0: {1}}
 
     """
-    node_dict = defaultdict(list)
+    node_dict = defaultdict(set)
     for edge_id, members in edge_dict.items():
         for node in members:
-            node_dict[node].append(edge_id)
+            node_dict[node].add(edge_id)
 
     return dict(node_dict)
 
