+++ tests/utils/test_utils.py
@@ -5,15 +5,15 @@ import xgi
 
 def test_dual_dict(dict5):
     dual = xgi.dual_dict(dict5)
-    assert dual[0] == [0]
-    assert dual[1] == [0]
-    assert dual[2] == [0]
-    assert dual[3] == [0]
-    assert dual[4] == [1]
-    assert dual[5] == [2]
-    assert dual[6] == [2, 3]
-    assert dual[7] == [3]
-    assert dual[8] == [3]
+    assert dual[0] == {0}
+    assert dual[1] == {0}
+    assert dual[2] == {0}
+    assert dual[3] == {0}
+    assert dual[4] == {1}
+    assert dual[5] == {2}
+    assert dual[6] == {2, 3}
+    assert dual[7] == {3}
+    assert dual[8] == {3}
 
 
 def test_powerset():
