+++ tests/classes/test_hypergraph.py
@@ -464,8 +464,12 @@ def test_double_edge_swap(edgelist1):
     with pytest.raises(IDNotFound):
         H.double_edge_swap(10, 3, 0, 1)
 
-    with pytest.raises(XGIError):
+    with pytest.raises(IDNotFound):
         H.double_edge_swap(8, 3, 0, 1)
+    
+    H = xgi.Hypergraph(edgelist1)
+    with pytest.raises(XGIError):
+        H.double_edge_swap(6, 7, 2, 3)
 
 
 def test_duplicate_edges(edgelist1):
