### 2023-05-04 50dfc08
Fix double edge swap (#349)
* fixed the double edge swap method

* added additional unit tests

* switched lists to sets

* fixed unit tests
