+++ requirements/documentation.txt
@@ -1,6 +1,6 @@
 sphinx~=6.0
 sphinx_copybutton
-sphinx-rtd-theme>=1.0
+sphinx-rtd-theme>=1.2
 numpydoc>=1.1
 pillow>=8.2
 matplotlib>=3.3
