+++ xgi/classes/reportviews.py
@@ -206,7 +206,7 @@ class IDView(Mapping, Set):
         See Also
         --------
         IDView.filterby_attr : For more details, see the `tutorial
-        <https://github.com/xgi-org/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+        <https://github.com/xgi-org/xgi-tutorials/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
 
         Examples
         --------
@@ -289,7 +289,7 @@ class IDView(Mapping, Set):
         See Also
         --------
         IDView.filterby : Identical method.  For more details, see the `tutorial
-        <https://github.com/xgi-org/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+        <https://github.com/xgi-org/xgi-tutorials/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
 
         Notes
         -----
@@ -537,7 +537,7 @@ class NodeView(IDView):
     In addition to the methods listed in this page, other methods defined in the `stats`
     package are also accessible via the `NodeView` class.  For more details, see the
     `tutorial
-    <https://github.com/xgi-org/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+    <https://github.com/xgi-org/xgi-tutorials/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
 
     """
 
