+++ .github/workflows/test.yml
@@ -1,4 +1,4 @@
-name: test
+name: Test codebase
 
 on:
   push:
@@ -9,48 +9,6 @@ on:
       - main
 
 jobs:
-  tutorials:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ['3.8', '3.9', '3.10', '3.11']
-    steps:
-    - uses: actions/checkout@v2
-
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-
-    - name: Install packages
-      run: |
-        sudo apt-get install -y rename
-        python -m pip install --upgrade pip wheel setuptools
-        python -m pip install -r requirements/default.txt -r requirements/test.txt -r requirements/tutorial.txt
-        python -m pip install .
-        python -m pip list
-
-    - name: Prepare tutorial notebooks
-      run: |
-        # go into tutorial location
-        cd tutorials/
-        # convert each notebook into a python file
-        jupyter nbconvert --to script *.ipynb
-        # rename the files so they are found by pytest: add 'test' at the beginning
-        rename 's/(.*)\.py/test $1.py/' *
-        # rename the files so they are found by pytest: change spaces to underscores
-        rename 's/ /_/g' test*
-        # modify each file by adding one level of indentation to each line
-        sed -i 's/^/    /' test*.py
-        # modify each file by adding 'def test_func():' at the top
-        sed -i '1s;^;def test_func():\n;' test*.py
-
-    - name: Test tutorial notebooks
-      run: |
-        # run pytest ONLY on the tutorials
-        cd tutorials/
-        pytest --color=yes
-
   pytest:
     runs-on: ${{ matrix.os }}-latest
     strategy:
