+++ xgi/classes/reportviews.py
@@ -116,7 +116,7 @@ class IDView(Mapping, Set):
         always use `x in view`.  The latter is always faster.
 
         """
-        return set(self._id_dict.keys()) if self._ids is None else self._ids
+        return set(self._id_dict) if self._ids is None else self._ids
 
     def __len__(self):
         """The number of IDs."""
@@ -125,7 +125,7 @@ class IDView(Mapping, Set):
     def __iter__(self):
         """Returns an iterator over the IDs."""
         if self._ids is None:
-            return iter({}) if self._id_dict is None else iter(self._id_dict.keys())
+            return iter({}) if self._id_dict is None else iter(self._id_dict)
         else:
             return iter(self._ids)
 
@@ -497,7 +497,7 @@ class IDView(Mapping, Set):
         newview._id_attr = view._id_attr
         newview._bi_id_dict = view._bi_id_dict
         newview._bi_id_attr = view._bi_id_attr
-        all_ids = set(view._id_dict.keys())
+        all_ids = set(view._id_dict)
         if bunch is None:
             newview._ids = all_ids
         else:
