### 2023-05-05 0219bf8
added empty_edges keyword
