+++ xgi/classes/hypergraph.py
@@ -441,7 +441,7 @@ class Hypergraph:
 
         strong : bool, optional
             Whether to execute weak or strong removal. By default, False.
-        
+
         empty_edges : bool, optional
             Whether to allow empty edges after all nodes are removed from
             an edge, by default False.
@@ -844,15 +844,14 @@ class Hypergraph:
             temp_members1.add(n_id2)
             temp_members2.add(n_id1)
 
-            # Now we handle the memberships   
+            # Now we handle the memberships
             # remove old nodes from edges
             temp_memberships1.remove(e_id1)
             temp_memberships2.remove(e_id2)
 
             # swap nodes
             temp_memberships1.add(e_id2)
-            temp_memberships2.add(e_id1)    
-
+            temp_memberships2.add(e_id1)
 
         except KeyError as e:
 
@@ -860,11 +859,12 @@ class Hypergraph:
                 "One of the nodes specified doesn't belong to the specified edge."
             ) from e
 
-        if (len(temp_memberships1) != len(self._node[n_id1]) or
-            len(temp_memberships2) != len(self._node[n_id2]) or
-            len(temp_members1) != len(self._edge[e_id1]) or
-            len(temp_members2) != len(self._edge[e_id2])
-            ):
+        if (
+            len(temp_memberships1) != len(self._node[n_id1])
+            or len(temp_memberships2) != len(self._node[n_id2])
+            or len(temp_members1) != len(self._edge[e_id1])
+            or len(temp_members2) != len(self._edge[e_id2])
+        ):
             raise XGIError("This swap does not preserve edge sizes.")
 
         self._node[n_id1] = temp_memberships1
@@ -1002,7 +1002,7 @@ class Hypergraph:
             raise XGIError(f"Edge {edge} not in the hypergraph") from e
         except ValueError as e:
             raise XGIError(f"Edge {edge} does not contain node {node}") from e
-        
+
         if not self._edge[edge] and not empty_edges:
             del self._edge[edge]
             del self._edge_attr[edge]
