### 2023-05-06 a58a4b8
Clean up classes (#353)
* moved IDDict

* remove keys

* remove extraneous arguments

* style: format with black
