+++ xgi/utils/utilities.py
@@ -3,12 +3,44 @@
 from collections import defaultdict
 from itertools import chain, combinations, count
 
+from xgi.exception import IDNotFound, XGIError
+
 __all__ = [
+    "IDDict",
     "dual_dict",
     "powerset",
     "update_uid_counter",
     "find_triangles",
 ]
+
+
+class IDDict(dict):
+    """A dict that holds (node or edge) IDs.
+
+    For internal use only.  Adds input validation functionality to the internal dicts
+    that hold nodes and edges in a network.
+
+    """
+
+    def __getitem__(self, item):
+        try:
+            return dict.__getitem__(self, item)
+        except KeyError as e:
+            raise IDNotFound(f"ID {item} not found") from e
+
+    def __setitem__(self, item, value):
+        if item is None:
+            raise XGIError("None cannot be a node or edge")
+        try:
+            return dict.__setitem__(self, item, value)
+        except TypeError as e:
+            raise TypeError(f"ID {item} not a valid type") from e
+
+    def __delitem__(self, item):
+        try:
+            return dict.__delitem__(self, item)
+        except KeyError as e:
+            raise IDNotFound(f"ID {item} not found") from e
 
 
 def dual_dict(edge_dict):
