+++ xgi/classes/hypergraph.py
@@ -5,7 +5,7 @@ from copy import copy, deepcopy
 from itertools import count
 from warnings import warn
 
-from ..exception import XGIError
+from ..exception import IDNotFound, XGIError
 from ..utils import IDDict, update_uid_counter
 from .reportviews import EdgeView, NodeView
 
