+++ xgi/classes/direportviews.py
@@ -79,15 +79,19 @@ class IDView(Mapping, Set):
     def __init__(self, network, ids=None):
         self._net = network
 
-        if self._id_kind == "node":
-            self._id_dict = None if self._net is None else network._node
+        if self._id_kind == "dinode":
+            self._in_id_dict = None if self._net is None else network._node_in
+            self._out_id_dict = None if self._net is None else network._node_out
             self._id_attr = None if self._net is None else network._node_attr
-            self._bi_id_dict = None if self._net is None else network._edge
+            self._bi_in_id_dict = None if self._net is None else network._edge_in
+            self._bi_out_id_dict = None if self._net is None else network._edge_out
             self._bi_id_attr = None if self._net is None else network._edge_attr
-        elif self._id_kind == "edge":
-            self._id_dict = None if self._net is None else network._edge
+        elif self._id_kind == "diedge":
+            self._in_id_dict = None if self._net is None else network._edge_in
+            self._out_id_dict = None if self._net is None else network._edge_out
             self._id_attr = None if self._net is None else network._edge_attr
-            self._bi_id_dict = None if self._net is None else network._node
+            self._bi_in_id_dict = None if self._net is None else network._node_in
+            self._bi_out_id_dict = None if self._net is None else network._node_out
             self._bi_id_attr = None if self._net is None else network._node_attr
 
         if ids is None:
@@ -333,141 +337,6 @@ class IDView(Mapping, Set):
             )
         return type(self).from_view(self, bunch)
 
-    def neighbors(self, id):
-        """Find the neighbors of an ID.
-
-        The neighbors of an ID are those IDs that share at least one bipartite ID.
-
-        Parameters
-        ----------
-        id : hashable
-            ID to find neighbors of.
-        Returns
-        -------
-        set
-            A set of the neighboring IDs
-
-        See Also
-        --------
-        ~xgi.classes.function.edge_neighborhood
-
-        Examples
-        --------
-        >>> import xgi
-        >>> hyperedge_list = [[1, 2], [2, 3, 4]]
-        >>> H = xgi.Hypergraph(hyperedge_list)
-        >>> H.nodes.neighbors(1)
-        {2}
-        >>> H.nodes.neighbors(2)
-        {1, 3, 4}
-
-        """
-        return {i for n in self._id_dict[id] for i in self._bi_id_dict[n]}.difference(
-            {id}
-        )
-
-    def duplicates(self):
-        """Find IDs that have a duplicate.
-
-        An ID has a 'duplicate' if there exists another ID with the same bipartite
-        neighbors.
-
-        Returns
-        -------
-        IDView
-            A view containing only those IDs with a duplicate.
-
-        Raises
-        ------
-        TypeError
-            When IDs are of different types. For example, ("a", 1).
-
-        Notes
-        -----
-        The IDs returned are in an arbitrary order, that is duplicates are not
-        guaranteed to be consecutive. For IDs with the same bipartite neighbors,
-        only the first ID added is not a duplicate.
-
-        See Also
-        --------
-        IDView.lookup
-
-        Examples
-        --------
-        >>> import xgi
-        >>> H = xgi.Hypergraph([[0, 1, 2], [3, 4, 2], [0, 1, 2]])
-        >>> H.edges.duplicates()
-        EdgeView((2,))
-
-        Order does not matter:
-
-        >>> H = xgi.Hypergraph([[2, 1, 0], [0, 1, 2]])
-        >>> H.edges.duplicates()
-        EdgeView((1,))
-
-        Repetitions matter:
-
-        >>> H = xgi.Hypergraph([[0, 1], [1, 0]])
-        >>> H.edges.duplicates()
-        EdgeView((1,))
-
-        """
-        dups = []
[…]
