+++ xgi/classes/reportviews.py
@@ -501,7 +501,7 @@ class IDView(Mapping, Set):
             bunch = set(bunch)
             wrong = bunch - all_ids
             if wrong:
-                raise IDNotFound(f"Nodes {wrong} not in the hypergraph")
+                raise IDNotFound(f"IDs {wrong} not in the hypergraph")
             newview._ids = bunch
         return newview
 
@@ -569,7 +569,7 @@ class NodeView(IDView):
 
         """
         return (
-            {key: self._id_dict[key] for key in self}
+            {key: self._id_dict[key].copy() for key in self}
             if n is None
             else self._id_dict[n].copy()
         )
@@ -675,9 +675,9 @@ class EdgeView(IDView):
         """
         if e is None:
             if dtype is dict:
-                return {key: self._id_dict[key] for key in self}
+                return {key: self._id_dict[key].copy() for key in self}
             elif dtype is list:
-                return [self._id_dict[key] for key in self]
+                return [self._id_dict[key].copy() for key in self]
             else:
                 raise XGIError(f"Unrecognized dtype {dtype}")
 
