### 2023-05-07 1ed990c
added directed node and edge stats
