+++ xgi/stats/__init__.py
@@ -50,7 +50,7 @@ from scipy.stats import moment as spmoment
 
 from xgi.exception import IDNotFound
 
-from . import edgestats, nodestats
+from . import edgestats, diedgestats, dinodestats, nodestats
 
 __all__ = ["nodestat_func", "edgestat_func", "dispatch_stat", "dispatch_many_stats"]
 
@@ -208,7 +208,27 @@ class NodeStat(IDStat):
     """
 
 
+class DiNodeStat(IDStat):
+    """An arbitrary node-quantity mapping.
+
+    `NodeStat` objects represent a mapping that assigns a value to each node in a
+    network.  For more details, see the `tutorial
+    <https://github.com/xgi-org/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+
+    """
+
+
 class EdgeStat(IDStat):
+    """An arbitrary edge-quantity mapping.
+
+    `EdgeStat` objects represent a mapping that assigns a value to each edge in a
+    network.  For more details, see the `tutorial
+    <https://github.com/xgi-org/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+
+    """
+
+
+class DiEdgeStat(IDStat):
     """An arbitrary edge-quantity mapping.
 
     `EdgeStat` objects represent a mapping that assigns a value to each edge in a
@@ -400,6 +420,18 @@ class MultiNodeStat(MultiIDStat):
     statsmodule = nodestats
 
 
+class MultiDiNodeStat(MultiIDStat):
+    """Multiple node-quantity mappings.
+
+    For more details, see the `tutorial
+    <https://github.com/xgi-org/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+
+    """
+
+    statsclass = DiNodeStat
+    statsmodule = dinodestats
+
+
 class MultiEdgeStat(MultiIDStat):
     """Multiple edge-quantity mappings.
 
@@ -412,16 +444,38 @@ class MultiEdgeStat(MultiIDStat):
     statsmodule = edgestats
 
 
+class MultiDiEdgeStat(MultiIDStat):
+    """Multiple edge-quantity mappings.
+
+    For more details, see the `tutorial
+    <https://github.com/xgi-org/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
+
+    """
+
+    statsclass = DiEdgeStat
+    statsmodule = diedgestats
+
+
 _dispatch_data = {
     "node": {
         "module": nodestats,
         "statclass": NodeStat,
         "multistatclass": MultiNodeStat,
     },
+    "dinode": {
+        "module": dinodestats,
+        "statclass": DiNodeStat,
+        "multistatclass": MultiDiNodeStat,
+    },
     "edge": {
         "module": edgestats,
         "statclass": EdgeStat,
         "multistatclass": MultiEdgeStat,
+    },
+    "diedge": {
+        "module": diedgestats,
+        "statclass": DiEdgeStat,
+        "multistatclass": MultiDiEdgeStat,
     },
 }
 
