+++ xgi/classes/dihypergraph.py
@@ -500,92 +500,6 @@ class DiHypergraph:
 
             return
 
-        # in formats 1-4 we only know that ebunch_to_add is an iterable, so we iterate
-        # over it and use the firs element to determine which format we are working with
-        new_edges = iter(ebunch_to_add)
-        try:
-            first_edge = next(new_edges)
-        except StopIteration:
-            return
-        try:
-            first_elem = list(first_edge)[0]
-        except TypeError:
-            first_elem = None
-
-        format1, format2, format3, format4 = False, False, False, False
-        if not isinstance(first_edge, dict):
-            if len(first_edge) == 2 and issubclass(type(first_edge[1]), Hashable):
-                format2 = True
-            elif len(first_edge) == 2:
-                format3 = True
-            elif len(first_edge) == 3:
-                format4 = True
-        else:
-            format1 = True
-
-        if (format1 and isinstance(first_edge, str)) or (
-            not format1 and isinstance(first_elem, str)
-        ):
-            raise XGIError("Members cannot be specified as a string")
-
-        # now we may iterate over the rest
-        e = first_edge
-        while True:
-            if format1:
-                members, id, eattr = e, next(self._edge_uid), {}
-            elif format2:
-                members, id, eattr = e[0], e[1], {}
-            elif format3:
-                members, id, eattr = e[0], next(self._edge_uid), e[1]
-            elif format4:
-                members, id, eattr = e[0], e[1], e[2]
-
-            if id in self._edge_in.keys():  # check that uid is not present yet
-                warn(f"uid {id} already exists, cannot add edge {members}.")
-            else:
-                if isinstance(members, (tuple, list)):
-                    tail = members[0]
-                    head = members[1]
-                elif type(members) == dict:
-                    tail = set(members["tail"])
-                    head = set(members["head"])
-                else:
-                    raise XGIError(
-                        "Directed edge must be a dictionary, list, or tuple!"
-                    )
-
-                try:
-                    self._edge_in[id] = set(head)
-                    self._edge_out[id] = set(tail)
-                except TypeError as e:
-                    raise XGIError("Invalid ebunch format") from e
-
-                for n in head:
-                    if n not in self._node_out:
-                        self._node_in[n] = set()
-                        self._node_out[n] = set()
-                        self._node_attr[n] = self._node_attr_dict_factory()
-                    self._node_out[n].add(id)
-
-                for n in tail:
-                    if n not in self._node_in:
-                        self._node_in[n] = set()
-                        self._node_out[n] = set()
-                        self._node_attr[n] = self._node_attr_dict_factory()
-                    self._node_in[n].add(id)
-
-                self._edge_attr[id] = self._hyperedge_attr_dict_factory()
-                self._edge_attr[id].update(attr)
-                self._edge_attr[id].update(eattr)
-
-            try:
-                e = next(new_edges)
-            except StopIteration:
-
-                if format2 or format4:
-                    update_uid_counter(self, id)
-                break
-
     def remove_edge(self, id):
         """Remove one edge.
 

[…]
