### 2023-05-07 d0e6cc4
added directed node and edge stats
