+++ xgi/classes/reportviews.py
@@ -16,6 +16,8 @@ from ..stats import IDStat, dispatch_many_stats, dispatch_stat
 __all__ = [
     "NodeView",
     "EdgeView",
+    "DiNodeView",
+    "DiEdgeView",
 ]
 
 
@@ -771,3 +773,680 @@ class EdgeView(IDView):
                         max_edges.update(dups[frozenset(e)])
 
         return self.from_view(self, bunch=max_edges)
+
+
+class DiIDView(Mapping, Set):
+    """Base View class for accessing the ids (nodes or edges) of a Hypergraph.
+
+    Can optionally keep track of a subset of ids.  By default all node ids or all edge
+    ids are kept track of.
+
+    Parameters
+    ----------
+    id_dict : dict
+        The original dict this is a view of.
+    id_attrs : dict
+        The original attribute dict this is a view of.
+    ids : iterable
+        A subset of the keys in id_dict to keep track of.
+
+    Raises
+    ------
+    XGIError
+        If ids is not a subset of the keys of id_dict.
+
+    """
+
+    _id_kind = None
+
+    __slots__ = (
+        "_net",
+        "_ids",
+    )
+
+    def __getstate__(self):
+        """Function that allows pickling.
+
+        Returns
+        -------
+        dict
+            The keys access the IDs and their attributes respectively
+            and the values are dictionarys from the Hypergraph class.
+
+        """
+        return {
+            "_net": self._net,
+            "_ids": self._ids,
+        }
+
+    def __setstate__(self, state):
+        """Function that allows unpickling.
+
+        Parameters
+        ----------
+        dict
+            The keys access the IDs and their attributes respectively
+            and the values are dictionarys from the Hypergraph class.
+
+        """
+        self._net = state["_net"]
+        self._id_kind = state["_id_kind"]
+
+    def __init__(self, network, ids=None):
+        self._net = network
+
+        if self._id_kind == "dinode":
+            self._in_id_dict = None if self._net is None else network._node_in
+            self._out_id_dict = None if self._net is None else network._node_out
+            self._id_attr = None if self._net is None else network._node_attr
+            self._bi_in_id_dict = None if self._net is None else network._edge_in
+            self._bi_out_id_dict = None if self._net is None else network._edge_out
+            self._bi_id_attr = None if self._net is None else network._edge_attr
+        elif self._id_kind == "diedge":
+            self._in_id_dict = None if self._net is None else network._edge_in
+            self._out_id_dict = None if self._net is None else network._edge_out
+            self._id_attr = None if self._net is None else network._edge_attr
+            self._bi_in_id_dict = None if self._net is None else network._node_in
+            self._bi_out_id_dict = None if self._net is None else network._node_out
+            self._bi_id_attr = None if self._net is None else network._node_attr
+
+        if ids is None:
+            self._ids = self._in_id_dict
+        else:
+            self._ids = ids
+
+    def __getattr__(self, attr):
+        stat = dispatch_stat(self._id_kind, self._net, self, attr)
+        self.__dict__[attr] = stat
+        return stat
+
+    def multi(self, names):
+        return dispatch_many_stats(self._id_kind, self._net, self, names)
+
+    @property
+    def ids(self):
+        """The ids in this view.
+
+        Notes
+        -----
+        Do not use this property for membership check. Instead of `x in view.ids`,
+        always use `x in view`.  The latter is always faster.
+
+        """
+        return set(self._in_id_dict) if self._ids is None else self._ids
+
+    def __len__(self):
+        """The number of IDs."""
+        return len(self._in_id_dict) if self._ids is None else len(self._ids)
+
+    def __iter__(self):
[…]
