+++ docs/source/api/classes/xgi.classes.reportviews.rst
@@ -14,3 +14,6 @@
         IDView
         NodeView
         EdgeView
+        DiIDView
+        DiNodeView
+        DiEdgeView
