+++ docs/source/api/classes.rst
@@ -8,6 +8,7 @@ classes package
    :toctree: classes
 
    ~xgi.classes.hypergraph
+   ~xgi.classes.dihypergraph
    ~xgi.classes.simplicialcomplex
    ~xgi.classes.reportviews
    ~xgi.classes.hypergraphviews
