+++ docs/source/api/classes/xgi.classes.dihypergraph.DiHypergraph.rst
@@ -1,0 +1,33 @@
+﻿xgi.classes.dihypergraph.DiHypergraph
+=====================================
+
+.. currentmodule:: xgi.classes.dihypergraph
+
+.. autoclass:: DiHypergraph
+   :show-inheritance:
+   :members:
+
+
+   .. rubric:: Attributes
+
+   .. autosummary::
+
+      ~Hypergraph.edges
+      ~Hypergraph.nodes
+      ~Hypergraph.num_edges
+      ~Hypergraph.num_nodes
+
+
+   .. rubric:: Methods that modify the structure
+
+   .. autosummary::
+      :nosignatures:
+
+      ~Hypergraph.add_node
+      ~Hypergraph.add_edge
+      ~Hypergraph.add_nodes_from
+      ~Hypergraph.add_edges_from
+      ~Hypergraph.remove_node
+      ~Hypergraph.remove_edge
+      ~Hypergraph.remove_nodes_from
+      ~Hypergraph.remove_edges_from
