+++ docs/source/api/classes/xgi.classes.reportviews.DiNodeView.rst
@@ -1,0 +1,26 @@
+xgi.classes.reportviews.DiNodeView
+==================================
+
+.. currentmodule:: xgi.classes.reportviews
+
+.. autoclass:: DiNodeView
+   :show-inheritance:
+   :members:
+
+
+   .. rubric:: Attributes
+
+   .. autosummary::
+
+      ~DiNodeView.ids
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+      :nosignatures:
+
+      ~DiNodeView.memberships
+      ~DiNodeView.dimemberships
+      ~IDView.filterby
+      ~IDView.filterby_attr
