+++ docs/source/api/classes/xgi.classes.reportviews.DiIDView.rst
@@ -1,0 +1,18 @@
+xgi.classes.reportviews.DiIDView
+================================
+
+.. currentmodule:: xgi.classes.reportviews
+
+.. autoclass:: DiIDView
+   :show-inheritance:
+   :members:
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+      :nosignatures:
+
+      ~IDView.from_view
+      ~IDView.filterby
+      ~IDView.filterby_attr
