+++ docs/source/api/classes/xgi.classes.reportviews.DiEdgeView.rst
@@ -1,0 +1,28 @@
+xgi.classes.reportviews.DiEdgeView
+==================================
+
+.. currentmodule:: xgi.classes.reportviews
+
+.. autoclass:: DiEdgeView
+   :show-inheritance:
+   :members:
+
+
+   .. rubric:: Attributes
+
+   .. autosummary::
+
+      ~DiEdgeView.ids
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+      :nosignatures:
+
+      ~DiEdgeView.members
+      ~DiEdgeView.dimembers
+      ~DiEdgeView.head
+      ~DiEdgeView.tail
+      ~IDView.filterby
+      ~IDView.filterby_attr
