+++ docs/source/api/classes/xgi.classes.dihypergraph.DiHypergraph.rst
@@ -12,10 +12,10 @@
 
    .. autosummary::
 
-      ~Hypergraph.edges
-      ~Hypergraph.nodes
-      ~Hypergraph.num_edges
-      ~Hypergraph.num_nodes
+      ~DiHypergraph.edges
+      ~DiHypergraph.nodes
+      ~DiHypergraph.num_edges
+      ~DiHypergraph.num_nodes
 
 
    .. rubric:: Methods that modify the structure
@@ -23,11 +23,12 @@
    .. autosummary::
       :nosignatures:
 
-      ~Hypergraph.add_node
-      ~Hypergraph.add_edge
-      ~Hypergraph.add_nodes_from
-      ~Hypergraph.add_edges_from
-      ~Hypergraph.remove_node
-      ~Hypergraph.remove_edge
-      ~Hypergraph.remove_nodes_from
-      ~Hypergraph.remove_edges_from
+      ~DiHypergraph.add_node
+      ~DiHypergraph.add_edge
+      ~DiHypergraph.add_nodes_from
+      ~DiHypergraph.add_edges_from
+      ~DiHypergraph.remove_node
+      ~DiHypergraph.remove_edge
+      ~DiHypergraph.remove_nodes_from
+      ~DiHypergraph.remove_edges_from
+      ~DiHypergraph.clear
