### 2023-05-08 8da01cf
perf: faster random_flag_complex by adding only necessary cliques. tests: moved for consistency and added corresponding
