+++ tests/generators/test_simplicial_complexes.py
@@ -1,4 +1,5 @@
 import networkx as nx
+import pytest
 
 import xgi
 from xgi.exception import XGIError
@@ -44,3 +45,113 @@ def test_flag_complex_d2():
     S2 = xgi.flag_complex_d2(G)
 
     assert set(S.edges.members()) == set(S2.edges.members())
+
+
+def test_random_simplicial_complex():
+    # seed
+    S1 = xgi.random_simplicial_complex(10, [0.1, 0.001], seed=1)
+    S2 = xgi.random_simplicial_complex(10, [0.1, 0.001], seed=2)
+    S3 = xgi.random_simplicial_complex(10, [0.1, 0.001], seed=2)
+
+    assert S1._edge != S2._edge
+    assert S2._edge == S3._edge
+
+    # wrong input
+    with pytest.raises(ValueError):
+        S1 = xgi.random_simplicial_complex(10, [1, 1.1])
+    with pytest.raises(ValueError):
+        S1 = xgi.random_simplicial_complex(10, [1, -2])
+
+
+def test_random_flag_complex():
+
+    S = xgi.random_flag_complex(10, 0.4, seed=2)
+    simplices = {frozenset({0, 4}),
+                 frozenset({0, 7}),
+                 frozenset({1, 8}),
+                 frozenset({2, 5}),
+                 frozenset({2, 9}),
+                 frozenset({3, 5}),
+                 frozenset({3, 6}),
+                 frozenset({3, 7}),
+                 frozenset({3, 8}),
+                 frozenset({4, 5}),
+                 frozenset({4, 7}),
+                 frozenset({4, 8}),
+                 frozenset({6, 7}),
+                 frozenset({6, 8}),
+                 frozenset({7, 8}),
+                 frozenset({0, 4, 7}),
+                 frozenset({3, 6, 7}),
+                 frozenset({3, 6, 8}),
+                 frozenset({3, 7, 8}),
+                 frozenset({4, 7, 8}),
+                 frozenset({6, 7, 8})}
+
+    assert set(S.edges.members()) == simplices
+
+    # max_order
+    S = xgi.random_flag_complex(10, 0.4, seed=2, max_order=3)
+    assert set(S.edges.members()) == simplices.union({frozenset({3, 6, 7, 8})})
+
+
+    # seed
+    S1 = xgi.random_flag_complex(10, 0.1, seed=1)
+    S2 = xgi.random_flag_complex(10, 0.1, seed=2)
+    S3 = xgi.random_flag_complex(10, 0.1, seed=2)
+
+    assert S1._edge != S2._edge
+    assert S2._edge == S3._edge
+
+    # wrong input
+    with pytest.raises(ValueError):
+        S1 = xgi.random_flag_complex(10, 1.1)
+    with pytest.raises(ValueError):
+        S1 = xgi.random_flag_complex(10, -2)
+
+
+def test_random_flag_complex_d2():
+
+    S = xgi.random_flag_complex_d2(10, 0.4, seed=2)
+    simplices = {frozenset({0, 4}),
+                 frozenset({0, 7}),
+                 frozenset({1, 8}),
+                 frozenset({2, 5}),
+                 frozenset({2, 9}),
+                 frozenset({3, 5}),
+                 frozenset({3, 6}),
+                 frozenset({3, 7}),
+                 frozenset({3, 8}),
+                 frozenset({4, 5}),
+                 frozenset({4, 7}),
+                 frozenset({4, 8}),
+                 frozenset({6, 7}),
+                 frozenset({6, 8}),
+                 frozenset({7, 8}),
+                 frozenset({0, 4, 7}),
[…]
