### 2023-05-08 36583fe
perf: same for flag_complex
