+++ docs/source/api/stats/xgi.stats.diedgestat_func.rst
@@ -1,0 +1,6 @@
+﻿xgi.stats.diedgestat\_func
+==========================
+
+.. currentmodule:: xgi.stats
+
+.. autofunction:: diedgestat_func
