+++ docs/source/api/stats/xgi.stats.DiNodeStat.rst
@@ -1,0 +1,22 @@
+﻿xgi.stats.DiNodeStat
+====================
+
+.. currentmodule:: xgi.stats
+
+.. autoclass:: DiNodeStat
+   :show-inheritance:
+   :members:
+
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+      
+      ~DiNodeStat.name
+   
+
+   .. rubric:: Methods
+
+   .. autosummary::
+      :nosignatures:
+      
