+++ docs/source/api/utils/xgi.utils.utilities.IDDict.rst
@@ -1,0 +1,14 @@
+﻿xgi.utils.utilities.IDDict
+==========================
+
+.. currentmodule:: xgi.utils.utilities
+
+.. autoclass:: IDDict
+   :show-inheritance:
+   :members:
+
+   .. rubric:: Methods
+
+   .. autosummary::
+      :nosignatures:
+      
