+++ docs/source/api/classes/xgi.classes.hypergraph.IDDict.rst
@@ -1,14 +1,0 @@
-﻿xgi.classes.hypergraph.IDDict
-=============================
-
-.. currentmodule:: xgi.classes.hypergraph
-
-.. autoclass:: IDDict
-   :show-inheritance:
-   :members:
-
-
-   .. rubric:: Methods
-
-   .. autosummary::
-      :nosignatures:
