+++ xgi/classes/dihypergraph.py
@@ -13,20 +13,16 @@ __all__ = ["DiHypergraph"]
 
 class DiHypergraph:
     r"""A directed hypergraph (dihypergraph) is a collection of directed
-    interactions
+    interactions where each interaction has both senders and receivers.
 
+    More formally, a dihypergraph is a pair :math:`(V, E)`, where :math:`V` is
+    a set of elements called *nodes* or *vertices*, and :math:`E` is the set of
+    directed hyperedges. A directed hyperedge is an ordered pair, $(e^+, e^-)$,
+    where $e^+ \subset V$, the set of senders, is known as the "tail" and $e^-\subset V$,
+    the set of receivers, is known as the "head". The equivalent undirected edge,
+    is $e = e^+ \cap e^-$ and the edge size is defined as $|e|$.
 
-    ordered pairs,
-    $(e^+, e^-)$, where $e^+$ is known as the tail and is the set of senders in
-    this interaction, and $e^-$ is known as the head and is the set of receivers
-    in the interaction.
-
-    A hypergraph is a pair :math:`(V, E)`, where :math:`V` is a set of elements called
-    *nodes* or *vertices*, and :math:`E` is a set whose elements are subsets of
-    :math:`V`, that is, each :math:`e \in E` satisfies :math:`e \subset V`.  The
-    elements of :math:`E` are called *hyperedges* or simply *edges*.
-
-    The Hypergraph class allows any hashable object as a node and can associate
+    The DiHypergraph class allows any hashable object as a node and can associate
     attributes to each node, edge, or the hypergraph itself, in the form of key/value
     pairs.
 
@@ -39,11 +35,9 @@ class DiHypergraph:
         hypergraph is created, i.e. one with no nodes or edges.
         The data can be in the following formats:
 
-        * hyperedge list
-        * hyperedge dictionary
-        * 2-column Pandas dataframe (bipartite edges)
-        * Scipy/Numpy incidence matrix
-        * Hypergraph object.
+        * directed hyperedge list
+        * directed hyperedge dictionary
+        * DiHypergraph object.
 
     **attr : dict, optional, default: None
         Attributes to add to the hypergraph as key, value pairs.
@@ -57,19 +51,18 @@ class DiHypergraph:
     attributes see :meth:`add_node` and :meth:`add_edge`.
 
     In addition to the methods listed in this page, other methods defined in the `stats`
-    package are also accessible via the `Hypergraph` class.  For more details, see the
+    package are also accessible via the `DiHypergraph` class.  For more details, see the
     `tutorial
     <https://github.com/ComplexGroupInteractions/xgi/blob/main/tutorials/Tutorial%206%20-%20Statistics.ipynb>`_.
 
     Examples
     --------
     >>> import xgi
-    >>> H = xgi.Hypergraph([[1, 2, 3], [4], [5, 6], [6, 7, 8]])
+    >>> H = xgi.DiHypergraph([([1, 2, 3], [4]), ([5, 6], [6, 7, 8])])
     >>> H.nodes
-    NodeView((1, 2, 3, 4, 5, 6, 7, 8))
+    DiNodeView((1, 2, 3, 4, 5, 6, 7, 8))
     >>> H.edges
-    EdgeView((0, 1, 2, 3))
-
+    DiEdgeView((0, 1))
     """
     _node_dict_factory = IDDict
     _node_attr_dict_factory = IDDict
@@ -119,8 +112,8 @@ class DiHypergraph:
         Examples
         --------
         >>> import xgi
-        >>> hyperedge_list = [[1, 2], [2, 3, 4]]
-        >>> H = xgi.Hypergraph(hyperedge_list)
[…]
