+++ xgi/classes/reportviews.py
@@ -185,8 +185,8 @@ class IDView(Mapping, Set):
 
         Parameters
         ----------
-        stat : str or :class:`xgi.stats.NodeStat`
-            `NodeStat` object, or name of a `NodeStat`.
+        stat : str or :class:`xgi.stats.NodeStat`/:class:`xgi.stats.EdgeStat`
+            `NodeStat`/`EdgeStat` object, or name of a `NodeStat`/`EdgeStat`.
         val : Any
             Value of the statistic.  Usually a single numeric value.  When mode is
             'between', must be a tuple of exactly two values.
@@ -778,7 +778,7 @@ class EdgeView(IDView):
 
 
 class DiIDView(Mapping, Set):
-    """Base View class for accessing the ids (nodes or edges) of a Hypergraph.
+    """Base View class for accessing the ids (nodes or edges) of a DiHypergraph.
 
     Can optionally keep track of a subset of ids.  By default all node ids or all edge
     ids are kept track of.
@@ -913,7 +913,7 @@ class DiIDView(Mapping, Set):
         return self._id_attr[id]
 
     def __contains__(self, id):
-        """Checks whether the ID is in the hypergraph"""
+        """Checks whether the ID is in the dihypergraph"""
         return id in self._ids
 
     def __str__(self):
@@ -945,8 +945,8 @@ class DiIDView(Mapping, Set):
 
         Parameters
         ----------
-        stat : str or :class:`xgi.stats.NodeStat`
-            `NodeStat` object, or name of a `NodeStat`.
+        stat : str or :class:`xgi.stats.DiNodeStat`/:class:`xgi.stats.DiEdgeStat`
+            `DiNodeStat`/`DiEdgeStat` object, or name of a `DiNodeStat`/`DiEdgeStat`.
         val : Any
             Value of the statistic.  Usually a single numeric value.  When mode is
             'between', must be a tuple of exactly two values.
@@ -973,33 +973,27 @@ class DiIDView(Mapping, Set):
         `val`.
 
         >>> import xgi
-        >>> H = xgi.Hypergraph([[1, 2, 3], [2, 3, 4, 5], [3, 4, 5]])
+        >>> H = xgi.DiHypergraph([([1, 2, 3], [2, 3, 4, 5]), ([3, 4, 5], [])])
         >>> n = H.nodes
         >>> n.filterby('degree', 2)
-        NodeView((2, 4, 5))
+        DiNodeView((3, 4, 5))
 
         Can choose other comparison methods via `mode`.
 
         >>> n.filterby('degree', 2, 'eq')
-        NodeView((2, 4, 5))
+        DiNodeView((3, 4, 5))
         >>> n.filterby('degree', 2, 'neq')
-        NodeView((1, 3))
+        DiNodeView((1, 2))
         >>> n.filterby('degree', 2, 'lt')
-        NodeView((1,))
+        DiNodeView((1, 2))
         >>> n.filterby('degree', 2, 'gt')
-        NodeView((3,))
+        DiNodeView(())
         >>> n.filterby('degree', 2, 'leq')
-        NodeView((1, 2, 4, 5))
+        DiNodeView((1, 2, 3, 4, 5))
         >>> n.filterby('degree', 2, 'geq')
-        NodeView((2, 3, 4, 5))
+        DiNodeView((3, 4, 5))
         >>> n.filterby('degree', (2, 3), 'between')
-        NodeView((2, 3, 4, 5))
-
-        Can also pass a :class:`NodeStat` object.
-
-        >>> n.filterby(n.degree(order=2), 2)
-        NodeView((3,))
-
+        DiNodeView((3, 4, 5))
         """
         if not isinstance(stat, IDStat):
             try:
[…]
