+++ tests/classes/test_dihypergraph.py
@@ -118,162 +118,166 @@ def test_dimemberships(diedgelist1):
 
 
 def test_add_edge():
-    for edge in [[1, 2, 3], {1, 2, 3}, iter([1, 2, 3])]:
-        H = xgi.Hypergraph()
+    for edge in [([1, 2, 3], [4]), [{1, 2, 3}, {4}], (iter([1, 2, 3]), iter([4]))]:
+        H = xgi.DiHypergraph()
         H.add_edge(edge)
-        assert (1 in H) and (2 in H) and (3 in H)
+        assert (1 in H) and (2 in H) and (3 in H) and (4 in H)
         assert 0 in H.edges
-        assert {1, 2, 3} in H.edges.members()
-        assert {1, 2, 3} == H.edges.members(0)
-        assert H.edges.members(dtype=dict) == {0: {1, 2, 3}}
+        assert {1, 2, 3, 4} in H.edges.members()
+        assert {1, 2, 3, 4} == H.edges.members(0)
+        assert H.edges.members(dtype=dict) == {0: {1, 2, 3, 4}}
+        assert H.edges.tail(dtype=dict) == {0: {1, 2, 3}}
+        assert H.edges.head(dtype=dict) == {0: {4}}
 
-    H = xgi.Hypergraph()
-    for edge in [[], set(), iter([])]:
+    H = xgi.DiHypergraph()
+    for edge in [[], {}, iter([])]:
         with pytest.raises(XGIError):
             H.add_edge(edge)
 
+    for edge in [[[], []], (set(), set())]:
+        with pytest.raises(XGIError):
+            H.add_edge(edge)
+
+    # type can't be a set.
+    with pytest.raises(XGIError):
+        H.add_edge({(1, 2), (3, 4)})
+
     # check that uid works correctly
-    H1 = xgi.Hypergraph()
-    H1.add_edge([1, 2], id=0)
-    H1.add_edge([3, 4], id=2)
-    H1.add_edge([5, 6])
-    assert H1._edge == {0: {1, 2}, 2: {3, 4}, 3: {5, 6}}
+    H1 = xgi.DiHypergraph()
+    H1.add_edge(([1, 2], [3]), id=0)
+    H1.add_edge(([3, 4], [4, 5]), id=2)
+    H1.add_edge([[5, 6], [2, 3]])
+    assert H1.edges.dimembers(dtype=dict) == {
+        0: ({1, 2}, {3}),
+        2: ({3, 4}, {4, 5}),
+        3: ({5, 6}, {2, 3}),
+    }
 
-    H2 = xgi.Hypergraph()
-    H2.add_edge([1, 2])
-    H2.add_edge([3, 4])
-    with pytest.warns(
-        UserWarning, match="uid 0 already exists, cannot add edge {5, 6}"
-    ):
-        H2.add_edge([5, 6], id=0)
+    H2 = xgi.DiHypergraph()
+    H2.add_edge(([1, 2], [3]))
+    H2.add_edge(([3, 4], [5, 6, 7]))
+    with pytest.warns(Warning):
+        H2.add_edge(([5, 6], [8]), id=0)
 
-    assert H2._edge == {0: {1, 2}, 1: {3, 4}}
+    assert H2._edge_out == {0: {1, 2}, 1: {3, 4}}
 
 
 def test_add_edge_with_id():
-    H = xgi.Hypergraph()
-    H.add_edge([1, 2, 3], id="myedge")
-    assert (1 in H) and (2 in H) and (3 in H)
+    H = xgi.DiHypergraph()
+    H.add_edge(([1, 2, 3], [3, 4]), id="myedge")
+    assert (1 in H) and (2 in H) and (3 in H) and (4 in H)
[…]
