+++ tests/classes/test_directed_reportviews.py
@@ -1,0 +1,280 @@
+import numpy as np
+import pytest
+
+import xgi
+from xgi.exception import IDNotFound, XGIError
+
+
+def test_edge_order(diedgelist1):
+    H = xgi.DiHypergraph(diedgelist1)
+
+    with pytest.raises(TypeError):
+        H.edges()
+
+    assert len(H.edges.filterby("order", 3)) == 2
+    assert len(H.edges.filterby("order", 2)) == 0
+
+    ord2 = H.edges.filterby("order", 3)
+    assert len(ord2) == 2
+    assert (0 in ord2) and (1 in ord2)
+
+    H.add_edge(([3, 7, 8, 9, 10], [11, 12]))
+
+    assert len(H.edges.filterby("order", 6)) == 1
+    assert 2 in H.edges.filterby("order", 6)
+
+
+def test_node_degree(diedgelist2):
+    H = xgi.DiHypergraph(diedgelist2)
+
+    with pytest.raises(TypeError):
+        H.edges()
+
+    assert H.degree() == {0: 1, 1: 2, 2: 3, 4: 2, 3: 1, 5: 1}
+    assert H.degree(1) == 2
+    assert H.degree(2) == 3
+    with pytest.raises(KeyError):
+        H.degree(-1)
+
+    assert len(H.nodes.filterby("degree", 1)) == 3
+    assert len(H.nodes.filterby("degree", 3)) == 1
+    deg2 = H.nodes.filterby("degree", 2)
+    assert len(deg2) == 2
+    assert (1 in deg2) and (4 in deg2)
+
+    H.add_edge(([3, 7], [9, 10]))
+
+    assert len(H.nodes.filterby("degree", 2)) == 3
+    assert len(H.nodes.filterby("degree", 3)) == 1
+    assert 3 in H.nodes.filterby("degree", 2)
+    assert 7 in H.nodes.filterby("degree", 1)
+
+
+def test_size_degree(diedgelist1, diedgelist2):
+    H1 = xgi.DiHypergraph(diedgelist1)
+    H2 = xgi.DiHypergraph(diedgelist2)
+    assert H1.edges.size.asdict() == {0: 3, 1: 4, 2: 3}
+    assert H2.edges.size.asdict() == {0: 3, 1: 1, 2: 2, 3: 3}
+    assert H1.edges.order.asdict() == {0: 2, 1: 3, 2: 2}
+    assert H2.edges.order.asdict() == {0: 2, 1: 0, 2: 1, 3: 2}
+
+
+def test_degree(edgelist1, edgelist4):
+    H1 = xgi.Hypergraph(edgelist4)
+    H2 = xgi.Hypergraph(edgelist1)
+    # test basic functionality
+    assert H1.degree(1) == 1
+    assert H1.degree(2) == 2
+    assert H1.degree(3) == 3
+    with pytest.raises(KeyError):
+        H1.degree(0)
+
+    # check len
+    assert len(H1.degree()) == 5
+
+    # test order
+    assert H2.nodes.degree(order=0).asdict() == {
+        1: 0,
+        2: 0,
+        3: 0,
+        4: 1,
+        5: 0,
+        6: 0,
+        7: 0,
+        8: 0,
+    }
+    assert H2.nodes.degree(order=1).asdict() == {
+        1: 0,
+        2: 0,
+        3: 0,
+        4: 0,
+        5: 1,
+        6: 1,
+        7: 0,
+        8: 0,
+    }
+    assert H2.nodes.degree(order=2).asdict() == {
+        1: 1,
+        2: 1,
+        3: 1,
+        4: 0,
+        5: 0,
+        6: 1,
+        7: 1,
+        8: 1,
+    }
+
[…]
