+++ tests/classes/test_directed_reportviews.py
@@ -53,82 +53,58 @@ def test_node_degree(diedgelist2):
 def test_size_degree(diedgelist1, diedgelist2):
     H1 = xgi.DiHypergraph(diedgelist1)
     H2 = xgi.DiHypergraph(diedgelist2)
-    assert H1.edges.size.asdict() == {0: 3, 1: 4, 2: 3}
-    assert H2.edges.size.asdict() == {0: 3, 1: 1, 2: 2, 3: 3}
-    assert H1.edges.order.asdict() == {0: 2, 1: 3, 2: 2}
-    assert H2.edges.order.asdict() == {0: 2, 1: 0, 2: 1, 3: 2}
+
+    assert H1.edges.size.asdict() == {0: 4, 1: 4}
+    assert H2.edges.size.asdict() == {0: 3, 1: 3, 2: 4}
+    assert H1.edges.order.asdict() == {0: 3, 1: 3}
+    assert H2.edges.order.asdict() == {0: 2, 1: 2, 2: 3}
 
 
-def test_degree(edgelist1, edgelist4):
-    H1 = xgi.Hypergraph(edgelist4)
-    H2 = xgi.Hypergraph(edgelist1)
+def test_degree(diedgelist1, diedgelist2):
+    H1 = xgi.DiHypergraph(diedgelist1)
+    H2 = xgi.DiHypergraph(diedgelist2)
     # test basic functionality
     assert H1.degree(1) == 1
-    assert H1.degree(2) == 2
-    assert H1.degree(3) == 3
+    assert H1.degree(2) == 1
+    assert H1.degree(3) == 1
     with pytest.raises(KeyError):
         H1.degree(0)
 
     # check len
-    assert len(H1.degree()) == 5
+    assert len(H1.degree()) == 8
 
     # test order
-    assert H2.nodes.degree(order=0).asdict() == {
-        1: 0,
-        2: 0,
-        3: 0,
-        4: 1,
-        5: 0,
-        6: 0,
-        7: 0,
-        8: 0,
-    }
-    assert H2.nodes.degree(order=1).asdict() == {
-        1: 0,
-        2: 0,
-        3: 0,
-        4: 0,
-        5: 1,
-        6: 1,
-        7: 0,
-        8: 0,
-    }
-    assert H2.nodes.degree(order=2).asdict() == {
-        1: 1,
-        2: 1,
-        3: 1,
-        4: 0,
-        5: 0,
-        6: 1,
-        7: 1,
-        8: 1,
-    }
+    assert H2.nodes.degree(order=1).asdict() == {0: 0, 1: 0, 2: 0, 4: 0, 3: 0, 5: 0}
+    assert H2.nodes.degree(order=2).asdict() == {0: 1, 1: 2, 2: 2, 4: 1, 3: 0, 5: 0}
+    assert H2.nodes.degree(order=3).asdict() == {0: 0, 1: 0, 2: 1, 4: 1, 3: 1, 5: 1}
 
     # test weights
     attr_dict1 = {0: {"weight": -2}, 1: {"weight": 4.0}, 2: {"weight": 0.3}}
-    xgi.set_edge_attributes(H1, attr_dict1)
+    xgi.set_edge_attributes(H2, attr_dict1)
 
-    assert H1.nodes.degree(weight="weight").asdict() == {
-        1: -2,
-        2: 2,
-        3: 2.3,
+    assert H2.nodes.degree(weight="weight").asdict() == {
+        0: -2,
+        1: 2.0,
+        2: 2.3,
         4: 4.3,
-        5: 4.3,
-    }
-    assert H1.nodes.degree(weight="weight", order=2).asdict() == {
-        1: -2,
-        2: -2,
-        3: -1.7,
-        4: 0.3,
[…]
