+++ tests/stats/test_dinodestats.py
@@ -1,0 +1,178 @@
+import numpy as np
+import pandas as pd
+import pytest
+
+import xgi
+
+
+def test_filterby_wrong_stat():
+    H = xgi.DiHypergraph()
+    with pytest.raises(AttributeError):
+        H.nodes.filterby("__I_DO_NOT_EXIST__", None)
+
+
+def test_filterby(diedgelist1, diedgelist2):
+    H = xgi.DiHypergraph(diedgelist2)
+    assert H.nodes.filterby("degree", 2).in_degree.asdict() == {1: 2, 4: 1}
+    assert H.nodes.filterby("degree", 2).out_degree.asdict() == {1: 0, 4: 2}
+    assert H.nodes.filterby("in_degree", 1).degree.asdict() == {0: 1, 3: 1, 4: 2}
+    assert H.nodes.filterby("out_degree", 1).degree.asdict() == {2: 3, 5: 1}
+
+    H = xgi.DiHypergraph(diedgelist1)
+    assert H.nodes.filterby("degree", 1).in_degree.asdict() == {
+        1: 1,
+        2: 1,
+        3: 1,
+        4: 0,
+        5: 1,
+        6: 1,
+        7: 0,
+        8: 0,
+    }
+    assert H.nodes.filterby("degree", 1).out_degree.asdict() == {
+        1: 0,
+        2: 0,
+        3: 0,
+        4: 1,
+        5: 0,
+        6: 1,
+        7: 1,
+        8: 1,
+    }
+    assert H.nodes.filterby("in_degree", 1).degree.asdict() == {
+        1: 1,
+        2: 1,
+        3: 1,
+        5: 1,
+        6: 1,
+    }
+    assert H.nodes.filterby("out_degree", 1).degree.asdict() == {8: 1, 4: 1, 6: 1, 7: 1}
+
+
+def test_filterby_with_nodestat(diedgelist2):
+    H = xgi.DiHypergraph(diedgelist2)
+    assert list(H.nodes.filterby(H.nodes.degree(order=2), 1)) == [0, 4]
+
+
+def test_filterby_modes(diedgelist2):
+    H = xgi.DiHypergraph(diedgelist2)
+    assert list(H.nodes.filterby("degree", 2)) == [1, 4]
+    assert list(H.nodes.filterby("degree", 2, "eq")) == [1, 4]
+    assert list(H.nodes.filterby("degree", 1, "neq")) == [1, 2, 4]
+    assert list(H.nodes.filterby("degree", 3, "geq")) == [2]
+    assert list(H.nodes.filterby("degree", 3, "gt")) == []
+    assert list(H.nodes.filterby("degree", 0, "leq")) == []
+    assert list(H.nodes.filterby("degree", 1, "lt")) == []
+    assert set(H.nodes.filterby("degree", (1, 3), "between")) == set(H.nodes)
+
+
+def test_call_filterby(diedgelist1, diedgelist2):
+    H = xgi.DiHypergraph(diedgelist1)
+
+    filtered = H.nodes([4, 5, 6]).filterby("in_degree", 1).degree
+    assert filtered.asdict() == {5: 1, 6: 1}
+
+    H = xgi.DiHypergraph(diedgelist2)
+    assert set(H.nodes([1, 2, 3]).filterby("degree", 2)) == {1}
+
+
+def test_single_node(diedgelist1):
+    H = xgi.DiHypergraph(diedgelist1)
+    assert H.degree()[1] == 1
+    assert H.nodes.degree[1] == 1
+    with pytest.raises(KeyError):
+        H.degree()[-1]
+    with pytest.raises(KeyError):
+        H.nodes.degree[-1]
+
+
[…]
