+++ tests/stats/test_dinodestats.py
@@ -102,6 +102,24 @@ def test_degree(diedgelist2):
     assert H.nodes.degree.aslist() == list(degs.values())
 
 
+def test_in_degree(diedgelist2):
+    H = xgi.DiHypergraph(diedgelist2)
+    degs = {0: 1, 1: 2, 2: 2, 4: 1, 3: 1, 5: 0}
+    assert H.in_degree() == degs
+    assert H.in_degree(order=2) == {0: 1, 1: 2, 2: 1, 4: 0, 3: 0, 5: 0}
+    assert H.nodes.in_degree.asdict() == degs
+    assert H.nodes.in_degree.aslist() == list(degs.values())
+
+
+def test_out_degree(diedgelist2):
+    H = xgi.DiHypergraph(diedgelist2)
+    degs = {0: 0, 1: 0, 2: 1, 4: 2, 3: 0, 5: 1}
+    assert H.out_degree() == degs
+    assert H.out_degree(order=2) == {0: 0, 1: 0, 2: 1, 4: 1, 3: 0, 5: 0}
+    assert H.nodes.out_degree.asdict() == degs
+    assert H.nodes.out_degree.aslist() == list(degs.values())
+
+
 def test_aggregates(diedgelist2):
     H = xgi.DiHypergraph(diedgelist2)
     assert H.nodes.degree.max() == 3
