+++ xgi/generators/simplicial_complexes.py
@@ -133,17 +133,7 @@ def flag_complex(G, max_order=2, ps=None, seed=None):
     N = len(nodes)
     edges = G.edges()
 
-    # compute all cliques to fill
-    if max_order is None:
-        cliques_to_add = list(nx.find_cliques(G)) # max cliques
-    else: # avoid adding many unnecessary redundant cliques
-        cliques_to_add = []
-        for clique in nx.enumerate_all_cliques(G): # sorted by size
-            if len(clique)<=max_order+1: 
-                cliques_to_add.append(clique)
-            else:
-                break # dont go over whole list if not necessary
-        cliques_to_add = cliques_to_add[N:] # remove singletons
+    cliques_to_add = _cliques_to_fill(G, max_order)
 
     S = SimplicialComplex()
     S.add_nodes_from(nodes)
@@ -283,20 +273,46 @@ def random_flag_complex(N, p, max_order=2, seed=None):
 
     nodes = G.nodes()
 
-    # compute all cliques to fill
-    if max_order is None:
-        cliques = list(nx.find_cliques(G)) # max cliques
-    else: # avoid adding many unnecessary redundant cliques
-        cliques = []
-        for clique in nx.enumerate_all_cliques(G): # sorted by size
-            if len(clique)<=max_order+1: 
-                cliques.append(clique)
-            else:
-                break # dont go over whole list if not necessary
-        cliques = cliques[N:] # remove singletons
-
+    cliques = _cliques_to_fill(G, max_order)
+        
     S = SimplicialComplex()
     S.add_nodes_from(nodes)
     S.add_simplices_from(cliques, max_order=max_order)
 
     return S
+
+
+def _cliques_to_fill(G, max_order):
+    """Return cliques to fill for flag complexes, 
+    to be passed to `add_simplices_from`.
+
+    This function was written to speedup flag_complex functions
+    by avoiding adding redundant faces.
+
+    Parameters
+    ----------
+    G : networkx Graph
+        Graph to consider
+    max_order: int or None
+        If None, return maximal cliques. If int, return all cliques
+        up to max_order.
+
+    Returns
+    -------
+    cliques : list
+        List of cliques
+
+    """
+    if max_order is None:
+        cliques = list(nx.find_cliques(G)) # max cliques
+    else: # avoid adding many unnecessary redundant cliques
+        cliques = []
+        for clique in nx.enumerate_all_cliques(G): # sorted by size
+            if len(clique) == 1:
+                continue # don't add singletons
+            if len(clique) <=max_order+1: 
+                cliques.append(clique)
+            else:
+                break # dont go over whole list if not necessary
+
+    return cliques
