### 2023-05-10 26ba651
Faster flag complex (#355)
* feat: added complete hypergraph

* tests: added corresponding

* docs: added new function. style: black isort

* review comments

* review comments

* fix docs maths

* perf: faster random_flag_complex by adding only necessary cliques. tests: moved for consistency and added corresponding

* perf: same for flag_complex

* review comments

* style: black
