### 2023-05-10 7a1e85e
Try ruff (#346)
Fix most ruff errors: fix line length problems, prefer function definition over assigning to lambdas, unused variables and related issues, and add some new tests.
