+++ tests/algorithms/test_clustering.py
@@ -1,4 +1,3 @@
-import numpy as np
 import pytest
 
 import xgi
