### 2023-05-12 767c4ee
Update layout.py
wrote explicitly the circular layout and the spiral layout.
