### 2023-05-12 3e7b158
added tests + minor fix
