+++ tests/drawing/test_layout.py
@@ -135,3 +135,45 @@ def test_pca_transform():
     assert np.allclose(transform1_pos2[1], np.array([-0.67296626, -0.21706316]))
     assert np.allclose(transform1_pos2[2], np.array([-0.02177678, -2.23596193]))
     assert np.allclose(transform1_pos2[3], np.array([4.47192387, -0.04355357]))
+
+def test_circular_layout():
+    # hypergraph
+    H = xgi.random_hypergraph(10, [0.2], seed=1)
+    pos = xgi.circular_layout(H)
+    assert len(pos) == H.num_nodes
+
+    # simplicial complex
+    S = xgi.random_flag_complex_d2(10, 0.2)
+    pos = xgi.circular_layout(S)
+    assert len(pos) == S.num_nodes
+
+def test_spiral_layout():
+    # hypergraph
+    H = xgi.random_hypergraph(10, [0.2], seed=1)
+    pos1 = xgi.spiral_layout(H)
+    pos2 = xgi.spiral_layout(H, equidistant=True)
+    assert pos1.keys() == pos2.keys()
+    assert len(pos1) == H.num_nodes
+    assert len(pos2) == H.num_nodes
+
+    # simplicial complex
+    S = xgi.random_flag_complex_d2(10, 0.2)
+    pos = xgi.spiral_layout(S)
+    assert len(pos) == S.num_nodes
+    
+def test_barycenter_kamada_kawai_layout(hypergraph1):
+    H = xgi.random_hypergraph(10, [0.2], seed=1)
+
+    pos1 = xgi.barycenter_kamada_kawai_layout(H)
+    assert len(pos1) == H.num_nodes
+
+    # phantom
+    pos2, G = xgi.barycenter_kamada_kawai_layout(H, return_phantom_graph=True)
+    pos3 = xgi.barycenter_kamada_kawai_layout(H, return_phantom_graph=False)
+    assert pos3.keys() == pos3.keys()
+    assert np.allclose(list(pos3.values()), list(pos3.values()))
+
+    # simplicial complex
+    S = xgi.random_flag_complex_d2(10, 0.2)
+    pos = xgi.barycenter_kamada_kawai_layout(H)
+    assert len(pos) == S.num_nodes
