### 2023-05-12 6c6ae3e
added tests for empty and single node hypergraph
