### 2023-05-12 8a8868c
Update xgi/drawing/layout.py
Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>
