### 2023-05-15 5b5803d
_augmented_projection helper function
+ minor fix in test (missing line in test_pairwise_spring_layout())
+ coherence in docstrings
+ run black
