+++ xgi/drawing/layout.py
@@ -17,7 +17,7 @@ __all__ = [
     "pca_transform",
     "circular_layout",
     "spiral_layout",
-    "barycenter_kamada_kawai_layout"
+    "barycenter_kamada_kawai_layout",
 ]
 
 
@@ -128,54 +128,21 @@ def pairwise_spring_layout(H, seed=None):
     return pos
 
 
-def barycenter_spring_layout(H, return_phantom_graph=False, seed=None):
-    """
-    Position the nodes using Fruchterman-Reingold force-directed
-    algorithm using an augmented version of the the graph projection
-    of the hypergraph (or simplicial complex), where phantom nodes
-    (barycenters) are created for each edge composed by more than two nodes.
-    If a simplicial complex is provided the results will be based on the
-    hypergraph constructed from its maximal simplices.
+def _augmented_projection(H):
+    """Augmented version of the the graph projection of the hypergraph
+    (or simplicial complex), where phantom nodes (barycenters) are created
+    for each edge composed by more than two nodes.
+
 
     Parameters
     ----------
-    H : xgi Hypergraph or SimplicialComplex
-        A position will be assigned to every node in H.
-    return_phantom_graph: bool (default=False)
-        If True the function returns also the augmented version of the 
-        the graph projection of the hypergraph (or simplicial complex).
-    seed : int, RandomState instance or None  optional (default=None)
-        Set the random state for deterministic node layouts.
-        If int, `seed` is the seed used by the random number generator,
-        If None (default), random numbers are sampled from the
-        numpy random number generator without initialization.
+    H : Hypergraph or SimplicialComplex
 
     Returns
     -------
-    pos : dict
-        A dictionary of positions keyed by node
-
-    See Also
-    --------
-    random_layout
-    pairwise_spring_layout
-    weighted_barycenter_spring_layout
-
-    Examples
-    --------
-    >>> import xgi
-    >>> N = 50
-    >>> ps = [0.1, 0.01]
-    >>> H = xgi.random_hypergraph(N, ps)
-    >>> pos = xgi.barycenter_spring_layout(H)
-
+    G : networkx.Graph
+        The augmented version of the graph projection
     """
-    if seed is not None:
-        random.seed(seed)
-
-    if isinstance(H, SimplicialComplex):
-        H = convert.from_max_simplices(H)
-
     # Creating the projected networkx Graph, I will fill it manually
     G = nx.Graph()
 
@@ -202,6 +169,57 @@ def barycenter_spring_layout(H, return_phantom_graph=False, seed=None):
             for n in he:
                 G.add_edge(phantom_node_id, n)
             phantom_node_id += 1
+    return G
+
+
+def barycenter_spring_layout(H, return_phantom_graph=False, seed=None):
+    """
+    Position the nodes using Fruchterman-Reingold force-directed
+    algorithm using an augmented version of the the graph projection
+    of the hypergraph (or simplicial complex), where phantom nodes
+    (barycenters) are created for each edge composed by more than two nodes.
+    If a simplicial complex is provided the results will be based on the
+    hypergraph constructed from its maximal simplices.
+
+    Parameters
+    ----------
+    H : xgi Hypergraph or SimplicialComplex
+        A position will be assigned to every node in H.
+    return_phantom_graph: bool (default=False)
+        If True the function returns also the augmented version of the
+        the graph projection of the hypergraph (or simplicial complex).
+    seed : int, RandomState instance or None  optional (default=None)
+        Set the random state for deterministic node layouts.
+        If int, `seed` is the seed used by the random number generator,
[…]
