### 2023-05-15 e41134f
Update test_layout.py
as in main
