### 2023-05-15 #5162 @maximelucas
Thank you very much @vissarion and all three reviewers for your time.

I've updated two references from their arXiv version to their published one just now.
