### 2023-05-15 374a1ca
New layouts (#360)
* Update layout.py

* Update layout.py

wrote explicitly the circular layout and the spiral layout.

* Update layout.py

added center in spiral

* added tests + minor fix

* minor fix

* minor fix

* added tests for empty and single node hypergraph

* Update xgi/drawing/layout.py

Co-authored-by: Maxime Lucas <maximelucas@users.noreply.github.com>

* Update layout.py

* after review

* _augmented_projection helper function

+ minor fix in test (missing line in test_pairwise_spring_layout())
+ coherence in docstrings
+ run black

* Update test_layout.py

as in main

* Update 
[…]
