+++ xgi/drawing/layout.py
@@ -15,6 +15,9 @@ __all__ = [
     "barycenter_spring_layout",
     "weighted_barycenter_spring_layout",
     "pca_transform",
+    "circular_layout",
+    "spiral_layout",
+    "barycenter_kamada_kawai_layout",
 ]
 
 
@@ -131,6 +134,61 @@ def pairwise_spring_layout(H, seed=None):
     return pos
 
 
+def _augmented_projection(H, weighted=False):
+    """Augmented version of the the graph projection of the hypergraph
+    (or simplicial complex), where phantom nodes (barycenters) are created
+    for each edge composed by more than two nodes.
+
+    Parameters
+    ----------
+    H : Hypergraph or SimplicialComplex
+    weighted : bool (default=False)
+               If True weights are assigned to all hyperedges of order d=1 (links)
+               and to all connections to phantom nodes within each hyperedge of
+               order d>1 to keep them together. Weights scale as the order d.
+
+
+    Returns
+    -------
+    G : networkx.Graph
+        The augmented version of the graph projection
+    """
+    # Creating the projected networkx Graph, I will fill it manually
+    G = nx.Graph()
+
+    # Adding real nodes
+    G.add_nodes_from(list(H.nodes))
+
+    # Adding links (edges composed by two nodes only, for which we don't use phantom nodes
+    for i, j in H.edges.filterby("order", 1).members():
+        if weighted == True:
+            G.add_edge(i, j, weight=1)
+        else:
+            G.add_edge(i, j)
+
+    # Adding phantom nodes and connections therein
+    # I will start from the first int node-label available
+    try:
+        phantom_node_id = max([n for n in H.nodes if isinstance(n, int)]) + 1
+    except ValueError:
+        # The list of node-labels has no integers, so I start from 0
+        phantom_node_id = 0
+
+    # Looping over the hyperedges of different order (from triples up)
+    for d in range(2, max_edge_order(H) + 1):
+        # Hyperedges of order d (d=2: triplets, etc.)
+        for he_id, members in H.edges.filterby("order", d).members(dtype=dict).items():
+            # Adding one phantom node for each hyperedge and linking it to the nodes of
+            # the hyperedge
+            for n in members:
+                if weighted == True:
+                    G.add_edge(phantom_node_id, n, weight=d)
+                else:
+                    G.add_edge(phantom_node_id, n)
+            phantom_node_id += 1
+    return G
+
+
 def barycenter_spring_layout(H, return_phantom_graph=False, seed=None):
     """
     Position the nodes using Fruchterman-Reingold force-directed
@@ -144,7 +202,10 @@ def barycenter_spring_layout(H, return_phantom_graph=False, seed=None):
     ----------
     H : xgi Hypergraph or SimplicialComplex
         A position will be assigned to every node in H.
-    seed : int, optional
+    return_phantom_graph: bool (default=False)
+        If True the function returns also the augmented version of the
+        the graph projection of the hypergraph (or simplicial complex).
+    seed : int, RandomState instance or None  optional (default=None)
         Set the random state for deterministic node layouts.
         If int, `seed` is the seed used by the random number generator,
         If None (default), random numbers are sampled from the
@@ -168,7 +229,6 @@ def barycenter_spring_layout(H, return_phantom_graph=False, seed=None):
     >>> ps = [0.1, 0.01]
     >>> H = xgi.random_hypergraph(N, ps)
     >>> pos = xgi.barycenter_spring_layout(H)
-
     """
     if seed is not None:
         random.seed(seed)
@@ -176,34 +236,7 @@ def barycenter_spring_layout(H, return_phantom_graph=False, seed=None):
     if isinstance(H, SimplicialComplex):
[…]
