+++ docs/source/api/drawing/xgi.drawing.layout.rst
@@ -12,3 +12,6 @@ xgi.drawing.layout
    .. autofunction:: barycenter_spring_layout
    .. autofunction:: weighted_barycenter_spring_layout
    .. autofunction:: pca_transform
+   .. autofunction:: circular_layout
+   .. autofunction:: spiral_layout
+   .. autofunction:: barycenter_kamada_kawai_layout
