+++ paper/references.bib
@@ -1,15 +1,12 @@
-@misc{adhikari_synchronization_2022,
-  title = {Synchronization of Phase Oscillators on Complex Hypergraphs},
-  author = {Adhikari, Sabina and Restrepo, Juan G. and Skardal, Per Sebastian},
-  year = {2022},
-  month = oct,
-  number = {arXiv:2208.00909},
-  eprint = {2208.00909},
-  eprinttype = {arxiv},
-  primaryclass = {cond-mat, physics:nlin},
-  publisher = {{arXiv}},
-  doi = {10.48550/arXiv.2208.00909},
-  archiveprefix = {arXiv}
+@article{adhikari_synchronization_2022,
+  title={Synchronization of phase oscillators on complex hypergraphs},
+  author={Adhikari, Sabina and Restrepo, Juan G and Skardal, Per Sebastian},
+  journal={Chaos: An Interdisciplinary Journal of Nonlinear Science},
+  volume={33},
+  number={3},
+  pages={033116},
+  year={2023},
+  doi={10.1063/5.0116747},
 }
 
 @article{arnaudon_connecting_2022,
@@ -244,15 +241,13 @@
   doi = {10.24166/im.01.2020}
 }
 
-@misc{zhang_higher-order_2022,
-  title = {Do Higher-Order Interactions Promote Synchronization?},
-  author = {Zhang, Yuanzhao and Lucas, Maxime and Battiston, Federico},
-  year = {2022},
-  month = mar,
-  number = {arXiv:2203.03060},
-  eprint = {2203.03060},
-  eprinttype = {arxiv},
-  publisher = {{arXiv}},
-  doi = {10.48550/arXiv.2203.03060},
-  archiveprefix = {arXiv}
+@article{zhang_higher-order_2022,
+  title = {Higher-order interactions shape collective dynamics differently in hypergraphs and simplicial complexes},
+  author = {Zhang, Y. and Lucas, M. and Battiston, F.},
+  year = {2023},
+  journal = {Nature Communications},
+  volume = {14},
+  number = {1},
+  pages = {1605},
+  doi = {10.1038/s41467-023-37190-9},
 }
