+++ paper/references.bib
@@ -1,7 +1,7 @@
 @article{adhikari_synchronization_2022,
   title={Synchronization of phase oscillators on complex hypergraphs},
   author={Adhikari, Sabina and Restrepo, Juan G and Skardal, Per Sebastian},
-  journal={Chaos: An Interdisciplinary Journal of Nonlinear Science},
+  journal={Chaos},
   volume={33},
   number={3},
   pages={033116},
