### 2023-05-15 #5162 @nwlandry
Hi @vissarion! Thanks so much! I have released XGI 0.6 on PyPI and have made an associated DOI: https://doi.org/10.5281/zenodo.7939055 - let me know if there's anything else I can provide.
