### 2023-05-16 xgi-org/xgi#363 @vissarion
[JOSS review] More information needed in some references
This issue is raised during a JOSS review https://github.com/openjournals/joss-reviews/issues/5162

The submitted paper will benefit from adding some extra information in some references.

In particular,

- "Landry, N., & Miller, J. (2022). Hypercontagion."
- Praggastis, B., Arendt, D., Joslyn, C., Purvine, E., Aksoy, S., & Monson, K. (n.d.). PNNL/Hy-129
perNetX.

These are libraries, so please cite them as libraries by either following the library authors instructions or by manually adding some text (e.g. "Hypercontagion (HNX): a Python library for hypergraphs") and a link to the documentation or the repository that host the library.

- Hagberg, A. A., Schult, D. A., & Swart, P. J. (2008). Exploring network structure, dynamics,...7th python in science conference (pp. 11–15)

Here python could be capitalized i.e. Python and a link could be provided (e.g. https://www.osti.gov/biblio/960616)

- Landry, N., Torres, L., Lucas, M., Iacopini, I., Petri, G., Patania, A., & Schwarze, A. (2023).115
XGI-DATA

This refers to data? Please provide more details and a url.

- Peixoto, T. P. (2021). Netzschleuder.

AFAIU this is a network catalogue, please add some description and a url.
