### 2023-05-16 #5162 @vissarion
Thanks  @nwlandry for the info. I created an [issue](https://github.com/xgi-org/xgi/issues/363) related to paper references.
