+++ paper/references.bib
@@ -54,7 +54,7 @@
   title = {Data!},
   author = {Benson, Austin},
   year = {2021},
-  howpublished = {https://www.cs.cornell.edu/\textasciitilde arb/data/}
+  url = {https://www.cs.cornell.edu/\textasciitilde arb/data/}
 }
 
 @article{benson_three_2019,
@@ -85,7 +85,8 @@
 @misc{clauset_colorado_2016,
   title = {The {{Colorado Index}} of {{Complex Networks}}},
   author = {Clauset, Aaron and Tucker, Ellen and Sainz, Matthias},
-  year = {2016}
+  year = {2016},
+  url = {https://icon.colorado.edu/}
 }
 
 @article{diaz_hypergraphsjl_2022,
@@ -103,18 +104,21 @@
 }
 
 @article{doecode_22160,
-  title = {{{PNNL}}/{{HyperNetX}}},
-  author = {Praggastis, Brenda and Arendt, Dustin and Joslyn, Cliff and Purvine, Emilie and Aksoy, Sinan and Monson, Kyle}
+  title = {{{HyperNetX (HNX): a Python library for hypergraphs}}},
+  author = {Praggastis, Brenda and Arendt, Dustin and Joslyn, Cliff and Purvine, Emilie and Aksoy, Sinan and Monson, Kyle},
+  url = {https://github.com/pnnl/HyperNetX},
+  year = {2023}
 }
 
 @inproceedings{hagberg_exploring_2008,
   title = {Exploring Network Structure, Dynamics, and Function Using {{NetworkX}}},
-  booktitle = {Proceedings of the 7th Python in Science Conference},
+  booktitle = {Proceedings of the 7th {{Python}} in {{Science}} {{Conference}}},
   author = {Hagberg, Aric A. and Schult, Daniel A. and Swart, Pieter J.},
   editor = {Varoquaux, Ga{\"e}l and Vaught, Travis and Millman, Jarrod},
   year = {2008},
   pages = {11--15},
-  address = {{Pasadena, CA USA}}
+  address = {{Pasadena, CA USA}},
+  url = {https://www.osti.gov/biblio/960616}
 }
 
 @article{hickok_bounded-confidence_2022,
@@ -153,16 +157,17 @@
 }
 
 @misc{landry_hypercontagion_2022,
-  title = {Hypercontagion},
+  title = {{{Hypercontagion: a Python package for contagion processes on higher-order networks.}}},
   author = {Landry, Nicholas and Miller, Joel},
   year = {2022},
-  month = aug
+  url = {https://github.com/nwlandry/hypercontagion}
 }
 
 @misc{landry_xgi-data_2023,
-  title = {{{XGI-DATA}}},
+  title = {{{XGI-DATA: a repository of hypergraph datasets in JSON format}}},
   author = {Landry, Nicholas and Torres, Leo and Lucas, Maxime and Iacopini, Iacopo and Petri, Giovanni and Patania, Alice and Schwarze, Alice},
-  year = {2023}
+  year = {2023},
+  url = {https://github.com/xgi-org/xgi-data}
 }
 
 @article{lucas_multiorder_2020,
@@ -205,7 +210,7 @@
 }
 
 @article{peixoto_graph-tool_2014,
-  title = {The Graph-Tool Python Library},
+  title = {{{The graph-tool Python library}}},
   author = {Peixoto, Tiago P.},
   year = {2014},
   journal = {figshare},
@@ -213,9 +218,10 @@
 }
 
 @misc{peixoto_netzschleuder_2021,
-  title = {Netzschleuder},
+  title = {The Netzschleuder network catalogue and repository},
   author = {Peixoto, Tiago P.},
[…]
