+++ paper/references.bib
@@ -54,7 +54,7 @@
   title = {Data!},
   author = {Benson, Austin},
   year = {2021},
-  url = {https://www.cs.cornell.edu/\textasciitilde arb/data/}
+  url = {https://www.cs.cornell.edu/%7Earb/data/}
 }
 
 @article{benson_three_2019,
@@ -218,7 +218,7 @@
 }
 
 @misc{peixoto_netzschleuder_2021,
-  title = {The Netzschleuder network catalogue and repository},
+  title = {{{The Netzschleuder}} network catalogue and repository},
   author = {Peixoto, Tiago P.},
   year = {2020},
   url = {https://networks.skewed.de/}
