+++ paper/references.bib
@@ -58,7 +58,7 @@
 }
 
 @article{benson_three_2019,
-  title = {Three {{Hypergraph Eigenvector Centralities}}},
+  title = {Three Hypergraph Eigenvector Centralities},
   author = {Benson, Austin R.},
   year = {2019},
   month = jan,
@@ -83,7 +83,7 @@
 }
 
 @misc{clauset_colorado_2016,
-  title = {The {{Colorado Index}} of {{Complex Networks}}},
+  title = {The {C}olorado Index of Complex Networks},
   author = {Clauset, Aaron and Tucker, Ellen and Sainz, Matthias},
   year = {2016},
   url = {https://icon.colorado.edu/}
@@ -122,7 +122,7 @@
 }
 
 @article{hickok_bounded-confidence_2022,
-  title = {A {{Bounded-Confidence Model}} of {{Opinion Dynamics}} on {{Hypergraphs}}},
+  title = {A Bounded-Confidence Model of Opinion Dynamics on Hypergraphs},
   author = {Hickok, Abigail and Kureh, Yacoub and Brooks, Heather Z. and Feng, Michelle and Porter, Mason A.},
   year = {2022},
   month = jan,
@@ -157,7 +157,7 @@
 }
 
 @misc{landry_hypercontagion_2022,
-  title = {{{Hypercontagion: a Python package for contagion processes on higher-order networks.}}},
+  title = {Hypercontagion: a {P}ython package for contagion processes on higher-order networks},
   author = {Landry, Nicholas and Miller, Joel},
   year = {2022},
   url = {https://github.com/nwlandry/hypercontagion}
@@ -184,7 +184,7 @@
 }
 
 @article{millan_explosive_2020,
-  title = {Explosive {{Higher-Order Kuramoto Dynamics}} on {{Simplicial Complexes}}},
+  title = {Explosive Higher-Order {K}uramoto Dynamics on Simplicial Complexes},
   author = {Mill{\'a}n, Ana P. and Torres, Joaqu{\'i}n J. and Bianconi, Ginestra},
   year = {2020},
   month = may,
@@ -237,7 +237,7 @@
 }
 
 @article{spagnuolo_analyzing_2020,
-  title = {Analyzing, {{Exploring}}, and {{Visualizing Complex Networks}} via {{Hypergraphs}} Using {{SimpleHypergraphs}}.Jl},
+  title = {Analyzing, Exploring, and Visualizing Complex Networks via Hypergraphs Using {SimpleHypergraphs.jl}},
   author = {Spagnuolo, Carmine and Cordasco, Gennaro and Szufel, Przemyslaw and Pra{\l}at, Pawe{\l} and Scarano, Vittorio and Kami{\'n}ski, Bogumi{\l} and Antelmi, Alessia},
   year = {2020},
   month = mar,
